refactor: update session creation logic and modify start modes
