+++ server_core/intelligence/decision_engine_constants.py
@@ -1,0 +1,303 @@
+from typing import Any, Dict, List
+
+from shared.target_types import TargetType, TechnologyStack
+
+
+def initialize_tool_effectiveness() -> Dict[str, Dict[str, float]]:
+    """Initialize tool effectiveness ratings for different target types."""
+    return {
+        TargetType.WEB_APPLICATION.value: {
+            "nmap": 0.8,
+            "gobuster": 0.9,
+            "nuclei": 0.95,
+            "nikto": 0.85,
+            "sqlmap": 0.9,
+            "ffuf": 0.9,
+            "feroxbuster": 0.85,
+            "katana": 0.88,
+            "httpx": 0.85,
+            "wpscan": 0.95,
+            "burpsuite": 0.9,
+            "dirsearch": 0.87,
+            "gau": 0.82,
+            "waybackurls": 0.8,
+            "arjun": 0.9,
+            "paramspider": 0.85,
+            "x8": 0.88,
+            "jaeles": 0.92,
+            "dalfox": 0.93,
+            "anew": 0.7,
+            "qsreplace": 0.75,
+            "uro": 0.7,
+        },
+        TargetType.NETWORK_HOST.value: {
+            "nmap": 0.95,
+            "nmap_advanced": 0.97,
+            "masscan": 0.92,
+            "rustscan": 0.9,
+            "autorecon": 0.95,
+            "enum4linux": 0.8,
+            "enum4linux-ng": 0.88,
+            "smbmap": 0.85,
+            "rpcclient": 0.82,
+            "nbtscan": 0.75,
+            "arp-scan": 0.85,
+            "responder": 0.88,
+            "hydra": 0.8,
+            "netexec": 0.85,
+            "amass": 0.7,
+        },
+        TargetType.API_ENDPOINT.value: {
+            "nuclei": 0.9,
+            "ffuf": 0.85,
+            "arjun": 0.95,
+            "paramspider": 0.88,
+            "httpx": 0.9,
+            "x8": 0.92,
+            "katana": 0.85,
+            "jaeles": 0.88,
+            "postman": 0.8,
+        },
+        TargetType.CLOUD_SERVICE.value: {
+            "prowler": 0.95,
+            "scout-suite": 0.92,
+            "cloudmapper": 0.88,
+            "pacu": 0.85,
+            "trivy": 0.9,
+            "clair": 0.85,
+            "kube-hunter": 0.9,
+            "kube-bench": 0.88,
+            "docker-bench-security": 0.85,
+            "falco": 0.87,
+            "checkov": 0.9,
+            "terrascan": 0.88,
+        },
+        TargetType.BINARY_FILE.value: {
+            "ghidra": 0.95,
+            "radare2": 0.9,
+            "gdb": 0.85,
+            "gdb-peda": 0.92,
+            "angr": 0.88,
+            "pwntools": 0.9,
+            "ropgadget": 0.85,
+            "ropper": 0.88,
+            "one-gadget": 0.82,
+            "libc-database": 0.8,
+            "checksec": 0.75,
+            "strings": 0.7,
+            "objdump": 0.75,
+            "binwalk": 0.8,
+            "pwninit": 0.85,
+        },
+    }
+
+
+def initialize_technology_signatures() -> Dict[str, Dict[str, Any]]:
+    """Initialize technology detection signatures."""
+    return {
+        "headers": {
+            TechnologyStack.APACHE.value: ["Apache", "apache"],
+            TechnologyStack.NGINX.value: ["nginx", "Nginx"],
+            TechnologyStack.IIS.value: ["Microsoft-IIS", "IIS"],
+            TechnologyStack.PHP.value: ["PHP", "X-Powered-By: PHP"],
+            TechnologyStack.NODEJS.value: ["Express", "X-Powered-By: Express"],
+            TechnologyStack.PYTHON.value: ["Django", "Flask", "Werkzeug"],
+            TechnologyStack.JAVA.value: ["Tomcat", "JBoss", "WebLogic"],
+            TechnologyStack.DOTNET.value: ["ASP.NET", "X-AspNet-Version"],
+        },
+        "content": {
+            TechnologyStack.WORDPRESS.value: ["wp-content", "wp-includes", "WordPress"],
+            TechnologyStack.DRUPAL.value: ["Drupal", "drupal", "/sites/default"],
+            TechnologyStack.JOOMLA.value: ["Joomla", "joomla", "/administrator"],
+            TechnologyStack.REACT.value: ["React", "react", "__REACT_DEVTOOLS"],
+            TechnologyStack.ANGULAR.value: ["Angular", "angular", "ng-version"],
+            TechnologyStack.VUE.value: ["Vue", "vue", "__VUE__"],
+        },
+        "ports": {
+            TechnologyStack.APACHE.value: [80, 443, 8080, 8443],
+            TechnologyStack.NGINX.value: [80, 443, 8080],
+            TechnologyStack.IIS.value: [80, 443, 8080],
+            TechnologyStack.NODEJS.value: [3000, 8000, 8080, 9000],
+        },
+    }
+
+
+def initialize_attack_patterns() -> Dict[str, List[Dict[str, Any]]]:
+    """Initialize common attack patterns for different scenarios."""
+    return {
+        "web_reconnaissance": [
+            {"tool": "nmap", "priority": 1, "params": {"scan_type": "-sV -sC", "ports": "80,443,8080,8443"}},
+            {"tool": "httpx", "priority": 2, "params": {"probe": True, "tech_detect": True}},
+            {"tool": "katana", "priority": 3, "params": {"depth": 3, "js_crawl": True}},
+            {"tool": "gau", "priority": 4, "params": {"include_subs": True}},
+            {"tool": "waybackurls", "priority": 5, "params": {"get_versions": False}},
+            {"tool": "nuclei", "priority": 6, "params": {"severity": "critical,high", "tags": "tech"}},
+            {"tool": "dirsearch", "priority": 7, "params": {"extensions": "php,html,js,txt", "threads": 30}},
+            {"tool": "gobuster", "priority": 8, "params": {"mode": "dir", "extensions": "php,html,js,txt"}},
+        ],
+        "api_testing": [
+            {"tool": "httpx", "priority": 1, "params": {"probe": True, "tech_detect": True}},
+            {"tool": "arjun", "priority": 2, "params": {"method": "GET,POST", "stable": True}},
+            {"tool": "x8", "priority": 3, "params": {"method": "GET", "wordlist": "/usr/share/wordlists/x8/params.txt"}},
+            {"tool": "paramspider", "priority": 4, "params": {"level": 2}},
+            {"tool": "api-schema-analyzer", "priority": 5, "params": {"schema_type": "openapi"}},
+            {
+                "tool": "graphql-scanner",
+                "priority": 6,
+                "params": {"introspection": True, "query_depth": 10, "mutations": True},
+            },
+            {"tool": "jwt-analyzer", "priority": 7, "params": {}},
+            {"tool": "katana", "priority": 8, "params": {"depth": 3, "js_crawl": True, "form_extraction": True}},
+            {"tool": "nuclei", "priority": 5, "params": {"tags": "api,graphql,jwt", "severity": "high,critical"}},
+            {"tool": "ffuf", "priority": 6, "params": {"mode": "parameter", "method": "POST"}},
+        ],
+        "internal_network_ad_assessment": [
+            {
+                "tool": "nmap_advanced",
+                "priority": 1,
+                "params": {"scan_type": "-sS", "os_detection": True, "version_detection": True},
+            },
+            {"tool": "autorecon", "priority": 2, "params": {"port_scans": "top-1000-ports", "service_scans": "default"}},
+            {"tool": "enum4linux-ng", "priority": 3, "params": {"shares": True, "users": True, "groups": True, "policy": True}},
+            {"tool": "smbmap", "priority": 4, "params": {}},
+            {"tool": "rpcclient", "priority": 5, "params": {"commands": "enumdomusers;enumdomgroups;querydominfo"}},
+            {"tool": "nxc", "priority": 6, "params": {"protocol": "smb"}},
+            {"tool": "ldapdomaindump", "priority": 7, "params": {"authtype": "NTLM"}},
+            {"tool": "impacket-ad-enum", "priority": 8, "params": {}},
+            {"tool": "responder", "priority": 9, "params": {"wpad": True, "duration": 180}},
+            {"tool": "evil-winrm", "priority": 10, "params": {}},
+        ],
+        "network_discovery": [
+            {"tool": "arp-scan", "priority": 1, "params": {"local_network": True}},
+            {"tool": "rustscan", "priority": 2, "params": {"ulimit": 5000, "scripts": True}},
+            {
+                "tool": "nmap_advanced",
+                "priority": 3,
+                "params": {"scan_type": "-sS", "os_detection": True, "version_detection": True},
+            },
+            {"tool": "masscan", "priority": 4, "params": {"rate": 1000, "ports": "1-65535", "banners": True}},
+            {"tool": "enum4linux-ng", "priority": 5, "params": {"shares": True, "users": True, "groups": True}},
+            {"tool": "nbtscan", "priority": 6, "params": {"verbose": True}},
+            {"tool": "smbmap", "priority": 7, "params": {"recursive": True}},
+            {"tool": "rpcclient", "priority": 8, "params": {"commands": "enumdomusers;enumdomgroups;querydominfo"}},
+        ],
+        "vulnerability_assessment": [
+            {"tool": "nuclei", "priority": 1, "params": {"severity": "critical,high,medium", "update": True}},
+            {"tool": "jaeles", "priority": 2, "params": {"threads": 20, "timeout": 20}},
+            {"tool": "dalfox", "priority": 3, "params": {"mining_dom": True, "mining_dict": True}},
+            {"tool": "nikto", "priority": 4, "params": {"comprehensive": True}},
+            {"tool": "sqlmap", "priority": 5, "params": {"crawl": 2, "batch": True}},
+        ],
+        "comprehensive_network_pentest": [
+            {"tool": "autorecon", "priority": 1, "params": {"port_scans": "top-1000-ports", "service_scans": "default"}},
+            {"tool": "rustscan", "priority": 2, "params": {"ulimit": 5000, "scripts": True}},
+            {"tool": "nmap_advanced", "priority": 3, "params": {"aggressive": True, "nse_scripts": "vuln,exploit"}},
+            {"tool": "enum4linux-ng", "priority": 4, "params": {"shares": True, "users": True, "groups": True, "policy": True}},
+            {"tool": "responder", "priority": 5, "params": {"wpad": True, "duration": 180}},
+        ],
+        "binary_exploitation": [
+            {"tool": "checksec", "priority": 1, "params": {}},
+            {"tool": "ghidra", "priority": 2, "params": {"analysis_timeout": 300, "output_format": "xml"}},
+            {"tool": "ropper", "priority": 3, "params": {"gadget_type": "rop", "quality": 2}},
+            {"tool": "one-gadget", "priority": 4, "params": {"level": 1}},
+            {"tool": "pwntools", "priority": 5, "params": {"exploit_type": "local"}},
+            {"tool": "gdb-peda", "priority": 6, "params": {"commands": "checksec\ninfo functions\nquit"}},
+        ],
+        "ctf_pwn_challenge": [
+            {"tool": "pwninit", "priority": 1, "params": {"template_type": "python"}},
+            {"tool": "checksec", "priority": 2, "params": {}},
+            {"tool": "ghidra", "priority": 3, "params": {"analysis_timeout": 180}},
+            {"tool": "ropper", "priority": 4, "params": {"gadget_type": "all", "quality": 3}},
+            {"tool": "angr", "priority": 5, "params": {"analysis_type": "symbolic"}},
+            {"tool": "one-gadget", "priority": 6, "params": {"level": 2}},
+        ],
+        "aws_security_assessment": [
+            {"tool": "prowler", "priority": 1, "params": {"provider": "aws", "output_format": "json"}},
+            {"tool": "scout-suite", "priority": 2, "params": {"provider": "aws"}},
+            {"tool": "cloudmapper", "priority": 3, "params": {"action": "collect"}},
+            {"tool": "pacu", "priority": 4, "params": {"modules": "iam__enum_users_roles_policies_groups"}},
+        ],
+        "kubernetes_security_assessment": [
+            {"tool": "kube-bench", "priority": 1, "params": {"output_format": "json"}},
+            {"tool": "kube-hunter", "priority": 2, "params": {"report": "json"}},
+            {"tool": "falco", "priority": 3, "params": {"duration": 120, "output_format": "json"}},
+        ],
+        "container_security_assessment": [
+            {"tool": "trivy", "priority": 1, "params": {"scan_type": "image", "severity": "HIGH,CRITICAL"}},
+            {"tool": "clair", "priority": 2, "params": {"output_format": "json"}},
+            {"tool": "docker-bench-security", "priority": 3, "params": {}},
+        ],
+        "iac_security_assessment": [
+            {"tool": "checkov", "priority": 1, "params": {"output_format": "json"}},
+            {"tool": "terrascan", "priority": 2, "params": {"scan_type": "all", "output_format": "json"}},
+            {"tool": "trivy", "priority": 3, "params": {"scan_type": "config", "severity": "HIGH,CRITICAL"}},
+        ],
+        "multi_cloud_assessment": [
+            {"tool": "scout-suite", "priority": 1, "params": {"provider": "aws"}},
+            {"tool": "prowler", "priority": 2, "params": {"provider": "aws"}},
+            {"tool": "checkov", "priority": 3, "params": {"framework": "terraform"}},
+            {"tool": "terrascan", "priority": 4, "params": {"scan_type": "all"}},
+        ],
+        "bug_bounty_reconnaissance": [
+            {"tool": "amass", "priority": 1, "params": {"mode": "enum", "passive": False}},
+            {"tool": "subfinder", "priority": 2, "params": {"silent": True, "all_sources": True}},
+            {"tool": "httpx", "priority": 3, "params": {"probe": True, "tech_detect": True, "status_code": True}},
+            {"tool": "katana", "priority": 4, "params": {"depth": 3, "js_crawl": True, "form_extraction": True}},
+            {"tool": "gau", "priority": 5, "params": {"include_subs": True}},
+            {"tool": "waybackurls", "priority": 6, "params": {"get_versions": False}},
+            {"tool": "paramspider", "priority": 7, "params": {"level": 2}},
+            {"tool": "arjun", "priority": 8, "params": {"method": "GET,POST", "stable": True}},
+        ],
+        "bug_bounty_vulnerability_hunting": [
+            {"tool": "nuclei", "priority": 1, "params": {"severity": "critical,high", "tags": "rce,sqli,xss,ssrf"}},
+            {"tool": "dalfox", "priority": 2, "params": {"mining_dom": True, "mining_dict": True}},
+            {"tool": "sqlmap", "priority": 3, "params": {"batch": True, "level": 2, "risk": 2}},
+            {"tool": "jaeles", "priority": 4, "params": {"threads": 20, "timeout": 20}},
+            {"tool": "ffuf", "priority": 5, "params": {"match_codes": "200,204,301,302,307,401,403", "threads": 40}},
+        ],
+        "bug_bounty_high_impact": [
+            {"tool": "nuclei", "priority": 1, "params": {"severity": "critical", "tags": "rce,sqli,ssrf,lfi,xxe"}},
+            {"tool": "sqlmap", "priority": 2, "params": {"batch": True, "level": 3, "risk": 3, "tamper": "space2comment"}},
+            {"tool": "jaeles", "priority": 3, "params": {"signatures": "rce,sqli,ssrf", "threads": 30}},
+            {
+                "tool": "dalfox",
+                "priority": 4,
+                "params": {"blind": True, "mining_dom": True, "custom_payload": "alert(document.domain)"},
+            },
+        ],
+    }
+
+
+TIME_ESTIMATES: Dict[str, int] = {
+    "nmap": 120,
+    "gobuster": 300,
+    "nuclei": 180,
+    "nikto": 240,
+    "sqlmap": 600,
+    "ffuf": 200,
+    "hydra": 900,
+    "amass": 300,
+    "ghidra": 300,
+    "radare2": 180,
+    "gdb": 120,
+    "gdb-peda": 150,
+    "angr": 600,
+    "pwntools": 240,
+    "ropper": 120,
+    "one-gadget": 60,
+    "checksec": 30,
+    "pwninit": 60,
+    "libc-database": 90,
+    "prowler": 600,
+    "scout-suite": 480,
+    "cloudmapper": 300,
+    "pacu": 420,
+    "trivy": 180,
+    "clair": 240,
+    "kube-hunter": 300,
+    "kube-bench": 120,
+    "docker-bench-security": 180,
+    "falco": 120,
+    "checkov": 240,
+    "terrascan": 200,
+}
