+++ server_static/assets/index-Bn3Gkg7X.js
@@ -44,4 +44,4 @@ python3 hexstrike_server.py
 hexstrike-env/bin/python3 hexstrike_mcp.py \\
 +  --server http://localhost:8888 \\
 +  --auth-token your-secret-token \\
-+  --profile full`})]})}function xn(){return(0,Z.jsxs)(`section`,{className:`section help-about-section`,children:[(0,Z.jsx)(`div`,{className:`section-header`,children:(0,Z.jsx)(`h3`,{children:`Demo Mode`})}),(0,Z.jsxs)(`div`,{className:`help-about`,children:[(0,Z.jsx)(`p`,{className:`help-about-desc`,children:`Activate demo mode to explore the dashboard. All data is synthetic but designed to feel realistic. Ideal for learning, demos, or just satisfying your curiosity!`}),(0,Z.jsxs)(`button`,{className:`help-demo-btn`,onClick:()=>{window.location.href=window.location.pathname+`?demo=1`+window.location.hash},children:[(0,Z.jsx)(_,{size:13}),`Try demo mode`]})]})]})}var Sn=[{id:`claude`,label:`Claude Desktop`,icon:`🤖`,configPath:`~/.config/Claude/claude_desktop_config.json`,note:`Also works for Cursor — same config format.`,json:e=>JSON.stringify({mcpServers:{"hexstrike-ai":{command:`${e}/hexstrike-env/bin/python3`,args:[`${e}/hexstrike_mcp.py`,`--server`,`http://localhost:8888`,`--profile`,`full`],description:`HexStrike AI Community Edition`,timeout:300,disabled:!1}}},null,2)},{id:`vscode`,label:`VS Code Copilot`,icon:`🔷`,configPath:`.vscode/settings.json  (workspace) or User settings`,note:`Place in your workspace .vscode/settings.json or open User Settings JSON.`,json:e=>JSON.stringify({servers:{hexstrike:{type:`stdio`,command:`${e}/hexstrike-env/bin/python3`,args:[`${e}/hexstrike_mcp.py`,`--server`,`http://127.0.0.1:8888`,`--profile`,`full`]}},inputs:[]},null,2)},{id:`opencode`,label:`OpenCode`,icon:`⚡`,configPath:`~/.config/opencode/opencode.json`,note:`OpenCode reads this on startup.`,json:e=>JSON.stringify({$schema:`https://opencode.ai/config.json`,mcp:{hexstrike:{type:`local`,command:[`${e}/hexstrike-env/bin/python3`,`${e}/hexstrike_mcp.py`,`--server`,`http://127.0.0.1:8888`,`--profile`,`default`],enabled:!0}}},null,2)},{id:`roo`,label:`Roo Code`,icon:`🦘`,configPath:`MCP Servers panel  →  Edit Config`,note:`Open Roo Code → MCP Servers → Edit Config and paste the block below.`,json:e=>JSON.stringify({mcpServers:{hexstrike:{command:`${e}/hexstrike-env/bin/python3`,args:[`${e}/hexstrike_mcp.py`,`--server`,`http://127.0.0.1:8888`,`--profile`,`full`],timeout:300}}},null,2)}];function Cn(){let[e,t]=(0,W.useState)(`claude`),[n,r]=(0,W.useState)(`/path/to/hexstrike-ai-community-edition`),[i,a]=(0,W.useState)(!1),o=Sn.find(t=>t.id===e)??Sn[0];return(0,W.useEffect)(()=>{let e=!0;return J.getSettings().then(t=>{let n=t.settings.server.working_dir?.trim();!e||!n||(r(n),a(!0))}).catch(()=>{e&&a(!1)}),()=>{e=!1}},[]),(0,Z.jsxs)(`div`,{className:`help-page`,children:[(0,Z.jsx)(vn,{installPath:n,setInstallPath:r,pathDetected:i,activeIde:e,setActiveIde:t,ideConfigs:Sn,selectedIde:o}),(0,Z.jsx)(yn,{}),(0,Z.jsx)(bn,{}),(0,Z.jsx)(xn,{})]})}function wn(e){let[t,n]=(0,W.useState)(e?.processes??null),[r,i]=(0,W.useState)(e?{workers:4,queued:2,completed:38}:null),[a,o]=(0,W.useState)(!e),[s,c]=(0,W.useState)(null),[l,u]=(0,W.useState)(null),[d,f]=(0,W.useState)(e?`polling`:`streaming`),p=(0,W.useRef)(null),m=(0,W.useCallback)(async()=>{if(!e)try{let[e,t]=await Promise.all([J.processDashboard(),J.processPoolStats()]);n(e),i(t),c(null)}catch(e){c(String(e))}finally{o(!1)}},[e]);(0,W.useEffect)(()=>{if(e)return;let t=null,r=null;function a(){t?.close(),r?.close()}function s(){p.current||=(f(`polling`),setInterval(m,3e3))}try{t=J.processDashboardStream(),r=J.processPoolStatsStream();let e=!1,l=!1;t.onopen=()=>{e=!0,e&&l&&f(`streaming`)},r.onopen=()=>{l=!0,e&&l&&f(`streaming`)};let u=()=>{f(`error`),a(),s()};t.onerror=u,r.onerror=u,t.onmessage=e=>{try{n(JSON.parse(e.data)),c(null),o(!1)}catch{c(`Stream parse error`)}},r.onmessage=e=>{try{i(JSON.parse(e.data)),c(null)}catch{c(`Pool stats stream error`)}}}catch{f(`error`),s()}return()=>{a(),p.current&&=(clearInterval(p.current),null)}},[e,m]),(0,W.useEffect)(()=>{if(!e){if(d!==`polling`){p.current&&=(clearInterval(p.current),null);return}return m(),p.current||=setInterval(m,3e3),()=>{p.current&&=(clearInterval(p.current),null)}}},[e,m,d]);let h=(0,W.useCallback)(async(e,t)=>{try{let n=await e();u(n.success?n.message??`${t} OK`:n.error??`${t} failed`)}catch(e){u(String(e))}setTimeout(()=>u(null),3e3),await m(),t===`Terminated`&&n(e=>e&&{...e,processes:e.processes.filter(e=>e.status!==`terminated`)})},[m]);return{data:t,poolStats:r,loading:a,error:s,actionMsg:l,streamStatus:d,fetchData:m,pauseProcess:e=>h(()=>J.pauseProcess(e),`Paused`),resumeProcess:e=>h(()=>J.resumeProcess(e),`Resumed`),terminateProcess:e=>h(()=>J.terminateProcess(e),`Terminated`)}}function Tn({data:e,streamStatus:t}){let n=e?.system_load,r=t===`streaming`?`Live`:t===`polling`?`Polling`:`Offline`,i=t===`streaming`?`var(--green)`:t===`polling`?`var(--blue)`:`var(--red)`,a=e?.timestamp?.slice(11,19)??`n/a`;return(0,Z.jsxs)(`div`,{className:`kpi-row`,children:[(0,Z.jsx)(Q,{icon:(0,Z.jsx)(M,{size:20}),label:`Worker Pool`,value:r,sub:`updated ${a}`,accent:i}),(0,Z.jsx)(Q,{icon:(0,Z.jsx)(l,{size:20}),label:`CPU`,value:n?`${n.cpu_percent.toFixed(1)}%`:`n/a`,sub:`system load`,accent:`var(--green)`}),(0,Z.jsx)(Q,{icon:(0,Z.jsx)(V,{size:20}),label:`Memory`,value:n?`${n.memory_percent.toFixed(1)}%`:`n/a`,sub:`system memory`,accent:`var(--blue)`}),(0,Z.jsx)(Q,{icon:(0,Z.jsx)(F,{size:20}),label:`Connections`,value:n?n.active_connections:`n/a`,sub:`active sockets`,accent:`var(--amber)`})]})}function En({processes:e,actionMsg:t,onRefresh:n,onPause:r,onResume:i,onTerminate:a}){return(0,Z.jsxs)(`section`,{className:`section`,children:[(0,Z.jsxs)(`div`,{className:`section-header`,children:[(0,Z.jsxs)(`h3`,{children:[`Active Processes `,(0,Z.jsx)(`span`,{className:`badge`,children:e.length})]}),(0,Z.jsxs)(`div`,{style:{display:`flex`,gap:8,alignItems:`center`},children:[t&&(0,Z.jsx)(`span`,{className:`section-meta`,style:{color:`var(--amber)`},children:t}),(0,Z.jsx)(`button`,{className:`icon-btn`,onClick:n,title:`Refresh`,children:(0,Z.jsx)(U,{size:14})})]})]}),e.length===0?(0,Z.jsxs)(`div`,{className:`tasks-empty`,children:[(0,Z.jsx)(o,{size:32,color:`var(--text-dim)`}),(0,Z.jsx)(`p`,{children:`No active processes`})]}):(0,Z.jsxs)(`div`,{className:`tasks-table`,children:[(0,Z.jsxs)(`div`,{className:`tasks-thead`,children:[(0,Z.jsx)(`span`,{children:`PID`}),(0,Z.jsx)(`span`,{children:`Command`}),(0,Z.jsx)(`span`,{children:`Status`}),(0,Z.jsx)(`span`,{children:`Progress`}),(0,Z.jsx)(`span`,{children:`Runtime`}),(0,Z.jsx)(`span`,{children:`ETA`}),(0,Z.jsx)(`span`,{children:`Actions`})]}),e.map(e=>{let t=parseFloat(e.progress_percent)||0,n=e.status===`running`?`var(--green)`:e.status===`paused`?`var(--amber)`:`var(--text-dim)`;return(0,Z.jsxs)(`div`,{className:`tasks-row`,children:[(0,Z.jsx)(`span`,{className:`mono tasks-pid`,children:e.pid}),(0,Z.jsx)(`span`,{className:`mono tasks-cmd`,title:e.command,children:e.command}),(0,Z.jsx)(`span`,{className:`tasks-status tasks-status--${e.status}`,children:e.status}),(0,Z.jsxs)(`div`,{className:`tasks-progress`,children:[(0,Z.jsx)(`div`,{className:`tasks-progress-bar-bg`,children:(0,Z.jsx)(`div`,{className:`tasks-progress-bar-fill`,style:{width:`${Math.min(100,t)}%`,background:n}})}),(0,Z.jsx)(`span`,{className:`tasks-pct mono`,children:e.progress_percent})]}),(0,Z.jsx)(`span`,{className:`mono`,children:e.runtime}),(0,Z.jsx)(`span`,{className:`mono`,children:e.eta}),(0,Z.jsxs)(`div`,{className:`tasks-actions`,children:[e.status!==`paused`&&(0,Z.jsx)(`button`,{className:`tasks-btn tasks-btn--pause`,title:`Pause`,onClick:()=>r(e.pid),children:(0,Z.jsx)(E,{size:14})}),e.status===`paused`&&(0,Z.jsx)(`button`,{className:`tasks-btn tasks-btn--resume`,title:`Resume`,onClick:()=>i(e.pid),children:(0,Z.jsx)(d,{size:14})}),(0,Z.jsx)(`button`,{className:`tasks-btn tasks-btn--stop`,title:`Terminate`,onClick:()=>a(e.pid),children:(0,Z.jsx)(f,{size:14})})]})]},e.pid)})]})]})}function Dn({demoData:e}){let{data:t,poolStats:n,loading:r,error:i,actionMsg:a,streamStatus:o,fetchData:s,pauseProcess:l,resumeProcess:u,terminateProcess:d}=wn(e);if(r&&!t)return(0,Z.jsxs)(`div`,{className:`loading-state`,children:[(0,Z.jsx)(U,{size:20,className:`spin`,color:`var(--green)`}),(0,Z.jsx)(`p`,{children:`Loading tasks…`})]});if(i&&!t)return(0,Z.jsxs)(`div`,{className:`error-banner`,children:[(0,Z.jsx)(c,{size:16}),` `,i]});let f=(t?.processes??[]).slice(-100);return(0,Z.jsxs)(`div`,{className:`page-content`,children:[n&&(0,Z.jsx)(Tn,{data:t,streamStatus:o}),(0,Z.jsx)(En,{processes:f,actionMsg:a,onRefresh:s,onPause:l,onResume:u,onTerminate:d})]})}function On({tools:e,filtered:t,categories:n,activeCat:r,setActiveCat:i,search:a,setSearch:o,missingOnly:s,setMissingOnly:l,missingCount:u,toolsStatus:d,onSelectTool:f}){return(0,Z.jsxs)(`section`,{className:`section`,children:[(0,Z.jsx)(`div`,{className:`section-header`,children:(0,Z.jsxs)(`h3`,{children:[`Tool Registry `,(0,Z.jsxs)(`span`,{className:`badge`,children:[t.length,` / `,e.length]})]})}),(0,Z.jsxs)(`div`,{className:`registry-controls`,children:[(0,Z.jsxs)(`div`,{className:`registry-controls-top`,children:[(0,Z.jsxs)(`div`,{className:`search-input-wrap`,children:[(0,Z.jsx)(`input`,{className:`search-input mono`,placeholder:`Search tools…`,value:a,onChange:e=>o(e.target.value)}),a.trim().length>0&&(0,Z.jsx)(`button`,{className:`search-clear-btn`,onClick:()=>o(``),title:`Clear search`,"aria-label":`Clear search`,children:(0,Z.jsx)(c,{size:12})})]}),(0,Z.jsxs)(`button`,{className:`registry-missing-toggle${s?` active`:``}`,onClick:()=>l(e=>!e),title:`Show only tools not installed`,children:[(0,Z.jsx)(c,{size:12}),`Not installed`,u>0&&(0,Z.jsx)(`span`,{className:`badge`,children:u})]})]}),(0,Z.jsx)(`div`,{className:`cat-tabs`,children:n.map(e=>(0,Z.jsx)(`button`,{className:`cat-tab ${r===e?`active`:``}`,onClick:()=>i(r===e?`all`:e),children:e.replace(/_/g,` `)},e))})]}),(0,Z.jsxs)(`div`,{className:`registry-grid registry-grid--wide`,children:[t.map(e=>(0,Z.jsxs)(`div`,{className:`registry-card registry-card--clickable`,onClick:()=>f(e),title:`Click for details on ${e.name}`,children:[(0,Z.jsxs)(`div`,{className:`registry-card-top`,children:[(0,Z.jsx)(`span`,{className:`registry-name mono`,title:e.name,children:e.name}),(0,Z.jsxs)(`div`,{className:`registry-meta`,children:[e.parent_tool&&(0,Z.jsx)(`span`,{className:`registry-cat`,title:`Based on ${e.parent_tool}`,children:e.parent_tool}),d[e.name]===!0&&(0,Z.jsx)(`span`,{className:`registry-installed`,title:`Installed`,children:(0,Z.jsx)(T,{size:11,color:`var(--green)`})}),d[e.name]===!1&&(0,Z.jsx)(`span`,{className:`registry-installed`,title:`Not installed`,children:(0,Z.jsx)(c,{size:11,color:`var(--red)`})})]})]}),(0,Z.jsx)(`p`,{className:`registry-desc`,children:e.desc}),(0,Z.jsxs)(`div`,{className:`registry-footer`,children:[(0,Z.jsx)(`span`,{className:`registry-cat`,children:e.category.replace(/_/g,` `)}),(0,Z.jsxs)(`span`,{className:`registry-eff`,title:`Effectiveness`,children:[`█`.repeat(Math.round(e.effectiveness*5)),`░`.repeat(5-Math.round(e.effectiveness*5))]})]})]},e.name)),t.length===0&&(0,Z.jsx)(`p`,{className:`empty-state`,children:`No tools match your filter.`})]})]})}function kn({health:e,tools:t,toolsStatus:n}){let[r,i]=(0,W.useState)(``),[a,o]=(0,W.useState)(`all`),[s,l]=(0,W.useState)(null),[u,d]=(0,W.useState)(!1),f=It(t),p=Lt(t,{toolsStatus:n,activeCategory:a,search:r,missingOnly:u,includeParentToolSearch:!0}),m=e.total_tools_count-e.total_tools_available;return(0,Z.jsxs)(`div`,{className:`page-content tools-page`,children:[s&&(0,Z.jsx)(Tt,{tool:s,onClose:()=>l(null),installed:n[s.name]}),(0,Z.jsxs)(`div`,{className:`kpi-row`,children:[(0,Z.jsx)(Q,{icon:(0,Z.jsx)(z,{size:20}),label:`Total Server Tools`,value:t.length,sub:`in registry`,accent:`var(--blue)`}),(0,Z.jsx)(Q,{icon:(0,Z.jsx)(se,{size:20}),label:`Kali Tools Installed`,value:`${e.total_tools_available} / ${e.total_tools_count}`,sub:`${(e.total_tools_available/Math.max(e.total_tools_count,1)*100).toFixed(0)}% coverage`,accent:`var(--green)`}),(0,Z.jsx)(Q,{icon:(0,Z.jsx)(c,{size:20}),label:`Missing`,value:m,sub:`not installed`,accent:m>0?`var(--amber)`:`var(--text-dim)`}),(0,Z.jsx)(Q,{icon:(0,Z.jsx)(C,{size:20}),label:`Categories`,value:f.length-1,sub:`tool categories`,accent:`var(--purple)`})]}),(0,Z.jsx)(On,{tools:t,filtered:p,categories:f,activeCat:a,setActiveCat:o,search:r,setSearch:i,missingOnly:u,setMissingOnly:d,missingCount:m,toolsStatus:n,onSelectTool:l})]})}function An(e){for(let t of[`target`,`url`,`host`,`ip`,`domain`,`file`]){let n=e.params[t];if(n)return String(n)}let t=Object.values(e.params)[0];return t?String(t):`(no target)`}function jn(e){let t=new Map;for(let n of e){let e=n.ts.toLocaleDateString(`en-GB`,{day:`2-digit`,month:`short`,year:`numeric`});t.has(e)||t.set(e,[]),t.get(e)?.push(n)}return Array.from(t.entries()).map(([e,t])=>({label:e,entries:t}))}function Mn(e){let t=e.filter(e=>e.result.success).length,n=e.reduce((e,t)=>e+t.result.execution_time,0)/e.length,r=e.reduce((e,t)=>e.ts>t.ts?e:t);return{total:e.length,ok:t,failed:e.length-t,avgTime:n,last:r}}function Nn({runHistory:e,onOpenEntry:t}){let n=[...e].sort((e,t)=>e.ts.getTime()-t.ts.getTime()).slice(-60),r=jn(n);return(0,Z.jsxs)(`section`,{className:`section`,children:[(0,Z.jsx)(`div`,{className:`section-header`,children:(0,Z.jsxs)(`h3`,{children:[`Run Timeline `,(0,Z.jsxs)(`span`,{className:`section-meta`,children:[`last `,n.length]})]})}),(0,Z.jsx)(`div`,{className:`reports-timeline-wrap`,children:r.map(({label:e,entries:n})=>(0,Z.jsxs)(`div`,{className:`reports-timeline-group`,children:[(0,Z.jsx)(`span`,{className:`reports-timeline-date`,children:e}),(0,Z.jsx)(`div`,{className:`reports-timeline-dots`,children:n.map((e,n)=>(0,Z.jsx)(`button`,{className:`reports-timeline-dot ${e.result.success?`ok`:`fail`}`,title:`${e.tool} — ${e.ts.toLocaleTimeString(`en-GB`)} — ${e.result.success?`ok`:`failed`} (${e.result.execution_time.toFixed(1)}s)`,onClick:()=>t(e)},n))})]},e))})]})}function Pn({grouped:e,keys:n,groupBy:r,search:i,setSearch:a,setGroupBy:o,expanded:s,toggleExpanded:c,onOpenEntry:l}){return(0,Z.jsxs)(`section`,{className:`section`,children:[(0,Z.jsxs)(`div`,{className:`section-header`,children:[(0,Z.jsx)(`h3`,{children:`Breakdown`}),(0,Z.jsxs)(`div`,{style:{display:`flex`,gap:8,alignItems:`center`},children:[(0,Z.jsx)(`input`,{className:`search-input mono`,style:{width:180},placeholder:`Search…`,value:i,onChange:e=>a(e.target.value)}),(0,Z.jsx)(`button`,{className:`cat-tab ${r===`tool`?`active`:``}`,onClick:()=>o(`tool`),children:`By Tool`}),(0,Z.jsx)(`button`,{className:`cat-tab ${r===`target`?`active`:``}`,onClick:()=>o(`target`),children:`By Target`})]})]}),(0,Z.jsxs)(`div`,{className:`reports-table`,children:[(0,Z.jsxs)(`div`,{className:`reports-thead`,children:[(0,Z.jsx)(`span`,{}),(0,Z.jsx)(`span`,{children:r===`tool`?`Tool`:`Target`}),(0,Z.jsx)(`span`,{children:`Runs`}),(0,Z.jsx)(`span`,{children:`Success`}),(0,Z.jsx)(`span`,{children:`Failed`}),(0,Z.jsx)(`span`,{children:`Avg Time`}),(0,Z.jsx)(`span`,{children:`Last Run`})]}),n.map(n=>{let r=e[n],i=Mn(r),a=i.ok/i.total*100,o=a>=80?`var(--green)`:a>=50?`var(--amber)`:`var(--red)`,u=s.has(n),d=[...r].sort((e,t)=>t.ts.getTime()-e.ts.getTime());return(0,Z.jsxs)(`div`,{className:`reports-group`,children:[(0,Z.jsxs)(`button`,{className:`reports-row reports-row--clickable`,onClick:()=>c(n),children:[(0,Z.jsx)(`span`,{className:`reports-chevron`,children:u?(0,Z.jsx)(N,{size:12}):(0,Z.jsx)(t,{size:12})}),(0,Z.jsx)(`span`,{className:`mono reports-key`,children:n}),(0,Z.jsx)(`span`,{className:`mono`,children:i.total}),(0,Z.jsx)(`span`,{className:`mono`,style:{color:`var(--green)`},children:i.ok}),(0,Z.jsx)(`span`,{className:`mono`,style:{color:i.failed>0?`var(--red)`:`var(--text-dim)`},children:i.failed}),(0,Z.jsxs)(`span`,{className:`mono`,children:[i.avgTime.toFixed(1),`s`]}),(0,Z.jsxs)(`div`,{className:`reports-last-cell`,children:[(0,Z.jsx)(`span`,{className:`reports-pct-bar-bg`,children:(0,Z.jsx)(`span`,{className:`reports-pct-bar-fill`,style:{width:`${a}%`,background:o}})}),(0,Z.jsxs)(`span`,{className:`mono`,style:{fontSize:11,color:`var(--text-dim)`},children:[i.last.ts.toLocaleDateString(`en-GB`),` `,i.last.ts.toLocaleTimeString(`en-GB`)]})]})]}),u&&(0,Z.jsx)(`div`,{className:`reports-runs`,children:d.map((e,t)=>(0,Z.jsxs)(`button`,{className:`reports-run-row`,onClick:()=>l(e),children:[(0,Z.jsx)(`span`,{className:`reports-run-dot ${e.result.success?`ok`:`fail`}`}),(0,Z.jsx)(`span`,{className:`mono reports-run-tool`,children:e.tool}),(0,Z.jsxs)(`span`,{className:`mono reports-run-time`,children:[e.ts.toLocaleDateString(`en-GB`),` `,e.ts.toLocaleTimeString(`en-GB`)]}),(0,Z.jsxs)(`span`,{className:`mono reports-run-duration`,children:[e.result.execution_time.toFixed(1),`s`]}),Object.entries(e.params).map(([e,t])=>(0,Z.jsxs)(`span`,{className:`mono reports-run-param`,children:[e,`=`,(0,Z.jsx)(`em`,{children:String(t)})]},e))]},t))})]},n)})]})]})}function Fn({runHistory:e}){let[t,n]=(0,W.useState)(`tool`),[r,i]=(0,W.useState)(``),[a,o]=(0,W.useState)(new Set),[s,c]=(0,W.useState)(null),l=e.reduce((e,t)=>((e[t.tool]=e[t.tool]||[]).push(t),e),{}),d=e.reduce((e,t)=>{let n=An(t);return(e[n]=e[n]||[]).push(t),e},{}),f=t===`tool`?l:d;function p(e){o(t=>{let n=new Set(t);return n.has(e)?n.delete(e):n.add(e),n})}let m=r.toLowerCase(),h=Object.keys(f).filter(e=>!m||e.toLowerCase().includes(m)).sort();return e.length===0?(0,Z.jsx)(`div`,{className:`page-content`,children:(0,Z.jsxs)(`div`,{className:`tasks-empty`,children:[(0,Z.jsx)(u,{size:32,color:`var(--text-dim)`}),(0,Z.jsx)(`p`,{children:`No run history yet. Execute tools from the Run tab to see reports.`})]})}):(0,Z.jsxs)(`div`,{className:`page-content`,children:[s&&(0,Z.jsx)(jt,{entry:s,onClose:()=>c(null)}),(0,Z.jsxs)(`div`,{className:`kpi-row`,children:[(0,Z.jsx)(Q,{icon:(0,Z.jsx)(R,{size:20}),label:`Total Runs`,value:e.length,sub:`all time`,accent:`var(--blue)`}),(0,Z.jsx)(Q,{icon:(0,Z.jsx)(T,{size:20}),label:`Success Rate`,value:`${(e.filter(e=>e.result.success).length/e.length*100).toFixed(0)}%`,sub:`${e.filter(e=>e.result.success).length} ok · ${e.filter(e=>!e.result.success).length} failed`,accent:`var(--green)`}),(0,Z.jsx)(Q,{icon:(0,Z.jsx)(w,{size:20}),label:`Avg Time`,value:`${(e.reduce((e,t)=>e+t.result.execution_time,0)/e.length).toFixed(1)}s`,sub:`per run`,accent:`var(--purple)`}),(0,Z.jsx)(Q,{icon:(0,Z.jsx)(ee,{size:20}),label:`Unique Tools`,value:Object.keys(l).length,sub:`used`,accent:`var(--amber)`})]}),(0,Z.jsx)(Nn,{runHistory:e,onOpenEntry:c}),(0,Z.jsx)(Pn,{grouped:f,keys:h,groupBy:t,search:r,setSearch:i,setGroupBy:n,expanded:a,toggleExpanded:p,onOpenEntry:c})]})}var In=[{key:`intelligence`,title:`Intelligent Profiling`,description:`Automated target profiling and attack chain generation.`,details:`Best for quick, comprehensive insights on a target.`,modalDescription:`Leverages AI to analyze the target and generate a customized attack chain. This mode is ideal when you want a smart, efficient workflow that adapts to the target’s unique characteristics, providing actionable insights with minimal setup.`,tools:[`analyze-target`,`smart-scan`,`technology-detection`,`create-attack-chain`,`vulnx`,`httpx`],placeholder:`Domain or org target (target.tld)`},{key:`comprehensive`,title:`Comprehensive Assessment`,description:`Balanced full-chain workflow from recon to vulnerability checks.`,details:`Best default for unknown targets.`,modalDescription:`Builds a broad, practical workflow that starts with target profiling and surface mapping, then moves into prioritized vulnerability validation. This is designed for cases where you want full context and a structured path from discovery to actionable findings.`,tools:[`nmap`,`httpx`,`katana`,`nuclei`,`nikto`,`ffuf`,`testssl`,`arjun`],placeholder:`Target URL/domain/IP (example.com)`},{key:`reconnaissance`,title:`Reconnaissance`,description:`Discovery-first workflow for assets, endpoints, and technologies.`,details:`Use when mapping attack surface.`,modalDescription:`Focuses on enumeration and intelligence gathering with minimal intrusive testing. It maps hosts, services, paths, and technologies so you can decide where deeper testing should happen next.`,tools:[`subfinder`,`amass`,`httpx`,`katana`,`waybackurls`,`gau`,`arjun`,`whatweb`,`paramspider`,`wafw00f`],placeholder:`Scope target (example.com or 10.0.0.0/24)`},{key:`vulnerability_hunting`,title:`Vulnerability Hunting`,description:`Vulnerability-focused chain prioritizing exploitable findings.`,details:`Use when recon is already known.`,modalDescription:`Runs targeted security checks against known attack surface to quickly identify high-value weaknesses. This mode biases toward validating likely vulnerabilities and producing results you can triage and act on fast.`,tools:[`nuclei`,`sqlmap`,`dalfox`,`jaeles`,`nikto`,`wpscan`,`arjun`,`jwt-analyzer`,`graphql-scanner`,`zaproxy`],placeholder:`Web/API target (https://target.tld)`},{key:`OSINT`,title:`OSINT Collection`,description:`Intelligence and external footprint gathering for a target.`,details:`Useful before active probing.`,modalDescription:`Collects passive intelligence from public sources to understand exposure before active scanning. This includes historical URLs, publicly indexed assets, and reconnaissance artifacts useful for planning follow-up testing.`,tools:[`theHarvester`,`subfinder`,`amass`,`waymore`,`waybackurls`,`whois`,`gau`,`dnsenum`],placeholder:`Domain or org target (target.tld)`},{key:`api_security`,title:`API Security`,description:`API-focused workflow for endpoint, auth, and parameter weakness detection.`,details:`Best for REST/GraphQL services and modern API stacks.`,modalDescription:`Targets API attack surfaces end-to-end: endpoint discovery, hidden parameter mining, schema analysis, auth/token checks, and vulnerability validation. Use this when APIs are primary scope or where mobile/backend services expose extensive routes.`,tools:[`httpx`,`katana`,`arjun`,`x8`,`paramspider`,`api-schema-analyzer`,`graphql-scanner`,`jwt-analyzer`,`nuclei`,`sqlmap`,`ffuf`],placeholder:`API base URL (https://api.target.tld)`},{key:`internal_network_ad`,title:`Internal Network + AD`,description:`Internal host discovery with SMB/AD enumeration and lateral movement checks.`,details:`Best for enterprise network and Active Directory testing.`,modalDescription:`Builds an internal assessment chain focused on host/service discovery, SMB/NetBIOS/RPC enumeration, domain intel extraction, and AD-aware access checks. Use for corp ranges, domain-joined segments, and post-foothold validation.`,tools:[`nmap_advanced`,`autorecon`,`enum4linux-ng`,`smbmap`,`rpcclient`,`nxc`,`ldapdomaindump`,`impacket-ad-enum`,`responder`,`evil-winrm`],placeholder:`Internal target (10.0.0.15 or 10.0.0.0/24)`},{key:`manual`,title:`Manual Session`,description:`Start an empty session and add tools yourself.`,details:`Best when you want full manual control.`,modalDescription:`Creates a clean session with only target context. No predefined workflow is added, so you can build your own tool chain step-by-step from the session detail page.`,tools:[],placeholder:`Target URL/domain/IP (example.com)`},{key:`from_template`,title:`From Template`,description:`Start from a saved tool template.`,details:`Reuse your recurring workflows.`,modalDescription:`Creates a session from an existing template. You only set target and choose the saved template; all template tools are copied into the new session.`,tools:[],placeholder:`Target URL/domain/IP (example.com)`}];function Ln(e){let t=e.metadata??{},n=t.session_name;if(typeof n==`string`&&n.trim()){let e=n.trim();return e.charAt(0).toUpperCase()+e.slice(1)}let r=typeof e.objective==`string`&&e.objective||(typeof t.mode==`string`?t.mode:``);if(r){let e=r.replace(/_/g,` `).trim();return e.charAt(0).toUpperCase()+e.slice(1)}return`Session`}function Rn({session:e,onOpen:t}){let n=e.metadata?.tool_status??{},r=e.metadata?.last_run??null;return(0,Z.jsxs)(`div`,{className:`session-card session-card--compact registry-card--clickable`,onClick:()=>t(e.session_id),children:[(0,Z.jsxs)(`div`,{className:`session-card-header`,children:[(0,Z.jsxs)(`div`,{className:`session-target`,children:[(0,Z.jsx)(te,{size:13,color:`var(--blue)`}),(0,Z.jsx)(`span`,{className:`mono`,children:e.target})]}),e.status&&(0,Z.jsx)(`span`,{className:`session-status session-status--${e.status}`,children:e.status})]}),(0,Z.jsxs)(`div`,{className:`session-card-meta`,children:[(0,Z.jsxs)(`span`,{children:[(0,Z.jsx)(M,{size:11}),` `,e.total_findings,` findings`]}),(0,Z.jsxs)(`span`,{children:[(0,Z.jsx)(U,{size:11}),` `,e.iterations,` iterations`]}),(0,Z.jsxs)(`span`,{children:[(0,Z.jsx)(w,{size:11}),` `,dt(e.updated_at)]})]}),(0,Z.jsxs)(`div`,{className:`session-tools`,children:[e.tools_executed.slice(0,7).map(e=>(0,Z.jsx)(`span`,{className:`session-tool-chip mono session-tool-chip--${n[e]===`success`?`success`:n[e]===`failed`?`failed`:`idle`}`,children:e},e)),e.tools_executed.length>7&&(0,Z.jsxs)(`span`,{className:`session-tool-chip session-tool-chip--more`,children:[`+`,e.tools_executed.length-7]})]}),(0,Z.jsxs)(`div`,{className:`session-card-footer`,children:[(0,Z.jsx)(`span`,{className:`session-id mono`,children:e.session_id}),(0,Z.jsx)(`span`,{className:`session-tool-chip mono`,children:Ln(e)})]}),r?.tool&&(0,Z.jsxs)(`div`,{className:`session-last-run mono`,children:[`last: `,r.tool,` | `,r.success?`OK`:`FAIL`,` | exit `,r.return_code??0,` | `,(r.execution_time??0).toFixed(2),`s`]})]})}function zn({startModes:e,onOpenStartMode:t,createMsg:n}){return(0,Z.jsxs)(`section`,{className:`section`,children:[(0,Z.jsxs)(`div`,{className:`section-header`,children:[(0,Z.jsx)(`h3`,{children:`Start Session`}),(0,Z.jsx)(`span`,{className:`section-meta`,children:`Choose a workflow type, then provide target details.`})]}),(0,Z.jsx)(`div`,{className:`registry-grid registry-grid--wider start-mode-grid`,children:e.map(e=>(0,Z.jsxs)(`div`,{className:`registry-card registry-card--clickable start-mode-card`,onClick:()=>t(e),children:[(0,Z.jsxs)(`div`,{className:`registry-card-top`,children:[(0,Z.jsx)(`span`,{className:`registry-name`,children:e.title}),(0,Z.jsx)(`span`,{className:`registry-cat`,children:e.key.replace(/_/g,` `)})]}),(0,Z.jsx)(`p`,{className:`registry-desc`,children:e.description}),(0,Z.jsx)(`p`,{className:`start-mode-detail`,children:e.details})]},e.key))}),n&&(0,Z.jsx)(`p`,{className:`section-meta`,children:n})]})}function Bn({title:e,sessions:t,emptyText:n,onOpenSession:r,headerRight:i,onHeaderClick:a,footer:o}){return(0,Z.jsxs)(`section`,{className:`section`,children:[(0,Z.jsxs)(`div`,{className:`section-header${a?` sessions-collapsed-toggle`:``}`,...a?{role:`button`,tabIndex:0,onClick:a,onKeyDown:e=>{(e.key===`Enter`||e.key===` `)&&(e.preventDefault(),a())}}:{},children:[(0,Z.jsxs)(`h3`,{children:[e,` `,(0,Z.jsx)(`span`,{className:`badge`,children:t.length})]}),i]}),t.length===0?(0,Z.jsxs)(`div`,{className:`tasks-empty`,children:[(0,Z.jsx)(g,{size:28,color:`var(--text-dim)`}),(0,Z.jsx)(`p`,{children:n})]}):(0,Z.jsx)(`div`,{className:`sessions-grid`,children:t.map(e=>(0,Z.jsx)(Rn,{session:e,onOpen:r},e.session_id))}),o&&(0,Z.jsx)(`div`,{className:`section-meta session-list-footer-tip`,children:o})]})}function Vn({startMode:e,templates:t,selectedTemplateId:n,setSelectedTemplateId:r,modalTarget:i,setModalTarget:a,modalNote:o,setModalNote:s,modalError:c,creatingSession:l,onClose:u,onSubmit:d}){let f=(0,W.useRef)(null),p=(0,W.useRef)(null),m=(0,W.useMemo)(()=>t.find(e=>e.template_id===n),[t,n]),h=(0,W.useMemo)(()=>{if(e.key!==`from_template`||!m)return e.tools;let t=Array.isArray(m.workflow_steps)?m.workflow_steps:[],n=[];for(let e of t){let t=e?.tool;!t||n.includes(t)||n.push(t)}return n},[e,m]);return(0,W.useEffect)(()=>{let t=window.requestAnimationFrame(()=>{if(e.key===`from_template`){f.current?.focus();return}p.current?.focus()});return()=>window.cancelAnimationFrame(t)},[e.key]),(0,Z.jsx)(`div`,{className:`modal-backdrop`,onClick:e=>{e.target===e.currentTarget&&u()},children:(0,Z.jsxs)(`div`,{className:`modal`,children:[(0,Z.jsxs)(`div`,{className:`modal-header`,children:[(0,Z.jsx)(`div`,{className:`modal-title-row`,children:(0,Z.jsxs)(`span`,{className:`modal-name`,children:[`Start `,e.title]})}),(0,Z.jsx)(`button`,{className:`modal-close`,onClick:u,children:`x`})]}),(0,Z.jsxs)(`div`,{className:`modal-body`,children:[(0,Z.jsx)(`p`,{className:`modal-desc`,children:e.modalDescription}),(0,Z.jsxs)(`div`,{className:`modal-section`,children:[(0,Z.jsx)(`span`,{className:`modal-label`,children:`Typical Tooling`}),(0,Z.jsxs)(`div`,{className:`modal-params`,children:[h.length===0&&(0,Z.jsx)(`span`,{className:`modal-param mono`,children:`none preloaded`}),h.map(e=>(0,Z.jsx)(`span`,{className:`modal-param mono`,children:e},e))]})]}),e.key===`from_template`&&(0,Z.jsxs)(`div`,{className:`session-start-form`,children:[(0,Z.jsx)(`label`,{className:`mono`,children:`Template *`}),(0,Z.jsxs)(`select`,{ref:f,className:`session-objective-select`,value:n,onChange:e=>r(e.target.value),children:[(0,Z.jsx)(`option`,{value:``,children:`Select template`}),t.map(e=>(0,Z.jsx)(`option`,{value:e.template_id,children:e.name},e.template_id))]})]}),(0,Z.jsxs)(`div`,{className:`session-start-form`,children:[(0,Z.jsx)(`label`,{className:`mono`,htmlFor:`session-target-input`,children:`Target *`}),(0,Z.jsx)(`input`,{id:`session-target-input`,ref:p,className:`search-input mono`,value:i,onChange:e=>a(e.target.value),placeholder:e.placeholder}),(0,Z.jsx)(`label`,{className:`mono`,htmlFor:`session-note-input`,children:`Note (optional)`}),(0,Z.jsx)(`textarea`,{id:`session-note-input`,className:`session-step-params mono`,rows:3,value:o,onChange:e=>s(e.target.value),placeholder:`Context for this run`}),c&&(0,Z.jsx)(`div`,{className:`run-error`,children:c}),(0,Z.jsxs)(`div`,{className:`session-start-actions`,children:[(0,Z.jsx)(`button`,{className:`session-action-btn`,onClick:u,children:`Cancel`}),(0,Z.jsxs)(`button`,{className:`session-run-btn`,onClick:d,disabled:l,children:[l?(0,Z.jsx)(U,{size:13,className:`spin`}):(0,Z.jsx)(te,{size:13}),l?`Starting…`:`Start Session`]})]})]})]})]})})}function Hn({demoData:e,onOpenSession:t}){let{pushToast:n}=et(),[r,i]=(0,W.useState)(e?.sessions??null),[a,o]=(0,W.useState)(!1),[s,l]=(0,W.useState)([]),[u,d]=(0,W.useState)([]),[f,p]=(0,W.useState)(null),[m,h]=(0,W.useState)(null),[_,v]=(0,W.useState)(null),[y,x]=(0,W.useState)(null),[S,C]=(0,W.useState)(null),[E,D]=(0,W.useState)(``),[O,k]=(0,W.useState)(``),[A,j]=(0,W.useState)(``),[ee,N]=(0,W.useState)(null),[P,ne]=(0,W.useState)(``),[re,F]=(0,W.useState)(``),[I,L]=(0,W.useState)([]),[oe,se]=(0,W.useState)(``),[ce,R]=(0,W.useState)(null),[le,ue]=(0,W.useState)(!1),[de,fe]=(0,W.useState)(!e),[z,B]=(0,W.useState)(null),[pe,me]=(0,W.useState)(!0),[he,ge]=(0,W.useState)(!1),[_e,ve]=(0,W.useState)(!0),[V,ye]=(0,W.useState)(e?`polling`:`streaming`),be=(0,W.useRef)(null),xe=(0,W.useRef)(null),Se=(0,W.useRef)(null),Ce=(0,W.useRef)(0),we=(0,W.useRef)(!1),Te=(0,W.useRef)(!1);(0,W.useEffect)(()=>{e||Promise.all([J.sessions(),J.tools(),J.sessionTemplates()]).then(([e,t,n])=>{i(e),l(t.tools??[]),d(n.templates??[]),B(null)}).catch(e=>B(String(e))).finally(()=>fe(!1))},[e]),(0,W.useEffect)(()=>{if(we.current||de||!r)return;let e=(r?.active??[]).filter(e=>(e.status??`active`)!==`completed`).length;me(e>0),ve(u.length>0),we.current=!0},[de,r,u.length]),(0,W.useEffect)(()=>{de||Te.current||u.length>0&&ve(!0)},[de,u.length]);async function Ee(){e||d((await J.sessionTemplates()).templates??[])}function De(t=!1){e||(t||fe(!0),J.sessions().then(e=>{i(e),B(null)}).catch(e=>B(String(e))).finally(()=>{t||fe(!1)}))}function Oe(t=!1){e||De(t)}(0,W.useEffect)(()=>{if(e)return;let t=null,n=!1;function r(){t&&t.close(),t=null,be.current=null}function a(){xe.current&&=(clearInterval(xe.current),null)}function o(){Se.current&&=(clearTimeout(Se.current),null)}function s(){a(),ye(`polling`),Oe(!0),xe.current=setInterval(()=>Oe(!0),5e3)}function c(e){o();let t=Ce.current+1;Ce.current=t;let r=Math.min(3e4,1e3*2**Math.max(0,t-1));Se.current=setTimeout(()=>{n||e()},r)}function l(){r();try{t=J.sessionsStream(),be.current=t,t.onopen=()=>{Ce.current=0,o(),a(),ye(`streaming`)},t.onmessage=e=>{try{let t=JSON.parse(e.data);t?.success&&(i(t),B(null),fe(!1))}catch{B(`Session stream parse error`)}},t.onerror=()=>{n||(ye(`error`),r(),s(),c(l))}}catch{if(n)return;ye(`error`),s(),c(l)}}return l(),()=>{n=!0,r(),a(),o()}},[e]);function ke(e,t=``){C(e),D(``),k(``),j(t),N(null)}function Ae(){C(null),N(null)}function je(e){ne(e.template_id),F(e.name),L(e.workflow_steps??[]),se(``),R(null),h(null)}function Me(){ne(``),F(``),L([]),se(``),R(null)}function Ne(e){let t=In.find(e=>e.key===`from_template`);t&&ke(t,e)}async function Pe(t){if(!e){v(t),h(null);try{await J.deleteSessionTemplate(t),d(e=>e.filter(e=>e.template_id!==t)),A===t&&j(``),p(`Template deleted.`),n(`success`,`Template deleted`)}catch(e){let t=String(e);h(t),n(`error`,`Delete failed: ${t}`)}finally{v(null),x(null)}}}function Fe(e){L(t=>[...t,{tool:e,parameters:{}}])}function G(e){L(t=>t.filter((t,n)=>n!==e))}async function K(){if(!P)return;let e=re.trim();if(!e){R(`Template name is required`),n(`error`,`Template name is required`);return}if(I.length===0){R(`Template must include at least one tool`),n(`error`,`Template must include at least one tool`);return}ue(!0),R(null),h(null);try{await J.updateSessionTemplate(P,{name:e,workflow_steps:I}),await Ee(),p(`Template updated.`),n(`success`,`Template updated`),Me()}catch(e){let t=String(e);R(t),n(`error`,`Save failed: ${t}`)}finally{ue(!1)}}async function Ie(e,t,r){if(!t.trim()){N(`Target is required`),n(`error`,`Target is required`);return}p(null),N(null),o(!0);try{let i=t.trim(),a,o=0;if(e.key===`manual`)a=await J.createSession({target:i,workflow_steps:[],source:`web`,objective:e.key,metadata:{origin:`ui/sessions/create`,mode:e.key,note:r,session_name:`Manual session`}});else if(e.key===`from_template`){let e=u.find(e=>e.template_id===A);if(!e){N(`Template is required`),n(`error`,`Template is required`);return}o=e.workflow_steps.length,a=await J.createSessionFromTemplate({target:i,template_id:e.template_id,source:`web`,objective:`from_template`,metadata:{origin:`ui/sessions/create`,mode:`from_template`,note:r,template_id:e.template_id,session_name:`Template: ${e.name}`}})}else{let t=await J.createAttackChain(i,e.key);o=t.attack_chain.steps.length,a=await J.createSession({target:i,workflow_steps:t.attack_chain.steps,source:`web`,objective:e.key,session_id:t.session_id,metadata:{origin:`ui/sessions/create`,mode:e.key,note:r}})}let s=a.session.session_id;p(e.key===`manual`?`Session created: ${s} (empty workflow, add tools manually).`:e.key===`from_template`?`Session created: ${s} (${o} template tool calls loaded).`:`Session created: ${s} (${o} tool calls ready).`),n(`success`,`Session created: ${s}`),Ae(),Oe()}catch(e){let t=String(e);N(t),p(t),n(`error`,`Session start failed: ${t}`)}finally{o(!1)}}if(de)return(0,Z.jsxs)(`div`,{className:`loading-state`,children:[(0,Z.jsx)(U,{size:20,className:`spin`,color:`var(--green)`}),(0,Z.jsx)(`p`,{children:`Loading sessions…`})]});if(z)return(0,Z.jsxs)(`div`,{className:`error-banner`,children:[(0,Z.jsx)(c,{size:16}),` `,z]});let Le=r?.active??[],q=Le.filter(e=>(e.status??`active`)!==`completed`),Re=[...r?.completed??[],...Le.filter(e=>(e.status??`active`)===`completed`)].filter((e,t,n)=>n.findIndex(t=>t.session_id===e.session_id)===t),ze=[...q,...Re].reduce((e,t)=>e+t.total_findings,0),Y=new Set([...q,...Re].map(e=>e.target)).size,Be=P?u.find(e=>e.template_id===P)??null:null,Ve=s.filter(e=>{if(!oe.trim())return!0;let t=oe.toLowerCase();return e.name.toLowerCase().includes(t)||e.desc.toLowerCase().includes(t)||e.category.toLowerCase().includes(t)}).slice(0,16);return(0,Z.jsxs)(`div`,{className:`page-content`,children:[(0,Z.jsxs)(`div`,{className:`kpi-row`,children:[(0,Z.jsx)(Q,{icon:(0,Z.jsx)(g,{size:20}),label:`Active Sessions`,value:q.length,sub:`in progress`,accent:q.length>0?`var(--green)`:`var(--text-dim)`}),(0,Z.jsx)(Q,{icon:(0,Z.jsx)(T,{size:20}),label:`Completed`,value:Re.length,sub:`archived`,accent:`var(--blue)`}),(0,Z.jsx)(Q,{icon:(0,Z.jsx)(M,{size:20}),label:`Total Findings`,value:ze,sub:`across all sessions`,accent:`var(--amber)`}),(0,Z.jsx)(Q,{icon:(0,Z.jsx)(te,{size:20}),label:`Unique Targets`,value:Y,sub:`scanned`,accent:`var(--purple)`})]}),(0,Z.jsx)(zn,{startModes:In,onOpenStartMode:ke,createMsg:f}),S&&(0,Z.jsx)(Vn,{startMode:S,templates:u,selectedTemplateId:A,setSelectedTemplateId:j,modalTarget:E,setModalTarget:D,modalNote:O,setModalNote:k,modalError:ee,creatingSession:a,onClose:Ae,onSubmit:()=>Ie(S,E,O)}),pe?(0,Z.jsx)(Bn,{title:`Active Sessions`,sessions:q,emptyText:`No active sessions. Start a session from target to run tools manually.`,onOpenSession:t,onHeaderClick:()=>me(!1),footer:(0,Z.jsxs)(Z.Fragment,{children:[(0,Z.jsx)(ie,{size:12}),`Call MCP tool `,(0,Z.jsx)(`span`,{className:`mono`,children:`handover_session("<session_id>", "optional note")`}),` to continue the session with AI.`]}),headerRight:(0,Z.jsxs)(`div`,{className:`sessions-header-actions`,children:[(0,Z.jsx)(`span`,{className:`sessions-stream-status sessions-stream-status--${V}`,children:V===`streaming`?`Live`:V===`polling`?`Polling`:`Offline`}),(0,Z.jsx)(`span`,{className:`session-help-btn`,children:`Hide active`})]})}):(0,Z.jsx)(`section`,{className:`section`,children:(0,Z.jsxs)(`div`,{className:`section-header sessions-collapsed-toggle`,role:`button`,tabIndex:0,onClick:()=>me(!0),onKeyDown:e=>{(e.key===`Enter`||e.key===` `)&&(e.preventDefault(),me(!0))},children:[(0,Z.jsxs)(`h3`,{children:[`Active Sessions `,(0,Z.jsx)(`span`,{className:`badge`,children:q.length})]}),(0,Z.jsx)(`span`,{className:`session-help-btn`,children:`Show active`})]})}),he?(0,Z.jsx)(Bn,{title:`Completed Sessions`,sessions:Re,emptyText:`No completed sessions yet.`,onOpenSession:t,onHeaderClick:()=>ge(!1),headerRight:(0,Z.jsx)(`span`,{className:`session-help-btn`,children:`Hide completed`})}):(0,Z.jsx)(`section`,{className:`section`,children:(0,Z.jsxs)(`div`,{className:`section-header sessions-collapsed-toggle`,role:`button`,tabIndex:0,onClick:()=>ge(!0),onKeyDown:e=>{(e.key===`Enter`||e.key===` `)&&(e.preventDefault(),ge(!0))},children:[(0,Z.jsxs)(`h3`,{children:[`Completed Sessions `,(0,Z.jsx)(`span`,{className:`badge`,children:Re.length})]}),(0,Z.jsx)(`span`,{className:`session-help-btn`,children:`Show completed`})]})}),_e?(0,Z.jsxs)(`section`,{className:`section`,children:[(0,Z.jsxs)(`div`,{className:`section-header sessions-collapsed-toggle`,role:`button`,tabIndex:0,onClick:()=>{Te.current=!0,ve(!1)},onKeyDown:e=>{(e.key===`Enter`||e.key===` `)&&(e.preventDefault(),Te.current=!0,ve(!1))},children:[(0,Z.jsxs)(`h3`,{children:[`Custom Templates `,(0,Z.jsx)(`span`,{className:`badge`,children:u.length})]}),(0,Z.jsx)(`span`,{className:`section-meta`,children:`Reusable user templates for starting sessions quickly.`}),(0,Z.jsx)(`span`,{className:`session-help-btn`,children:`Hide templates`})]}),m&&(0,Z.jsx)(`div`,{className:`run-error`,children:m}),u.length===0?(0,Z.jsxs)(`div`,{className:`tasks-empty`,children:[(0,Z.jsx)(g,{size:28,color:`var(--text-dim)`}),(0,Z.jsx)(`p`,{children:`No custom templates yet. Open a session and click Create Template.`})]}):(0,Z.jsx)(`div`,{className:`sessions-grid`,children:u.map(e=>(0,Z.jsxs)(`div`,{className:`session-card session-card--compact template-session-card`,children:[(0,Z.jsxs)(`div`,{className:`session-card-header`,children:[(0,Z.jsxs)(`div`,{className:`session-target`,children:[(0,Z.jsx)(g,{size:13,color:`var(--blue)`}),(0,Z.jsx)(`span`,{className:`mono`,children:e.name})]}),(0,Z.jsx)(`span`,{className:`session-tool-chip mono`,children:`Custom Template`})]}),(0,Z.jsxs)(`div`,{className:`session-card-meta`,children:[(0,Z.jsxs)(`span`,{children:[(0,Z.jsx)(M,{size:11}),` `,e.workflow_steps.length,` tools`]}),(0,Z.jsxs)(`span`,{children:[(0,Z.jsx)(w,{size:11}),` `,dt(e.updated_at)]})]}),(0,Z.jsxs)(`div`,{className:`session-tools`,children:[e.workflow_steps.slice(0,7).map((t,n)=>(0,Z.jsx)(`span`,{className:`session-tool-chip`,children:t.tool},`${e.template_id}:${n}:${t.tool}`)),e.workflow_steps.length>7&&(0,Z.jsxs)(`span`,{className:`session-tool-chip`,children:[`+`,e.workflow_steps.length-7]})]}),(0,Z.jsx)(`div`,{className:`session-card-footer`,children:(0,Z.jsxs)(`div`,{className:`template-card-actions`,children:[(0,Z.jsxs)(`button`,{className:`session-action-btn`,onClick:()=>Ne(e.template_id),children:[(0,Z.jsx)(H,{size:12}),` Use`]}),(0,Z.jsxs)(`button`,{className:`session-action-btn`,onClick:()=>je(e),children:[(0,Z.jsx)(b,{size:12}),` Edit`]}),(0,Z.jsxs)(`button`,{className:`session-delete-btn`,onClick:()=>x(e),disabled:_===e.template_id,children:[(0,Z.jsx)(ae,{size:12}),` `,_===e.template_id?`Deleting…`:`Delete`]})]})})]},e.template_id))})]}):(0,Z.jsx)(`section`,{className:`section`,children:(0,Z.jsxs)(`div`,{className:`section-header sessions-collapsed-toggle`,role:`button`,tabIndex:0,onClick:()=>{Te.current=!0,ve(!0)},onKeyDown:e=>{(e.key===`Enter`||e.key===` `)&&(e.preventDefault(),Te.current=!0,ve(!0))},children:[(0,Z.jsxs)(`h3`,{children:[`Custom Templates `,(0,Z.jsx)(`span`,{className:`badge`,children:u.length})]}),(0,Z.jsx)(`span`,{className:`session-help-btn`,children:`Show templates`})]})}),Be&&(0,Z.jsx)(`div`,{className:`modal-backdrop`,onClick:e=>{e.target===e.currentTarget&&Me()},children:(0,Z.jsxs)(`div`,{className:`modal modal--wide`,children:[(0,Z.jsxs)(`div`,{className:`modal-header`,children:[(0,Z.jsx)(`div`,{className:`modal-title-row`,children:(0,Z.jsx)(`span`,{className:`modal-name`,children:`Edit Template`})}),(0,Z.jsx)(`button`,{className:`modal-close`,onClick:Me,children:`x`})]}),(0,Z.jsxs)(`div`,{className:`modal-body`,children:[(0,Z.jsxs)(`div`,{className:`session-start-form`,children:[(0,Z.jsx)(`label`,{className:`mono`,htmlFor:`template-name-input`,children:`Template name *`}),(0,Z.jsx)(`input`,{id:`template-name-input`,className:`search-input mono`,value:re,onChange:e=>F(e.target.value),placeholder:`Template name`})]}),(0,Z.jsxs)(`div`,{className:`template-editor-grid`,children:[(0,Z.jsxs)(`div`,{className:`template-editor-col`,children:[(0,Z.jsx)(`span`,{className:`modal-label`,children:`Tools in template`}),I.length===0?(0,Z.jsx)(`div`,{className:`tasks-empty tasks-empty--compact`,children:(0,Z.jsx)(`p`,{children:`No tools selected yet.`})}):(0,Z.jsx)(`div`,{className:`template-step-list`,children:I.map((e,t)=>(0,Z.jsxs)(`div`,{className:`template-step-row`,children:[(0,Z.jsx)(`span`,{className:`mono`,children:e.tool}),(0,Z.jsx)(`button`,{className:`session-remove-tool`,onClick:()=>G(t),children:`x`})]},`${e.tool}:${t}`))})]}),(0,Z.jsxs)(`div`,{className:`template-editor-col`,children:[(0,Z.jsx)(`span`,{className:`modal-label`,children:`Add tools`}),(0,Z.jsx)(`input`,{className:`search-input mono`,value:oe,onChange:e=>se(e.target.value),placeholder:`Search tools`}),(0,Z.jsx)(`div`,{className:`session-add-tool-list`,children:Ve.map(e=>(0,Z.jsxs)(`button`,{className:`session-add-tool-item`,onClick:()=>Fe(e.name),children:[(0,Z.jsx)(`span`,{className:`mono`,children:e.name}),(0,Z.jsx)(`span`,{children:e.category})]},`edit-template-tool:${e.name}`))})]})]}),ce&&(0,Z.jsx)(`div`,{className:`run-error`,children:ce}),(0,Z.jsxs)(`div`,{className:`session-start-actions`,children:[(0,Z.jsx)(`button`,{className:`session-action-btn`,onClick:Me,children:`Cancel`}),(0,Z.jsxs)(`button`,{className:`session-run-btn`,onClick:K,disabled:le,children:[le?(0,Z.jsx)(U,{size:13,className:`spin`}):(0,Z.jsx)(b,{size:13}),le?`Saving…`:`Save Template`]})]})]})]})}),(0,Z.jsx)(Ht,{isOpen:!!y,title:`Delete Template`,description:y?`Delete template "${y.name}"? This action cannot be undone.`:`Delete template?`,impactItems:y?[`Template ID: ${y.template_id}`,`Tools in template: ${y.workflow_steps.length}`,`Future sessions can no longer use this template`]:[],confirmLabel:`Yes, delete template`,cancelLabel:`Keep template`,confirmVariant:`danger`,isConfirming:!!y&&_===y.template_id,onConfirm:async()=>{y&&await Pe(y.template_id)},onClose:()=>{_||x(null)}})]})}function Un(e,t,n,r){ht({id:Date.now(),tool:t,params:n,result:r,ts:new Date(r.timestamp),source:`browser`},e)}function Wn({isCompleted:e,sessionId:t,steps:n,selectedStep:i,selectedStepIndex:a,setSelectedStepIndex:o,selectedStepKey:s,selectedTool:c,stepState:l,runningStepKey:u,targetValue:d,setTargetValue:f,stepFieldValues:p,setStepFieldValues:m,showOptionalByStep:h,setShowOptionalByStep:g,selectedResult:_,onRunStep:y,onStopRunningStep:b,onApplyAttackChainFromResult:x,onRemoveTool:S,showAddTool:C,setShowAddTool:w,addToolSearch:T,setAddToolSearch:E,addCandidates:D,onAddTool:O}){let k=s?u===s:!1,A=_?.result,j=(0,W.useRef)(null);return(0,W.useEffect)(()=>{if(!C)return;let e=window.requestAnimationFrame(()=>j.current?.focus());return()=>window.cancelAnimationFrame(e)},[C]),(0,Z.jsxs)(`section`,{className:`section`,children:[(0,Z.jsx)(`div`,{className:`section-header`,children:(0,Z.jsxs)(`h3`,{children:[`Manual Tool Execution `,(0,Z.jsx)(`span`,{className:`badge`,children:n.length})]})}),(0,Z.jsxs)(`div`,{className:`session-workbench`,children:[(0,Z.jsxs)(`aside`,{className:`session-workbench-tools`,children:[!e&&(0,Z.jsxs)(`div`,{className:`session-tool-manage`,children:[(0,Z.jsx)(`button`,{className:`session-add-tool-btn`,onClick:()=>w(e=>!e),children:`+ Add tool`}),C&&(0,Z.jsxs)(`div`,{className:`session-add-tool-panel`,children:[(0,Z.jsx)(`input`,{ref:j,className:`search-input mono`,placeholder:`Search tool...`,value:T,onChange:e=>E(e.target.value)}),(0,Z.jsx)(`div`,{className:`session-add-tool-list`,children:D.map(e=>(0,Z.jsxs)(`button`,{className:`session-add-tool-item`,onClick:()=>O(e),children:[(0,Z.jsx)(`span`,{className:`mono`,children:e.name}),(0,Z.jsx)(`span`,{children:e.category})]},e.name))})]})]}),n.map((n,r)=>{let i=`${t}:${r}`,s=u===i;return(0,Z.jsxs)(`button`,{className:`session-workbench-tool session-workbench-tool--${l[i]??`idle`} ${s?`session-workbench-tool--running`:``} ${a===r?`active`:``}`,onClick:()=>o(r),children:[(0,Z.jsx)(`span`,{className:`session-workbench-tool-name mono`,children:n.tool}),(0,Z.jsxs)(`span`,{className:`session-workbench-tool-actions`,children:[s&&(0,Z.jsxs)(`span`,{className:`session-running-indicator mono`,title:`Tool is running`,children:[(0,Z.jsx)(U,{size:11,className:`spin`}),` Running`]}),!e&&(0,Z.jsx)(`button`,{type:`button`,className:`session-remove-tool`,onClick:e=>{e.stopPropagation(),S(r)},title:`Remove tool`,children:(0,Z.jsx)(ae,{size:12})})]})]},i)})]}),(0,Z.jsx)(`div`,{className:`session-workbench-center`,children:!i||!s||!c?(0,Z.jsx)(`p`,{className:`empty-state`,children:`No tools in this session.`}):(0,Z.jsxs)(Z.Fragment,{children:[(0,Z.jsxs)(`div`,{className:`session-target-override`,children:[(0,Z.jsx)(`label`,{className:`mono`,children:`Target`}),(0,Z.jsx)(`input`,{className:`search-input mono`,value:d,onChange:e=>f(e.target.value),placeholder:`Target to run this session against`,disabled:e})]}),(0,Z.jsxs)(`div`,{className:`session-step-row`,children:[(0,Z.jsxs)(`div`,{className:`session-step-head`,children:[(0,Z.jsx)(`span`,{className:`session-tool-chip mono session-tool-chip--${l[s]??`idle`}`,children:i.tool}),(0,Z.jsxs)($,{variant:`default`,disabled:e||k,onClick:()=>y(i,a),children:[k?(0,Z.jsx)(U,{size:12,className:`spin`}):(0,Z.jsx)(H,{size:12}),e?`Completed`:k?`Running…`:`Run ${i.tool}`]}),k&&!e&&(0,Z.jsxs)($,{variant:`danger`,onClick:()=>{b()},children:[(0,Z.jsx)(de,{size:12}),` Stop`]})]}),c.desc&&(0,Z.jsx)(`p`,{className:`session-tool-description`,children:c.desc}),(0,Z.jsxs)(`div`,{className:`session-param-grid`,children:[Object.keys(c.params).map(t=>(0,Z.jsx)(Bt,{name:t,value:p[s]?.[t]??``,onChange:e=>m(n=>({...n,[s]:{...n[s]??{},[t]:e}})),required:!0,disabled:e},t)),Object.keys(c.optional).length>0&&(0,Z.jsxs)(`button`,{className:`run-opt-btn`,onClick:()=>g(e=>({...e,[s]:!e[s]})),children:[h[s]?(0,Z.jsx)(v,{size:12}):(0,Z.jsx)(N,{size:12}),`Optional parameters (`,Object.keys(c.optional).length,`)`]}),h[s]&&Object.keys(c.optional).map(t=>(0,Z.jsx)(Bt,{name:t,value:p[s]?.[t]??``,onChange:e=>m(n=>({...n,[s]:{...n[s]??{},[t]:e}})),disabled:e},t))]})]}),(0,Z.jsxs)(`div`,{className:`session-result-panel`,children:[(0,Z.jsx)(`h4`,{children:`Result`}),e&&(0,Z.jsx)(`p`,{className:`section-meta`,children:`Completed session: results are read-only.`}),_?.error&&(0,Z.jsx)(`div`,{className:`run-error`,children:_.error}),A?(0,Z.jsxs)(Z.Fragment,{children:[(0,Z.jsxs)(`div`,{className:`session-result-head`,children:[(0,Z.jsxs)(`div`,{className:`session-step-result mono`,children:[A.success?`OK`:`FAIL`,` | exit `,A.return_code,` | `,A.execution_time.toFixed(2),`s`]}),s&&A&&(0,Z.jsxs)(`div`,{className:`session-result-actions`,children:[(0,Z.jsxs)($,{variant:`default`,onClick:()=>Un(`txt`,i.tool,p[s]??{},A),children:[(0,Z.jsx)(r,{size:11}),` Export`]}),(0,Z.jsxs)($,{variant:`default`,onClick:()=>Un(`json`,i.tool,p[s]??{},A),children:[(0,Z.jsx)(r,{size:11}),` JSON`]}),i.tool===`create-attack-chain`&&A.success&&(0,Z.jsx)($,{disabled:e,variant:`success`,onClick:()=>{x()},children:`Use Chain`})]})]}),(0,Z.jsx)(`pre`,{className:`session-result-pre mono`,children:A.stdout||`(no stdout)`}),A.stderr&&(0,Z.jsx)(`pre`,{className:`session-result-pre mono`,children:A.stderr})]}):(0,Z.jsx)(`p`,{className:`section-meta`,children:`No result yet for this tool.`})]})]})})]})]})}function Gn({show:e,templateName:t,templateError:n,setTemplateName:r,onClose:i,onSave:a}){return e?(0,Z.jsx)(`div`,{className:`modal-backdrop`,onClick:e=>{e.target===e.currentTarget&&i()},children:(0,Z.jsxs)(`div`,{className:`modal`,children:[(0,Z.jsxs)(`div`,{className:`modal-header`,children:[(0,Z.jsx)(`div`,{className:`modal-title-row`,children:(0,Z.jsx)(`span`,{className:`modal-name`,children:`Create Template`})}),(0,Z.jsx)(`button`,{className:`modal-close`,onClick:i,children:`x`})]}),(0,Z.jsx)(`div`,{className:`modal-body`,children:(0,Z.jsxs)(`div`,{className:`session-start-form`,children:[(0,Z.jsx)(`label`,{className:`mono`,children:`Template Name *`}),(0,Z.jsx)(`input`,{className:`search-input mono`,placeholder:`e.g. SMB Enumeration Pack`,value:t,onChange:e=>r(e.target.value)}),n&&(0,Z.jsx)(`div`,{className:`run-error`,children:n}),(0,Z.jsxs)(`div`,{className:`session-start-actions`,children:[(0,Z.jsx)(`button`,{className:`session-action-btn`,onClick:i,children:`Cancel`}),(0,Z.jsx)(`button`,{className:`session-run-btn`,onClick:a,children:`Save Template`})]})]})})]})}):null}function Kn(e){return Array.isArray(e.workflow_steps)&&e.workflow_steps.length>0?e.workflow_steps:e.tools_executed.map(e=>({tool:e,parameters:{}}))}function qn(e){return e.toLowerCase().replace(/[^a-z0-9]/g,``)}function Jn(e,t){let n=e.trim();if(!n)return null;let r=t.find(e=>e.name===n);if(r)return r;let i=t.find(e=>e.endpoint===n);if(i)return i;let a=t.find(e=>e.parent_tool===n);if(a)return a;let o=qn(n),s=null;for(let e of t){let t=qn(e.name),n=qn(e.parent_tool??``),r=qn(e.endpoint),i=0;t===o&&(i=Math.max(i,80)),n===o&&(i=Math.max(i,75)),r===o&&(i=Math.max(i,70)),t.includes(o)&&(i=Math.max(i,62)),r.includes(o)&&(i=Math.max(i,58)),n&&n.includes(o)&&(i=Math.max(i,56)),o.includes(t)&&(i=Math.max(i,52)),i!==0&&(!s||i>s.score)&&(s={tool:e,score:i})}return s?.tool??null}function Yn({sessionId:e,tools:t,onBack:n,onToolRun:i}){let[a,o]=(0,W.useState)(null),[s,c]=(0,W.useState)(!0),[l,u]=(0,W.useState)(null),[d,f]=(0,W.useState)(``),[p,m]=(0,W.useState)({}),[h,g]=(0,W.useState)({}),[_,v]=(0,W.useState)(null),[y,b]=(0,W.useState)({}),[x,S]=(0,W.useState)({}),[C,T]=(0,W.useState)(0),[E,D]=(0,W.useState)(!1),[O,k]=(0,W.useState)(``),[A,j]=(0,W.useState)(null),[ee,N]=(0,W.useState)(!1),[P,ne]=(0,W.useState)(``),[re,F]=(0,W.useState)(!1),[I,L]=(0,W.useState)(null),[ie,ae]=(0,W.useState)(!1),[oe,se]=(0,W.useState)(!1),[ce,R]=(0,W.useState)(!1),[le,de]=(0,W.useState)(!1),fe=(0,W.useMemo)(()=>Object.fromEntries(t.map(e=>[e.name,e])),[t]),z=a?Kn(a):[];async function B(){c(!0);try{let t=await J.session(e);o(t.session),f(t.session.target);let n=t.session.metadata??{},r=n.tool_status??{},i=n.step_results??{},a=typeof n.running_step_key==`string`?n.running_step_key:null,s={},c={};for(let[e,t]of Object.entries(r))(t===`success`||t===`failed`)&&(s[e]=t);for(let[e,t]of Object.entries(i))!t||typeof t!=`object`||(c[e]={result:{success:!!t.success,return_code:Number(t.return_code??0),execution_time:Number(t.execution_time??0),timed_out:!1,partial_results:!1,stdout:String(t.stdout??``),stderr:String(t.stderr??``),timestamp:String(t.timestamp??new Date().toISOString())}});S(s),b(c);try{let e=await J.processList();Object.keys(e.active_processes??{}).length>0&&a?(s[a]=`running`,v(a),S({...s})):v(null)}catch{}u(null)}catch(e){u(String(e))}finally{c(!1)}}(0,W.useEffect)(()=>{B()},[e]);async function pe(){if(a){F(!0);try{let e=`Tools: ${z.map(e=>e.tool).join(`, `)}`,t=await J.handoverSession(a.session_id,e);L(`LLM handover done -> ${t.handover?.category??`unknown`} (${((t.handover?.confidence??0)*100).toFixed(0)}% confidence)`),await B()}catch(e){L(`Handover failed: ${String(e)}`)}finally{F(!1)}}}async function he(){if(a){R(!0);try{await J.updateSession(a.session_id,{status:`completed`}),n()}catch(e){L(`Complete failed: ${String(e)}`)}finally{R(!1),ae(!1)}}}async function ge(){if(a){de(!0);try{await J.deleteSession(a.session_id),n()}catch(e){L(`Delete failed: ${String(e)}`)}finally{de(!1),se(!1)}}}function _e(){if(a)try{let e=a.metadata??{},t=e.tool_status&&typeof e.tool_status==`object`?e.tool_status:{},n=e.step_results&&typeof e.step_results==`object`?e.step_results:{},r=z.map((e,r)=>{let i=`${a.session_id}:${r}`,o=y[i]?.result,s=n[i],c=o?{success:o.success,return_code:o.return_code,execution_time:o.execution_time,timestamp:o.timestamp,stdout:o.stdout,stderr:o.stderr,timed_out:o.timed_out,partial_results:o.partial_results}:s?{success:s.success,return_code:s.return_code,execution_time:s.execution_time,timestamp:s.timestamp??``,stdout:s.stdout??``,stderr:s.stderr??``,timed_out:!1,partial_results:!1}:null;return{step_index:r,step_key:i,tool:e.tool,parameters:e.parameters??{},status:x[i]??t[i]??`idle`,result:c}}),i={session_id:a.session_id,target:a.target,status:a.status??`active`,exported_at:new Date().toISOString(),logs:r},o=JSON.stringify(i,null,2),s=new Blob([o],{type:`application/json`}),c=URL.createObjectURL(s),l=document.createElement(`a`);l.href=c,l.download=`${a.session_id}_tool_logs.json`,l.click(),URL.revokeObjectURL(c),L(`Tool logs exported`)}catch(e){L(`Export failed: ${String(e)}`)}}async function ve(){if(!a)return;let e=O.trim();if(!e){j(`Template name is required`);return}try{let t=Kn(a);if(t.length===0){j(`Session has no tools to save as template`);return}try{await J.createSessionTemplate({name:e,workflow_steps:t,source_session_id:a.session_id})}catch{await J.createSessionTemplateCompat({name:e,workflow_steps:t,source_session_id:a.session_id})}j(null),k(``),D(!1),L(`Template created: ${e}`)}catch(e){j(String(e))}}let V=z[C]??null,H=V&&a?`${a.session_id}:${C}`:null,ye=H?y[H]:void 0,be=V?fe[V.tool]:null;function xe(e){f(e),a&&m(n=>{let r=!1,i={...n};for(let o=0;o<z.length;o+=1){let s=z[o],c=Jn(s.tool,t);if(!c)continue;let l=`${a.session_id}:${o}`,u=n[l]??Pt(c,s,e),d=!1,f={...u};for(let t of[...Object.keys(c.params),...Object.keys(c.optional)]){let n=Nt(t,e);n===void 0||f[t]===n||(f[t]=n,d=!0)}(d||!n[l])&&(i[l]=f,r=!0)}return r?i:n})}(0,W.useEffect)(()=>{!a||!V||!H||!be||(m(e=>e[H]?e:{...e,[H]:Pt(be,V,d.trim()||a.target)}),g(e=>({...e,[H]:e[H]??!1})))},[a,V,H,be,d]);async function Se(e,n){if(!a)return;let r=a,o=`${r.session_id}:${n}`,s=Jn(e.tool,t);if(!s){b(t=>({...t,[o]:{error:`Tool ${e.tool} could not be mapped to registry`}})),S(e=>({...e,[o]:`failed`}));return}let c=d.trim(),{payload:l,missing:u}=Ft(s,p[o]??Pt(s,e,c||r.target));if(u.length>0){b(e=>({...e,[o]:{error:`Missing required: ${u.join(`, `)}`}})),S(e=>({...e,[o]:`failed`}));return}let f=z.map((e,t)=>t===n?{...e,parameters:l}:e);try{await J.updateSession(r.session_id,{target:c,workflow_steps:f})}catch{}let m=r.metadata??{};try{await J.updateSession(r.session_id,{metadata:{...m,running_step_key:o,running_tool:e.tool,running_started_at:new Date().toISOString()}})}catch{}v(o),S(e=>({...e,[o]:`running`})),b(e=>({...e,[o]:{}}));try{let t=await J.runTool(s.endpoint,l);b(e=>({...e,[o]:{result:t}})),S(e=>({...e,[o]:t.success?`success`:`failed`})),i&&i(e.tool,l,t);let n=r.metadata??{},a=n.tool_status&&typeof n.tool_status==`object`?n.tool_status:{},c=n.step_results&&typeof n.step_results==`object`?n.step_results:{};await J.updateSession(r.session_id,{metadata:{...n,tool_status:{...a,[e.tool]:t.success?`success`:`failed`,[o]:t.success?`success`:`failed`},step_results:{...c,[o]:{success:t.success,return_code:t.return_code,execution_time:t.execution_time,timestamp:t.timestamp,stdout:t.stdout,stderr:t.stderr}},last_run:{step_key:o,tool:e.tool,success:t.success,return_code:t.return_code,execution_time:t.execution_time,timestamp:t.timestamp},running_step_key:null,running_tool:null,running_started_at:null}}),await B()}catch(e){b(t=>({...t,[o]:{error:String(e)}})),S(e=>({...e,[o]:`failed`}));try{let e=r.metadata??{};await J.updateSession(r.session_id,{metadata:{...e,running_step_key:null,running_tool:null,running_started_at:null}})}catch{}}finally{v(null)}}async function Ce(){let e=_;if(e)try{let t=(await J.processList()).active_processes??{},n=Object.values(t).filter(e=>e&&typeof e.command==`string`&&e.command.trim()!==``).map(e=>e.pid).filter(e=>typeof e==`number`);if(n.length===0){L(`No active process found to terminate`),v(null);return}let r=Math.max(...n);if(await J.terminateProcess(r),L(`Stopped running process ${r}`),v(null),S(t=>({...t,[e]:`failed`})),b(t=>({...t,[e]:{...t[e]??{},error:`Process ${r} terminated by user`}})),a)try{let e=a.metadata??{};await J.updateSession(a.session_id,{metadata:{...e,running_step_key:null,running_tool:null,running_started_at:null}})}catch{}}catch(e){L(`Stop failed: ${String(e)}`)}}async function we(e){if(!a)return;let t=[...z,{tool:e.name,parameters:{}}];try{await J.updateSession(a.session_id,{workflow_steps:t}),N(!1),ne(``),await B(),T(Math.max(0,t.length-1))}catch(e){L(`Add tool failed: ${String(e)}`)}}async function Te(){if(!a||!H||!ye?.result)return;let e=ye.result,t=null;try{t=e.stdout?JSON.parse(e.stdout):null}catch{t=null}let n=(t&&typeof t==`object`?t:null)?.attack_chain,r=n&&typeof n==`object`?n.steps:null,i=Array.isArray(r)?r.filter(e=>!!e&&typeof e==`object`&&typeof e.tool==`string`):[];if(i.length===0){L(`No attack-chain steps found in result output`);return}let o=new Set(z.map(e=>`${e.tool}::${JSON.stringify(e.parameters??{})}`)),s=i.filter(e=>{let t=`${e.tool}::${JSON.stringify(e.parameters??{})}`;return o.has(t)?!1:(o.add(t),!0)});if(s.length===0){L(`All attack-chain tools are already present in manual execution`);return}let c=[...z,...s];try{await J.updateSession(a.session_id,{workflow_steps:c}),await B(),T(Math.max(0,c.length-s.length)),L(`Added ${s.length} tool(s) from attack chain`)}catch(e){L(`Apply attack chain failed: ${String(e)}`)}}async function Ee(e){if(!a)return;let t=z.filter((t,n)=>n!==e);try{await J.updateSession(a.session_id,{workflow_steps:t}),await B(),T(n=>t.length===0?0:n>e?n-1:Math.min(n,t.length-1))}catch(e){L(`Remove tool failed: ${String(e)}`)}}if(s)return(0,Z.jsxs)(`div`,{className:`loading-state`,children:[(0,Z.jsx)(U,{size:20,className:`spin`,color:`var(--green)`}),(0,Z.jsx)(`p`,{children:`Loading session…`})]});if(l||!a)return(0,Z.jsx)(`div`,{className:`error-banner`,children:l??`Session not found`});let De=(a.status??`active`)===`completed`,Oe=t.filter(e=>{if(!P.trim())return!0;let t=P.toLowerCase();return e.name.toLowerCase().includes(t)||e.desc.toLowerCase().includes(t)||e.category.toLowerCase().includes(t)}).slice(0,12);return(0,Z.jsxs)(`div`,{className:`page-content`,children:[(0,Z.jsxs)(`section`,{className:`section`,children:[(0,Z.jsxs)(`div`,{className:`section-header`,children:[(0,Z.jsx)(`h3`,{children:`Session Detail`}),(0,Z.jsxs)(`button`,{className:`session-action-btn`,onClick:n,children:[(0,Z.jsx)(me,{size:12}),` Back`]})]}),(0,Z.jsxs)(`div`,{className:`session-detail-meta-row`,children:[(0,Z.jsxs)(`span`,{children:[(0,Z.jsx)(te,{size:12}),` `,(0,Z.jsx)(`span`,{className:`mono`,children:a.target})]}),(0,Z.jsxs)(`span`,{children:[(0,Z.jsx)(M,{size:12}),` `,a.total_findings,` findings`]}),(0,Z.jsxs)(`span`,{children:[(0,Z.jsx)(U,{size:12}),` `,a.iterations,` iterations`]}),(0,Z.jsxs)(`span`,{children:[(0,Z.jsx)(w,{size:12}),` `,dt(a.updated_at)]}),(0,Z.jsx)(`span`,{className:`session-status session-status--${a.status??`active`}`,children:a.status??`active`})]}),(0,Z.jsxs)(`div`,{className:`session-detail-actions`,children:[(0,Z.jsx)(`button`,{className:`session-action-btn`,onClick:()=>D(!0),children:`Create Template`}),(0,Z.jsxs)(`button`,{className:`session-action-btn`,onClick:_e,children:[(0,Z.jsx)(r,{size:12}),` Export Tool Logs`]}),(0,Z.jsxs)(`button`,{className:`session-action-btn`,onClick:pe,disabled:re,children:[(0,Z.jsx)(ue,{size:12}),` `,re?`Handing over…`:`Handover to LLM`]}),a.status!==`completed`&&(0,Z.jsx)(`button`,{className:`session-complete-btn`,onClick:()=>ae(!0),children:`Complete Session`}),(0,Z.jsx)(`button`,{className:`session-delete-btn`,onClick:()=>se(!0),children:`Delete Session`}),I&&(0,Z.jsx)(`span`,{className:`section-meta`,children:I})]}),(0,Z.jsx)(Gn,{show:E,templateName:O,templateError:A,setTemplateName:k,onClose:()=>D(!1),onSave:ve}),(0,Z.jsx)(Ht,{isOpen:ie,title:`Complete Session`,description:`Mark this session as completed? It will move to Completed Sessions.`,impactItems:[`Session ID: ${a.session_id}`,`Session will become read-only`],confirmLabel:`Yes, complete session`,cancelLabel:`Keep active`,confirmVariant:`danger`,isConfirming:ce,onConfirm:he,onClose:()=>ae(!1)}),(0,Z.jsx)(Ht,{isOpen:oe,title:`Delete Session`,description:`Delete this session permanently? This action cannot be undone.`,impactItems:[`Session ID: ${a.session_id}`,`All workflow state and run metadata will be removed`],confirmLabel:`Yes, delete session`,cancelLabel:`Keep session`,confirmVariant:`danger`,isConfirming:le,onConfirm:ge,onClose:()=>se(!1)})]}),(0,Z.jsx)(Wn,{isCompleted:De,sessionId:a.session_id,steps:z,selectedStep:V,selectedStepIndex:C,setSelectedStepIndex:T,selectedStepKey:H,selectedTool:be,stepState:x,runningStepKey:_,targetValue:d,setTargetValue:xe,stepFieldValues:p,setStepFieldValues:m,showOptionalByStep:h,setShowOptionalByStep:g,selectedResult:ye,onRunStep:Se,onStopRunningStep:Ce,onApplyAttackChainFromResult:Te,onRemoveTool:Ee,showAddTool:ee,setShowAddTool:N,addToolSearch:P,setAddToolSearch:ne,addCandidates:Oe,onAddTool:we})]})}function Xn({page:e,demo:t,tools:n,health:r,toolsStatusWithParents:i,runHistory:a,setRunHistory:o,fetchServerRunHistory:s,clearServerRunHistory:c,commandToolRequest:l,onCommandToolHandled:u,openSessionDetail:d,activeSessionId:f,setPage:p,addBrowserRunEntry:m,logLines:h,logAutoScroll:g,setLogAutoScroll:_,logLimit:v,setLogLimit:y,logEndRef:b,loading:x,error:S,history:C,toolCategories:w}){return(0,Z.jsxs)(`main`,{className:`main${e===`run`?` main--flush`:``}`,children:[e===`settings`&&(0,Z.jsx)(hn,{}),e===`help`&&(0,Z.jsx)(Cn,{}),e===`run`&&(0,Z.jsx)(Xt,{tools:n,toolsStatus:i,runHistory:a,setRunHistory:o,commandToolRequest:l,onCommandToolHandled:u,onRefresh:s,onClearHistory:c}),e===`tasks`&&(0,Z.jsx)(Dn,{demoData:t?{processes:Xe}:void 0}),e===`tools`&&r&&(0,Z.jsx)(kn,{health:r,tools:n,toolsStatus:i}),e===`reports`&&(0,Z.jsx)(Fn,{runHistory:a}),e===`sessions`&&(0,Z.jsx)(Hn,{demoData:t?{sessions:Ye}:void 0,onOpenSession:d}),e===`session-detail`&&f&&(0,Z.jsx)(Yn,{sessionId:f,tools:n,onBack:()=>p(`sessions`),onToolRun:m}),e===`logs`&&(0,Z.jsx)(nn,{logLines:h,logAutoScroll:g,setLogAutoScroll:_,logLimit:v,setLogLimit:y,logEndRef:b}),e===`dashboard`&&(0,Z.jsxs)(Z.Fragment,{children:[x&&!r&&(0,Z.jsxs)(`div`,{className:`loading-state`,children:[(0,Z.jsx)(U,{size:24,className:`spin`,color:`var(--green)`}),(0,Z.jsx)(`p`,{children:`Connecting to server…`})]}),S&&!r&&(0,Z.jsxs)(`div`,{className:`error-banner`,children:[S,` — is the server running on port 8888?`]}),r&&(0,Z.jsx)(kt,{health:r,tools:n,history:C,runHistory:a,loading:x,error:S,toolCategories:w})]})]})}var Zn=[{page:`dashboard`,label:`Open Dashboard`},{page:`run`,label:`Open Run`},{page:`reports`,label:`Open Reports`},{page:`tasks`,label:`Open Tasks`},{page:`tools`,label:`Open Tools`},{page:`sessions`,label:`Open Sessions`},{page:`settings`,label:`Open Settings`},{page:`logs`,label:`Open Logs`},{page:`help`,label:`Open Help`}];function Qn({open:e,onClose:t,setPage:n,tools:r,onSelectTool:i}){let[a,o]=(0,W.useState)(``),[s,c]=(0,W.useState)(0),l=(0,W.useMemo)(()=>{let e=Zn.map(e=>({id:`page-${e.page}`,label:e.label,hint:`Navigation`,run:()=>{n(e.page),t()}})),a=r.slice(0,120).map(e=>({id:`tool-${e.name}`,label:`Run ${e.name}`,hint:e.category.replace(/_/g,` `),run:()=>{n(`run`),i(e),t()}}));return[...e,...a]},[r,n,i,t]),u=(0,W.useMemo)(()=>{let e=a.trim().toLowerCase();return e?l.filter(t=>`${t.label} ${t.hint}`.toLowerCase().includes(e)).slice(0,18):l.slice(0,18)},[l,a]);return(0,W.useEffect)(()=>{e&&(o(``),c(0))},[e]),(0,W.useEffect)(()=>{if(!e)return;function n(e){if(e.key===`Escape`){e.preventDefault(),t();return}if(e.key===`ArrowDown`){e.preventDefault(),c(e=>Math.min(e+1,Math.max(u.length-1,0)));return}if(e.key===`ArrowUp`){e.preventDefault(),c(e=>Math.max(e-1,0));return}if(e.key===`Enter`){let t=u[s];t&&(e.preventDefault(),t.run())}}return window.addEventListener(`keydown`,n),()=>window.removeEventListener(`keydown`,n)},[e,u,s,t]),e?(0,At.createPortal)((0,Z.jsx)(`div`,{className:`modal-backdrop`,onClick:t,children:(0,Z.jsxs)(`div`,{className:`modal command-palette`,onClick:e=>e.stopPropagation(),children:[(0,Z.jsxs)(`div`,{className:`command-palette-input-wrap`,children:[(0,Z.jsx)(O,{size:14}),(0,Z.jsx)(`input`,{autoFocus:!0,className:`command-palette-input mono`,value:a,onChange:e=>{o(e.target.value),c(0)},placeholder:`Type a command or tool name...`})]}),(0,Z.jsxs)(`div`,{className:`command-palette-list`,children:[u.map((e,t)=>(0,Z.jsxs)(`button`,{className:`command-palette-item${t===s?` active`:``}`,onClick:e.run,children:[(0,Z.jsx)(`span`,{className:`command-palette-label`,children:e.label}),(0,Z.jsx)(`span`,{className:`command-palette-hint`,children:e.hint})]},e.id)),u.length===0&&(0,Z.jsx)(`div`,{className:`command-palette-empty`,children:`No results`})]})]})}),document.body):null}var $n=1e4,er=`hexstrike_palette_hint_seen`;function tr(){let[e]=(0,W.useState)(Re),[t,n]=(0,W.useState)(e||Fe()),[r,i]=(0,W.useState)(!1),a=nt(),[o,s]=(0,W.useState)(a.page),[c,l]=(0,W.useState)(a.sessionId);function u(e){e!==`session-detail`&&(window.location.hash=e===`dashboard`?``:`/${e}`,s(e),l(null))}function d(e){window.location.hash=`/sessions/${e}`,s(`session-detail`),l(e)}(0,W.useEffect)(()=>{function e(){let e=nt();s(e.page),l(e.sessionId)}return window.addEventListener(`hashchange`,e),()=>window.removeEventListener(`hashchange`,e)},[]);let[f,p]=(0,W.useState)(e?Ge:null),[m,h]=(0,W.useState)(e?Be:[]),[g,v]=(0,W.useState)(e?Ke():[]);(0,W.useEffect)(()=>{e||J.tools().then(e=>h(e.tools)).catch(()=>{})},[e]);let[y,b]=(0,W.useState)(()=>e?qe:[]),[x,S]=(0,W.useState)(e?new Date:null),[C,w]=(0,W.useState)(!e),[T,E]=(0,W.useState)(null),[D,O]=(0,W.useState)(e?Je:[]),[k,A]=(0,W.useState)(!0),[j,ee]=(0,W.useState)(500),M=(0,W.useRef)(null),te=(0,W.useRef)(null),N=(0,W.useRef)(null),P=(0,W.useRef)(null),[ne,re]=(0,W.useState)(!1),[F,I]=(0,W.useState)(null),[L,ie]=(0,W.useState)({}),[ae,oe]=(0,W.useState)(!1),[se,ce]=(0,W.useState)(null),[R,le]=(0,W.useState)(()=>{try{return localStorage.getItem(er)!==`1`}catch{return!0}}),[ue,de]=(0,W.useState)(()=>{try{let e=localStorage.getItem(at);if(e&&st(e))return e}catch{}return`dark`});(0,W.useEffect)(()=>{document.documentElement.setAttribute(`data-theme`,ue);try{localStorage.setItem(at,ue)}catch{}},[ue]),(0,W.useEffect)(()=>{function e(e){(e.ctrlKey||e.metaKey)&&e.key.toLowerCase()===`k`&&(e.preventDefault(),oe(!0))}return window.addEventListener(`keydown`,e),()=>window.removeEventListener(`keydown`,e)},[]);function fe(e){u(`run`),ce({toolName:e.name,requestId:Date.now()})}function z(){le(!1);try{localStorage.setItem(er,`1`)}catch{}}function B(){oe(!0),R&&z()}let pe=(0,W.useCallback)((e,t,n)=>{b(r=>[{id:Date.now(),tool:e,params:t,result:n,ts:new Date,source:`browser`},...r].slice(0,200))},[]),me=(0,W.useCallback)(async()=>{if(!e)try{let e=await J.dashboard();p(e),v(t=>[...t.slice(-29),{t:Date.now(),cpu:e.resources.cpu_percent,mem:e.resources.memory_percent,network_bytes_sent:e.resources.network_bytes_sent,network_bytes_recv:e.resources.network_bytes_recv}]),S(new Date),E(null)}catch(e){e instanceof Error&&e.message===`UNAUTHORIZED`?(i(!0),n(!1)):E(`Server unreachable`)}finally{w(!1)}},[]);(0,W.useEffect)(()=>{e||!t||(async()=>{try{ie((await J.getToolCategories()).categories)}catch{}})()},[e,t]);let he=(0,W.useCallback)(async()=>{if(!e)try{let e=await J.runHistory();if(!e.success)return;b(t=>{let n=new Set(t.filter(e=>e.source===`server`).map(e=>e.serverId)),r=e.runs.filter(e=>{if(n.has(e.id))return!1;let r=e.timestamp?new Date(e.timestamp).getTime():0;return!t.some(t=>t.source===`browser`&&t.tool===e.tool&&r>0&&Math.abs(t.ts.getTime()-r)<1e4)}).map(e=>({id:-e.id,serverId:e.id,source:`server`,tool:e.tool,params:e.params,ts:e.timestamp?new Date(e.timestamp):new Date,result:{stdout:e.stdout,stderr:e.stderr,return_code:e.return_code,success:e.success,timed_out:e.timed_out,partial_results:e.partial_results,execution_time:e.execution_time,timestamp:e.timestamp}}));return r.length===0?t:[...t,...r].sort((e,t)=>t.ts.getTime()-e.ts.getTime())})}catch{}},[]),ge=(0,W.useCallback)(async()=>{if(e){b([]);return}(await J.clearRunHistory()).success&&b([])},[e]);if((0,W.useEffect)(()=>{e||!t||he().catch(()=>{})},[e,t,he]),(0,W.useEffect)(()=>{e||Fe()||J.dashboard().then(e=>{p(e),n(!0),w(!1)}).catch(e=>{e instanceof Error&&e.message===`UNAUTHORIZED`?i(!0):n(!0),w(!1)})},[]),(0,W.useEffect)(()=>{if(e||!t)return;N.current&&N.current.close(),P.current&&=(clearInterval(P.current),null);function n(){P.current&&clearInterval(P.current),P.current=window.setInterval(()=>{me()},$n)}let r=J.dashboardStream();return N.current=r,r.onmessage=e=>{try{let t=JSON.parse(e.data);p(t),v(e=>[...e.slice(-29),{t:Date.now(),cpu:t.resources.cpu_percent,mem:t.resources.memory_percent,network_bytes_sent:t.resources.network_bytes_sent,network_bytes_recv:t.resources.network_bytes_recv}]),S(new Date),w(!1),E(null),re(!0),I(null),P.current&&=(clearInterval(P.current),null)}catch{I(`Malformed dashboard data`)}},r.onerror=()=>{re(!1),I(`Dashboard stream disconnected; using polling.`),P.current||n()},()=>{r.close(),P.current&&clearInterval(P.current)}},[e,t,me]),(0,W.useEffect)(()=>{if(e||o!==`logs`)return;let t=J.logStream(150);return te.current=t,t.onmessage=e=>{O(t=>{let n=[...t,e.data];return n.length>500?n.slice(-500):n})},()=>{t.close()}},[e,o]),(0,W.useEffect)(()=>{o===`logs`&&k&&M.current?.scrollIntoView({behavior:`smooth`})},[D,o,k]),r&&!t)return(0,Z.jsx)(Ze,{onUnlocked:()=>{n(!0),i(!1)}});let _e=rt(m,f?.tools_status??{});return(0,Z.jsx)($e,{children:(0,Z.jsxs)(`div`,{className:e?`layout layout--demo`:`layout`,children:[(0,Z.jsx)(Qn,{open:ae,onClose:()=>oe(!1),setPage:u,tools:m,onSelectTool:fe}),e&&(0,Z.jsxs)(`div`,{className:`demo-banner`,children:[(0,Z.jsx)(_,{size:13}),(0,Z.jsx)(`span`,{children:`Demo mode — all data is synthetic`}),(0,Z.jsx)(`button`,{onClick:()=>{ze(),window.location.href=window.location.pathname+window.location.hash},children:`Exit demo`})]}),(0,Z.jsx)(lt,{page:o,setPage:u,lastRefresh:x,demo:e,isStreaming:ne,streamingError:F,health:f,error:T,loading:C,fetchAll:me,themeId:ue,setThemeId:de,onOpenCommandPalette:B,onSignOut:()=>{n(!1),i(!0)}}),R&&(0,Z.jsxs)(`div`,{className:`palette-hint-floating`,role:`status`,"aria-live":`polite`,children:[(0,Z.jsxs)(`button`,{className:`palette-hint-main`,onClick:B,title:`Open command palette`,children:[(0,Z.jsx)(`span`,{className:`palette-hint-title`,children:`Try Command Palette`}),(0,Z.jsx)(`span`,{className:`palette-hint-kbd mono`,children:`Ctrl/Cmd + K`})]}),(0,Z.jsx)(`button`,{className:`palette-hint-close`,onClick:z,title:`Dismiss hint`,children:`x`})]}),(0,Z.jsx)(Xn,{page:o,demo:e,tools:m,health:f,toolsStatusWithParents:_e,runHistory:y,setRunHistory:b,fetchServerRunHistory:he,clearServerRunHistory:ge,commandToolRequest:se,onCommandToolHandled:()=>ce(null),openSessionDetail:d,activeSessionId:c,setPage:u,addBrowserRunEntry:pe,logLines:D,logAutoScroll:k,setLogAutoScroll:A,logLimit:j,setLogLimit:ee,logEndRef:M,loading:C,error:T,history:g,toolCategories:L})]})})}(0,ke.createRoot)(document.getElementById(`root`)).render((0,Z.jsx)(W.StrictMode,{children:(0,Z.jsx)(tr,{})}));
++  --profile full`})]})}function xn(){return(0,Z.jsxs)(`section`,{className:`section help-about-section`,children:[(0,Z.jsx)(`div`,{className:`section-header`,children:(0,Z.jsx)(`h3`,{children:`Demo Mode`})}),(0,Z.jsxs)(`div`,{className:`help-about`,children:[(0,Z.jsx)(`p`,{className:`help-about-desc`,children:`Activate demo mode to explore the dashboard. All data is synthetic but designed to feel realistic. Ideal for learning, demos, or just satisfying your curiosity!`}),(0,Z.jsxs)(`button`,{className:`help-demo-btn`,onClick:()=>{window.location.href=window.location.pathname+`?demo=1`+window.location.hash},children:[(0,Z.jsx)(_,{size:13}),`Try demo mode`]})]})]})}var Sn=[{id:`claude`,label:`Claude Desktop`,icon:`🤖`,configPath:`~/.config/Claude/claude_desktop_config.json`,note:`Also works for Cursor — same config format.`,json:e=>JSON.stringify({mcpServers:{"hexstrike-ai":{command:`${e}/hexstrike-env/bin/python3`,args:[`${e}/hexstrike_mcp.py`,`--server`,`http://localhost:8888`,`--profile`,`full`],description:`HexStrike AI Community Edition`,timeout:300,disabled:!1}}},null,2)},{id:`vscode`,label:`VS Code Copilot`,icon:`🔷`,configPath:`.vscode/settings.json  (workspace) or User settings`,note:`Place in your workspace .vscode/settings.json or open User Settings JSON.`,json:e=>JSON.stringify({servers:{hexstrike:{type:`stdio`,command:`${e}/hexstrike-env/bin/python3`,args:[`${e}/hexstrike_mcp.py`,`--server`,`http://127.0.0.1:8888`,`--profile`,`full`]}},inputs:[]},null,2)},{id:`opencode`,label:`OpenCode`,icon:`⚡`,configPath:`~/.config/opencode/opencode.json`,note:`OpenCode reads this on startup.`,json:e=>JSON.stringify({$schema:`https://opencode.ai/config.json`,mcp:{hexstrike:{type:`local`,command:[`${e}/hexstrike-env/bin/python3`,`${e}/hexstrike_mcp.py`,`--server`,`http://127.0.0.1:8888`,`--profile`,`default`],enabled:!0}}},null,2)},{id:`roo`,label:`Roo Code`,icon:`🦘`,configPath:`MCP Servers panel  →  Edit Config`,note:`Open Roo Code → MCP Servers → Edit Config and paste the block below.`,json:e=>JSON.stringify({mcpServers:{hexstrike:{command:`${e}/hexstrike-env/bin/python3`,args:[`${e}/hexstrike_mcp.py`,`--server`,`http://127.0.0.1:8888`,`--profile`,`full`],timeout:300}}},null,2)}];function Cn(){let[e,t]=(0,W.useState)(`claude`),[n,r]=(0,W.useState)(`/path/to/hexstrike-ai-community-edition`),[i,a]=(0,W.useState)(!1),o=Sn.find(t=>t.id===e)??Sn[0];return(0,W.useEffect)(()=>{let e=!0;return J.getSettings().then(t=>{let n=t.settings.server.working_dir?.trim();!e||!n||(r(n),a(!0))}).catch(()=>{e&&a(!1)}),()=>{e=!1}},[]),(0,Z.jsxs)(`div`,{className:`help-page`,children:[(0,Z.jsx)(vn,{installPath:n,setInstallPath:r,pathDetected:i,activeIde:e,setActiveIde:t,ideConfigs:Sn,selectedIde:o}),(0,Z.jsx)(yn,{}),(0,Z.jsx)(bn,{}),(0,Z.jsx)(xn,{})]})}function wn(e){let[t,n]=(0,W.useState)(e?.processes??null),[r,i]=(0,W.useState)(e?{workers:4,queued:2,completed:38}:null),[a,o]=(0,W.useState)(!e),[s,c]=(0,W.useState)(null),[l,u]=(0,W.useState)(null),[d,f]=(0,W.useState)(e?`polling`:`streaming`),p=(0,W.useRef)(null),m=(0,W.useCallback)(async()=>{if(!e)try{let[e,t]=await Promise.all([J.processDashboard(),J.processPoolStats()]);n(e),i(t),c(null)}catch(e){c(String(e))}finally{o(!1)}},[e]);(0,W.useEffect)(()=>{if(e)return;let t=null,r=null;function a(){t?.close(),r?.close()}function s(){p.current||=(f(`polling`),setInterval(m,3e3))}try{t=J.processDashboardStream(),r=J.processPoolStatsStream();let e=!1,l=!1;t.onopen=()=>{e=!0,e&&l&&f(`streaming`)},r.onopen=()=>{l=!0,e&&l&&f(`streaming`)};let u=()=>{f(`error`),a(),s()};t.onerror=u,r.onerror=u,t.onmessage=e=>{try{n(JSON.parse(e.data)),c(null),o(!1)}catch{c(`Stream parse error`)}},r.onmessage=e=>{try{i(JSON.parse(e.data)),c(null)}catch{c(`Pool stats stream error`)}}}catch{f(`error`),s()}return()=>{a(),p.current&&=(clearInterval(p.current),null)}},[e,m]),(0,W.useEffect)(()=>{if(!e){if(d!==`polling`){p.current&&=(clearInterval(p.current),null);return}return m(),p.current||=setInterval(m,3e3),()=>{p.current&&=(clearInterval(p.current),null)}}},[e,m,d]);let h=(0,W.useCallback)(async(e,t)=>{try{let n=await e();u(n.success?n.message??`${t} OK`:n.error??`${t} failed`)}catch(e){u(String(e))}setTimeout(()=>u(null),3e3),await m(),t===`Terminated`&&n(e=>e&&{...e,processes:e.processes.filter(e=>e.status!==`terminated`)})},[m]);return{data:t,poolStats:r,loading:a,error:s,actionMsg:l,streamStatus:d,fetchData:m,pauseProcess:e=>h(()=>J.pauseProcess(e),`Paused`),resumeProcess:e=>h(()=>J.resumeProcess(e),`Resumed`),terminateProcess:e=>h(()=>J.terminateProcess(e),`Terminated`)}}function Tn({data:e,streamStatus:t}){let n=e?.system_load,r=t===`streaming`?`Live`:t===`polling`?`Polling`:`Offline`,i=t===`streaming`?`var(--green)`:t===`polling`?`var(--blue)`:`var(--red)`,a=e?.timestamp?.slice(11,19)??`n/a`;return(0,Z.jsxs)(`div`,{className:`kpi-row`,children:[(0,Z.jsx)(Q,{icon:(0,Z.jsx)(M,{size:20}),label:`Worker Pool`,value:r,sub:`updated ${a}`,accent:i}),(0,Z.jsx)(Q,{icon:(0,Z.jsx)(l,{size:20}),label:`CPU`,value:n?`${n.cpu_percent.toFixed(1)}%`:`n/a`,sub:`system load`,accent:`var(--green)`}),(0,Z.jsx)(Q,{icon:(0,Z.jsx)(V,{size:20}),label:`Memory`,value:n?`${n.memory_percent.toFixed(1)}%`:`n/a`,sub:`system memory`,accent:`var(--blue)`}),(0,Z.jsx)(Q,{icon:(0,Z.jsx)(F,{size:20}),label:`Connections`,value:n?n.active_connections:`n/a`,sub:`active sockets`,accent:`var(--amber)`})]})}function En({processes:e,actionMsg:t,onRefresh:n,onPause:r,onResume:i,onTerminate:a}){return(0,Z.jsxs)(`section`,{className:`section`,children:[(0,Z.jsxs)(`div`,{className:`section-header`,children:[(0,Z.jsxs)(`h3`,{children:[`Active Processes `,(0,Z.jsx)(`span`,{className:`badge`,children:e.length})]}),(0,Z.jsxs)(`div`,{style:{display:`flex`,gap:8,alignItems:`center`},children:[t&&(0,Z.jsx)(`span`,{className:`section-meta`,style:{color:`var(--amber)`},children:t}),(0,Z.jsx)(`button`,{className:`icon-btn`,onClick:n,title:`Refresh`,children:(0,Z.jsx)(U,{size:14})})]})]}),e.length===0?(0,Z.jsxs)(`div`,{className:`tasks-empty`,children:[(0,Z.jsx)(o,{size:32,color:`var(--text-dim)`}),(0,Z.jsx)(`p`,{children:`No active processes`})]}):(0,Z.jsxs)(`div`,{className:`tasks-table`,children:[(0,Z.jsxs)(`div`,{className:`tasks-thead`,children:[(0,Z.jsx)(`span`,{children:`PID`}),(0,Z.jsx)(`span`,{children:`Command`}),(0,Z.jsx)(`span`,{children:`Status`}),(0,Z.jsx)(`span`,{children:`Progress`}),(0,Z.jsx)(`span`,{children:`Runtime`}),(0,Z.jsx)(`span`,{children:`ETA`}),(0,Z.jsx)(`span`,{children:`Actions`})]}),e.map(e=>{let t=parseFloat(e.progress_percent)||0,n=e.status===`running`?`var(--green)`:e.status===`paused`?`var(--amber)`:`var(--text-dim)`;return(0,Z.jsxs)(`div`,{className:`tasks-row`,children:[(0,Z.jsx)(`span`,{className:`mono tasks-pid`,children:e.pid}),(0,Z.jsx)(`span`,{className:`mono tasks-cmd`,title:e.command,children:e.command}),(0,Z.jsx)(`span`,{className:`tasks-status tasks-status--${e.status}`,children:e.status}),(0,Z.jsxs)(`div`,{className:`tasks-progress`,children:[(0,Z.jsx)(`div`,{className:`tasks-progress-bar-bg`,children:(0,Z.jsx)(`div`,{className:`tasks-progress-bar-fill`,style:{width:`${Math.min(100,t)}%`,background:n}})}),(0,Z.jsx)(`span`,{className:`tasks-pct mono`,children:e.progress_percent})]}),(0,Z.jsx)(`span`,{className:`mono`,children:e.runtime}),(0,Z.jsx)(`span`,{className:`mono`,children:e.eta}),(0,Z.jsxs)(`div`,{className:`tasks-actions`,children:[e.status!==`paused`&&(0,Z.jsx)(`button`,{className:`tasks-btn tasks-btn--pause`,title:`Pause`,onClick:()=>r(e.pid),children:(0,Z.jsx)(E,{size:14})}),e.status===`paused`&&(0,Z.jsx)(`button`,{className:`tasks-btn tasks-btn--resume`,title:`Resume`,onClick:()=>i(e.pid),children:(0,Z.jsx)(d,{size:14})}),(0,Z.jsx)(`button`,{className:`tasks-btn tasks-btn--stop`,title:`Terminate`,onClick:()=>a(e.pid),children:(0,Z.jsx)(f,{size:14})})]})]},e.pid)})]})]})}function Dn({demoData:e}){let{data:t,poolStats:n,loading:r,error:i,actionMsg:a,streamStatus:o,fetchData:s,pauseProcess:l,resumeProcess:u,terminateProcess:d}=wn(e);if(r&&!t)return(0,Z.jsxs)(`div`,{className:`loading-state`,children:[(0,Z.jsx)(U,{size:20,className:`spin`,color:`var(--green)`}),(0,Z.jsx)(`p`,{children:`Loading tasks…`})]});if(i&&!t)return(0,Z.jsxs)(`div`,{className:`error-banner`,children:[(0,Z.jsx)(c,{size:16}),` `,i]});let f=(t?.processes??[]).slice(-100);return(0,Z.jsxs)(`div`,{className:`page-content`,children:[n&&(0,Z.jsx)(Tn,{data:t,streamStatus:o}),(0,Z.jsx)(En,{processes:f,actionMsg:a,onRefresh:s,onPause:l,onResume:u,onTerminate:d})]})}function On({tools:e,filtered:t,categories:n,activeCat:r,setActiveCat:i,search:a,setSearch:o,missingOnly:s,setMissingOnly:l,missingCount:u,toolsStatus:d,onSelectTool:f}){return(0,Z.jsxs)(`section`,{className:`section`,children:[(0,Z.jsx)(`div`,{className:`section-header`,children:(0,Z.jsxs)(`h3`,{children:[`Tool Registry `,(0,Z.jsxs)(`span`,{className:`badge`,children:[t.length,` / `,e.length]})]})}),(0,Z.jsxs)(`div`,{className:`registry-controls`,children:[(0,Z.jsxs)(`div`,{className:`registry-controls-top`,children:[(0,Z.jsxs)(`div`,{className:`search-input-wrap`,children:[(0,Z.jsx)(`input`,{className:`search-input mono`,placeholder:`Search tools…`,value:a,onChange:e=>o(e.target.value)}),a.trim().length>0&&(0,Z.jsx)(`button`,{className:`search-clear-btn`,onClick:()=>o(``),title:`Clear search`,"aria-label":`Clear search`,children:(0,Z.jsx)(c,{size:12})})]}),(0,Z.jsxs)(`button`,{className:`registry-missing-toggle${s?` active`:``}`,onClick:()=>l(e=>!e),title:`Show only tools not installed`,children:[(0,Z.jsx)(c,{size:12}),`Not installed`,u>0&&(0,Z.jsx)(`span`,{className:`badge`,children:u})]})]}),(0,Z.jsx)(`div`,{className:`cat-tabs`,children:n.map(e=>(0,Z.jsx)(`button`,{className:`cat-tab ${r===e?`active`:``}`,onClick:()=>i(r===e?`all`:e),children:e.replace(/_/g,` `)},e))})]}),(0,Z.jsxs)(`div`,{className:`registry-grid registry-grid--wide`,children:[t.map(e=>(0,Z.jsxs)(`div`,{className:`registry-card registry-card--clickable`,onClick:()=>f(e),title:`Click for details on ${e.name}`,children:[(0,Z.jsxs)(`div`,{className:`registry-card-top`,children:[(0,Z.jsx)(`span`,{className:`registry-name mono`,title:e.name,children:e.name}),(0,Z.jsxs)(`div`,{className:`registry-meta`,children:[e.parent_tool&&(0,Z.jsx)(`span`,{className:`registry-cat`,title:`Based on ${e.parent_tool}`,children:e.parent_tool}),d[e.name]===!0&&(0,Z.jsx)(`span`,{className:`registry-installed`,title:`Installed`,children:(0,Z.jsx)(T,{size:11,color:`var(--green)`})}),d[e.name]===!1&&(0,Z.jsx)(`span`,{className:`registry-installed`,title:`Not installed`,children:(0,Z.jsx)(c,{size:11,color:`var(--red)`})})]})]}),(0,Z.jsx)(`p`,{className:`registry-desc`,children:e.desc}),(0,Z.jsxs)(`div`,{className:`registry-footer`,children:[(0,Z.jsx)(`span`,{className:`registry-cat`,children:e.category.replace(/_/g,` `)}),(0,Z.jsxs)(`span`,{className:`registry-eff`,title:`Effectiveness`,children:[`█`.repeat(Math.round(e.effectiveness*5)),`░`.repeat(5-Math.round(e.effectiveness*5))]})]})]},e.name)),t.length===0&&(0,Z.jsx)(`p`,{className:`empty-state`,children:`No tools match your filter.`})]})]})}function kn({health:e,tools:t,toolsStatus:n}){let[r,i]=(0,W.useState)(``),[a,o]=(0,W.useState)(`all`),[s,l]=(0,W.useState)(null),[u,d]=(0,W.useState)(!1),f=It(t),p=Lt(t,{toolsStatus:n,activeCategory:a,search:r,missingOnly:u,includeParentToolSearch:!0}),m=e.total_tools_count-e.total_tools_available;return(0,Z.jsxs)(`div`,{className:`page-content tools-page`,children:[s&&(0,Z.jsx)(Tt,{tool:s,onClose:()=>l(null),installed:n[s.name]}),(0,Z.jsxs)(`div`,{className:`kpi-row`,children:[(0,Z.jsx)(Q,{icon:(0,Z.jsx)(z,{size:20}),label:`Total Server Tools`,value:t.length,sub:`in registry`,accent:`var(--blue)`}),(0,Z.jsx)(Q,{icon:(0,Z.jsx)(se,{size:20}),label:`Kali Tools Installed`,value:`${e.total_tools_available} / ${e.total_tools_count}`,sub:`${(e.total_tools_available/Math.max(e.total_tools_count,1)*100).toFixed(0)}% coverage`,accent:`var(--green)`}),(0,Z.jsx)(Q,{icon:(0,Z.jsx)(c,{size:20}),label:`Missing`,value:m,sub:`not installed`,accent:m>0?`var(--amber)`:`var(--text-dim)`}),(0,Z.jsx)(Q,{icon:(0,Z.jsx)(C,{size:20}),label:`Categories`,value:f.length-1,sub:`tool categories`,accent:`var(--purple)`})]}),(0,Z.jsx)(On,{tools:t,filtered:p,categories:f,activeCat:a,setActiveCat:o,search:r,setSearch:i,missingOnly:u,setMissingOnly:d,missingCount:m,toolsStatus:n,onSelectTool:l})]})}function An(e){for(let t of[`target`,`url`,`host`,`ip`,`domain`,`file`]){let n=e.params[t];if(n)return String(n)}let t=Object.values(e.params)[0];return t?String(t):`(no target)`}function jn(e){let t=new Map;for(let n of e){let e=n.ts.toLocaleDateString(`en-GB`,{day:`2-digit`,month:`short`,year:`numeric`});t.has(e)||t.set(e,[]),t.get(e)?.push(n)}return Array.from(t.entries()).map(([e,t])=>({label:e,entries:t}))}function Mn(e){let t=e.filter(e=>e.result.success).length,n=e.reduce((e,t)=>e+t.result.execution_time,0)/e.length,r=e.reduce((e,t)=>e.ts>t.ts?e:t);return{total:e.length,ok:t,failed:e.length-t,avgTime:n,last:r}}function Nn({runHistory:e,onOpenEntry:t}){let n=[...e].sort((e,t)=>e.ts.getTime()-t.ts.getTime()).slice(-60),r=jn(n);return(0,Z.jsxs)(`section`,{className:`section`,children:[(0,Z.jsx)(`div`,{className:`section-header`,children:(0,Z.jsxs)(`h3`,{children:[`Run Timeline `,(0,Z.jsxs)(`span`,{className:`section-meta`,children:[`last `,n.length]})]})}),(0,Z.jsx)(`div`,{className:`reports-timeline-wrap`,children:r.map(({label:e,entries:n})=>(0,Z.jsxs)(`div`,{className:`reports-timeline-group`,children:[(0,Z.jsx)(`span`,{className:`reports-timeline-date`,children:e}),(0,Z.jsx)(`div`,{className:`reports-timeline-dots`,children:n.map((e,n)=>(0,Z.jsx)(`button`,{className:`reports-timeline-dot ${e.result.success?`ok`:`fail`}`,title:`${e.tool} — ${e.ts.toLocaleTimeString(`en-GB`)} — ${e.result.success?`ok`:`failed`} (${e.result.execution_time.toFixed(1)}s)`,onClick:()=>t(e)},n))})]},e))})]})}function Pn({grouped:e,keys:n,groupBy:r,search:i,setSearch:a,setGroupBy:o,expanded:s,toggleExpanded:c,onOpenEntry:l}){return(0,Z.jsxs)(`section`,{className:`section`,children:[(0,Z.jsxs)(`div`,{className:`section-header`,children:[(0,Z.jsx)(`h3`,{children:`Breakdown`}),(0,Z.jsxs)(`div`,{style:{display:`flex`,gap:8,alignItems:`center`},children:[(0,Z.jsx)(`input`,{className:`search-input mono`,style:{width:180},placeholder:`Search…`,value:i,onChange:e=>a(e.target.value)}),(0,Z.jsx)(`button`,{className:`cat-tab ${r===`tool`?`active`:``}`,onClick:()=>o(`tool`),children:`By Tool`}),(0,Z.jsx)(`button`,{className:`cat-tab ${r===`target`?`active`:``}`,onClick:()=>o(`target`),children:`By Target`})]})]}),(0,Z.jsxs)(`div`,{className:`reports-table`,children:[(0,Z.jsxs)(`div`,{className:`reports-thead`,children:[(0,Z.jsx)(`span`,{}),(0,Z.jsx)(`span`,{children:r===`tool`?`Tool`:`Target`}),(0,Z.jsx)(`span`,{children:`Runs`}),(0,Z.jsx)(`span`,{children:`Success`}),(0,Z.jsx)(`span`,{children:`Failed`}),(0,Z.jsx)(`span`,{children:`Avg Time`}),(0,Z.jsx)(`span`,{children:`Last Run`})]}),n.map(n=>{let r=e[n],i=Mn(r),a=i.ok/i.total*100,o=a>=80?`var(--green)`:a>=50?`var(--amber)`:`var(--red)`,u=s.has(n),d=[...r].sort((e,t)=>t.ts.getTime()-e.ts.getTime());return(0,Z.jsxs)(`div`,{className:`reports-group`,children:[(0,Z.jsxs)(`button`,{className:`reports-row reports-row--clickable`,onClick:()=>c(n),children:[(0,Z.jsx)(`span`,{className:`reports-chevron`,children:u?(0,Z.jsx)(N,{size:12}):(0,Z.jsx)(t,{size:12})}),(0,Z.jsx)(`span`,{className:`mono reports-key`,children:n}),(0,Z.jsx)(`span`,{className:`mono`,children:i.total}),(0,Z.jsx)(`span`,{className:`mono`,style:{color:`var(--green)`},children:i.ok}),(0,Z.jsx)(`span`,{className:`mono`,style:{color:i.failed>0?`var(--red)`:`var(--text-dim)`},children:i.failed}),(0,Z.jsxs)(`span`,{className:`mono`,children:[i.avgTime.toFixed(1),`s`]}),(0,Z.jsxs)(`div`,{className:`reports-last-cell`,children:[(0,Z.jsx)(`span`,{className:`reports-pct-bar-bg`,children:(0,Z.jsx)(`span`,{className:`reports-pct-bar-fill`,style:{width:`${a}%`,background:o}})}),(0,Z.jsxs)(`span`,{className:`mono`,style:{fontSize:11,color:`var(--text-dim)`},children:[i.last.ts.toLocaleDateString(`en-GB`),` `,i.last.ts.toLocaleTimeString(`en-GB`)]})]})]}),u&&(0,Z.jsx)(`div`,{className:`reports-runs`,children:d.map((e,t)=>(0,Z.jsxs)(`button`,{className:`reports-run-row`,onClick:()=>l(e),children:[(0,Z.jsx)(`span`,{className:`reports-run-dot ${e.result.success?`ok`:`fail`}`}),(0,Z.jsx)(`span`,{className:`mono reports-run-tool`,children:e.tool}),(0,Z.jsxs)(`span`,{className:`mono reports-run-time`,children:[e.ts.toLocaleDateString(`en-GB`),` `,e.ts.toLocaleTimeString(`en-GB`)]}),(0,Z.jsxs)(`span`,{className:`mono reports-run-duration`,children:[e.result.execution_time.toFixed(1),`s`]}),Object.entries(e.params).map(([e,t])=>(0,Z.jsxs)(`span`,{className:`mono reports-run-param`,children:[e,`=`,(0,Z.jsx)(`em`,{children:String(t)})]},e))]},t))})]},n)})]})]})}function Fn({runHistory:e}){let[t,n]=(0,W.useState)(`tool`),[r,i]=(0,W.useState)(``),[a,o]=(0,W.useState)(new Set),[s,c]=(0,W.useState)(null),l=e.reduce((e,t)=>((e[t.tool]=e[t.tool]||[]).push(t),e),{}),d=e.reduce((e,t)=>{let n=An(t);return(e[n]=e[n]||[]).push(t),e},{}),f=t===`tool`?l:d;function p(e){o(t=>{let n=new Set(t);return n.has(e)?n.delete(e):n.add(e),n})}let m=r.toLowerCase(),h=Object.keys(f).filter(e=>!m||e.toLowerCase().includes(m)).sort();return e.length===0?(0,Z.jsx)(`div`,{className:`page-content`,children:(0,Z.jsxs)(`div`,{className:`tasks-empty`,children:[(0,Z.jsx)(u,{size:32,color:`var(--text-dim)`}),(0,Z.jsx)(`p`,{children:`No run history yet. Execute tools from the Run tab to see reports.`})]})}):(0,Z.jsxs)(`div`,{className:`page-content`,children:[s&&(0,Z.jsx)(jt,{entry:s,onClose:()=>c(null)}),(0,Z.jsxs)(`div`,{className:`kpi-row`,children:[(0,Z.jsx)(Q,{icon:(0,Z.jsx)(R,{size:20}),label:`Total Runs`,value:e.length,sub:`all time`,accent:`var(--blue)`}),(0,Z.jsx)(Q,{icon:(0,Z.jsx)(T,{size:20}),label:`Success Rate`,value:`${(e.filter(e=>e.result.success).length/e.length*100).toFixed(0)}%`,sub:`${e.filter(e=>e.result.success).length} ok · ${e.filter(e=>!e.result.success).length} failed`,accent:`var(--green)`}),(0,Z.jsx)(Q,{icon:(0,Z.jsx)(w,{size:20}),label:`Avg Time`,value:`${(e.reduce((e,t)=>e+t.result.execution_time,0)/e.length).toFixed(1)}s`,sub:`per run`,accent:`var(--purple)`}),(0,Z.jsx)(Q,{icon:(0,Z.jsx)(ee,{size:20}),label:`Unique Tools`,value:Object.keys(l).length,sub:`used`,accent:`var(--amber)`})]}),(0,Z.jsx)(Nn,{runHistory:e,onOpenEntry:c}),(0,Z.jsx)(Pn,{grouped:f,keys:h,groupBy:t,search:r,setSearch:i,setGroupBy:n,expanded:a,toggleExpanded:p,onOpenEntry:c})]})}var In=[{key:`intelligence`,title:`Intelligent Attack-Chain`,description:`Automated target profiling and attack chain generation.`,details:`Best for quick, AI-driven assessments.`,modalDescription:`Leverages AI to analyze the target and generate a customized attack chain.`,tools:[],placeholder:`Domain or org target (target.tld)`},{key:`manual`,title:`Manual Session`,description:`Start an empty session and add tools yourself.`,details:`Best when you want full manual control.`,modalDescription:`Creates a clean session with only target context. No predefined workflow is added, so you can build your own tool chain step-by-step from the session detail page.`,tools:[],placeholder:`Target URL/domain/IP (example.com)`},{key:`from_template`,title:`From Template`,description:`Start from a saved tool template.`,details:`Reuse your recurring workflows.`,modalDescription:`Creates a session from an existing template. You only set target and choose the saved template; all template tools are copied into the new session.`,tools:[],placeholder:`Target URL/domain/IP (example.com)`}];function Ln(e){let t=e.metadata??{},n=t.session_name;if(typeof n==`string`&&n.trim()){let e=n.trim();return e.charAt(0).toUpperCase()+e.slice(1)}let r=typeof e.objective==`string`&&e.objective||(typeof t.mode==`string`?t.mode:``);if(r){let e=r.replace(/_/g,` `).trim();return e.charAt(0).toUpperCase()+e.slice(1)}return`Session`}function Rn({session:e,onOpen:t}){let n=e.metadata?.tool_status??{},r=e.metadata?.last_run??null;return(0,Z.jsxs)(`div`,{className:`session-card session-card--compact registry-card--clickable`,onClick:()=>t(e.session_id),children:[(0,Z.jsxs)(`div`,{className:`session-card-header`,children:[(0,Z.jsxs)(`div`,{className:`session-target`,children:[(0,Z.jsx)(te,{size:13,color:`var(--blue)`}),(0,Z.jsx)(`span`,{className:`mono`,children:e.target})]}),e.status&&(0,Z.jsx)(`span`,{className:`session-status session-status--${e.status}`,children:e.status})]}),(0,Z.jsxs)(`div`,{className:`session-card-meta`,children:[(0,Z.jsxs)(`span`,{children:[(0,Z.jsx)(M,{size:11}),` `,e.total_findings,` findings`]}),(0,Z.jsxs)(`span`,{children:[(0,Z.jsx)(U,{size:11}),` `,e.iterations,` iterations`]}),(0,Z.jsxs)(`span`,{children:[(0,Z.jsx)(w,{size:11}),` `,dt(e.updated_at)]})]}),(0,Z.jsxs)(`div`,{className:`session-tools`,children:[e.tools_executed.slice(0,7).map(e=>(0,Z.jsx)(`span`,{className:`session-tool-chip mono session-tool-chip--${n[e]===`success`?`success`:n[e]===`failed`?`failed`:`idle`}`,children:e},e)),e.tools_executed.length>7&&(0,Z.jsxs)(`span`,{className:`session-tool-chip session-tool-chip--more`,children:[`+`,e.tools_executed.length-7]})]}),(0,Z.jsxs)(`div`,{className:`session-card-footer`,children:[(0,Z.jsx)(`span`,{className:`session-id mono`,children:e.session_id}),(0,Z.jsx)(`span`,{className:`session-tool-chip mono`,children:Ln(e)})]}),r?.tool&&(0,Z.jsxs)(`div`,{className:`session-last-run mono`,children:[`last: `,r.tool,` | `,r.success?`OK`:`FAIL`,` | exit `,r.return_code??0,` | `,(r.execution_time??0).toFixed(2),`s`]})]})}function zn({startModes:e,onOpenStartMode:t,createMsg:n}){return(0,Z.jsxs)(`section`,{className:`section`,children:[(0,Z.jsxs)(`div`,{className:`section-header`,children:[(0,Z.jsx)(`h3`,{children:`Start Session`}),(0,Z.jsx)(`span`,{className:`section-meta`,children:`Choose a workflow type, then provide target details.`})]}),(0,Z.jsx)(`div`,{className:`registry-grid registry-grid--wider start-mode-grid`,children:e.map(e=>(0,Z.jsxs)(`div`,{className:`registry-card registry-card--clickable start-mode-card`,onClick:()=>t(e),children:[(0,Z.jsxs)(`div`,{className:`registry-card-top`,children:[(0,Z.jsx)(`span`,{className:`registry-name`,children:e.title}),(0,Z.jsx)(`span`,{className:`registry-cat`,children:e.key.replace(/_/g,` `)})]}),(0,Z.jsx)(`p`,{className:`registry-desc`,children:e.description}),(0,Z.jsx)(`p`,{className:`start-mode-detail`,children:e.details})]},e.key))}),n&&(0,Z.jsx)(`p`,{className:`section-meta`,children:n})]})}function Bn({title:e,sessions:t,emptyText:n,onOpenSession:r,headerRight:i,onHeaderClick:a,footer:o}){return(0,Z.jsxs)(`section`,{className:`section`,children:[(0,Z.jsxs)(`div`,{className:`section-header${a?` sessions-collapsed-toggle`:``}`,...a?{role:`button`,tabIndex:0,onClick:a,onKeyDown:e=>{(e.key===`Enter`||e.key===` `)&&(e.preventDefault(),a())}}:{},children:[(0,Z.jsxs)(`h3`,{children:[e,` `,(0,Z.jsx)(`span`,{className:`badge`,children:t.length})]}),i]}),t.length===0?(0,Z.jsxs)(`div`,{className:`tasks-empty`,children:[(0,Z.jsx)(g,{size:28,color:`var(--text-dim)`}),(0,Z.jsx)(`p`,{children:n})]}):(0,Z.jsx)(`div`,{className:`sessions-grid`,children:t.map(e=>(0,Z.jsx)(Rn,{session:e,onOpen:r},e.session_id))}),o&&(0,Z.jsx)(`div`,{className:`section-meta session-list-footer-tip`,children:o})]})}function Vn({startMode:e,templates:t,selectedTemplateId:n,setSelectedTemplateId:r,modalTarget:i,setModalTarget:a,modalNote:o,setModalNote:s,modalError:c,creatingSession:l,onClose:u,onSubmit:d}){let f=(0,W.useRef)(null),p=(0,W.useRef)(null),m=(0,W.useMemo)(()=>t.find(e=>e.template_id===n),[t,n]),h=(0,W.useMemo)(()=>{if(e.key!==`from_template`||!m)return e.tools;let t=Array.isArray(m.workflow_steps)?m.workflow_steps:[],n=[];for(let e of t){let t=e?.tool;!t||n.includes(t)||n.push(t)}return n},[e,m]);return(0,W.useEffect)(()=>{let t=window.requestAnimationFrame(()=>{if(e.key===`from_template`){f.current?.focus();return}p.current?.focus()});return()=>window.cancelAnimationFrame(t)},[e.key]),(0,Z.jsx)(`div`,{className:`modal-backdrop`,onClick:e=>{e.target===e.currentTarget&&u()},children:(0,Z.jsxs)(`div`,{className:`modal`,children:[(0,Z.jsxs)(`div`,{className:`modal-header`,children:[(0,Z.jsx)(`div`,{className:`modal-title-row`,children:(0,Z.jsxs)(`span`,{className:`modal-name`,children:[`Start `,e.title]})}),(0,Z.jsx)(`button`,{className:`modal-close`,onClick:u,children:`x`})]}),(0,Z.jsxs)(`div`,{className:`modal-body`,children:[(0,Z.jsx)(`p`,{className:`modal-desc`,children:e.modalDescription}),(0,Z.jsxs)(`div`,{className:`modal-section`,children:[(0,Z.jsx)(`span`,{className:`modal-label`,children:`Typical Tooling`}),(0,Z.jsxs)(`div`,{className:`modal-params`,children:[h.length===0&&(0,Z.jsx)(`span`,{className:`modal-param mono`,children:`none preloaded`}),h.map(e=>(0,Z.jsx)(`span`,{className:`modal-param mono`,children:e},e))]})]}),e.key===`from_template`&&(0,Z.jsxs)(`div`,{className:`session-start-form`,children:[(0,Z.jsx)(`label`,{className:`mono`,children:`Template *`}),(0,Z.jsxs)(`select`,{ref:f,className:`session-objective-select`,value:n,onChange:e=>r(e.target.value),children:[(0,Z.jsx)(`option`,{value:``,children:`Select template`}),t.map(e=>(0,Z.jsx)(`option`,{value:e.template_id,children:e.name},e.template_id))]})]}),(0,Z.jsxs)(`div`,{className:`session-start-form`,children:[(0,Z.jsx)(`label`,{className:`mono`,htmlFor:`session-target-input`,children:`Target *`}),(0,Z.jsx)(`input`,{id:`session-target-input`,ref:p,className:`search-input mono`,value:i,onChange:e=>a(e.target.value),placeholder:e.placeholder}),(0,Z.jsx)(`label`,{className:`mono`,htmlFor:`session-note-input`,children:`Note (optional)`}),(0,Z.jsx)(`textarea`,{id:`session-note-input`,className:`session-step-params mono`,rows:3,value:o,onChange:e=>s(e.target.value),placeholder:`Context for this run`}),c&&(0,Z.jsx)(`div`,{className:`run-error`,children:c}),(0,Z.jsxs)(`div`,{className:`session-start-actions`,children:[(0,Z.jsx)(`button`,{className:`session-action-btn`,onClick:u,children:`Cancel`}),(0,Z.jsxs)(`button`,{className:`session-run-btn`,onClick:d,disabled:l,children:[l?(0,Z.jsx)(U,{size:13,className:`spin`}):(0,Z.jsx)(te,{size:13}),l?`Starting…`:`Start Session`]})]})]})]})]})})}function Hn({demoData:e,onOpenSession:t}){let{pushToast:n}=et(),[r,i]=(0,W.useState)(e?.sessions??null),[a,o]=(0,W.useState)(!1),[s,l]=(0,W.useState)([]),[u,d]=(0,W.useState)([]),[f,p]=(0,W.useState)(null),[m,h]=(0,W.useState)(null),[_,v]=(0,W.useState)(null),[y,x]=(0,W.useState)(null),[S,C]=(0,W.useState)(null),[E,D]=(0,W.useState)(``),[O,k]=(0,W.useState)(``),[A,j]=(0,W.useState)(``),[ee,N]=(0,W.useState)(null),[P,ne]=(0,W.useState)(``),[re,F]=(0,W.useState)(``),[I,L]=(0,W.useState)([]),[oe,se]=(0,W.useState)(``),[ce,R]=(0,W.useState)(null),[le,ue]=(0,W.useState)(!1),[de,fe]=(0,W.useState)(!e),[z,B]=(0,W.useState)(null),[pe,me]=(0,W.useState)(!0),[he,ge]=(0,W.useState)(!1),[_e,ve]=(0,W.useState)(!0),[V,ye]=(0,W.useState)(e?`polling`:`streaming`),be=(0,W.useRef)(null),xe=(0,W.useRef)(null),Se=(0,W.useRef)(null),Ce=(0,W.useRef)(0),we=(0,W.useRef)(!1),Te=(0,W.useRef)(!1);(0,W.useEffect)(()=>{e||Promise.all([J.sessions(),J.tools(),J.sessionTemplates()]).then(([e,t,n])=>{i(e),l(t.tools??[]),d(n.templates??[]),B(null)}).catch(e=>B(String(e))).finally(()=>fe(!1))},[e]),(0,W.useEffect)(()=>{if(we.current||de||!r)return;let e=(r?.active??[]).filter(e=>(e.status??`active`)!==`completed`).length;me(e>0),ve(u.length>0),we.current=!0},[de,r,u.length]),(0,W.useEffect)(()=>{de||Te.current||u.length>0&&ve(!0)},[de,u.length]);async function Ee(){e||d((await J.sessionTemplates()).templates??[])}function De(t=!1){e||(t||fe(!0),J.sessions().then(e=>{i(e),B(null)}).catch(e=>B(String(e))).finally(()=>{t||fe(!1)}))}function Oe(t=!1){e||De(t)}(0,W.useEffect)(()=>{if(e)return;let t=null,n=!1;function r(){t&&t.close(),t=null,be.current=null}function a(){xe.current&&=(clearInterval(xe.current),null)}function o(){Se.current&&=(clearTimeout(Se.current),null)}function s(){a(),ye(`polling`),Oe(!0),xe.current=setInterval(()=>Oe(!0),5e3)}function c(e){o();let t=Ce.current+1;Ce.current=t;let r=Math.min(3e4,1e3*2**Math.max(0,t-1));Se.current=setTimeout(()=>{n||e()},r)}function l(){r();try{t=J.sessionsStream(),be.current=t,t.onopen=()=>{Ce.current=0,o(),a(),ye(`streaming`)},t.onmessage=e=>{try{let t=JSON.parse(e.data);t?.success&&(i(t),B(null),fe(!1))}catch{B(`Session stream parse error`)}},t.onerror=()=>{n||(ye(`error`),r(),s(),c(l))}}catch{if(n)return;ye(`error`),s(),c(l)}}return l(),()=>{n=!0,r(),a(),o()}},[e]);function ke(e,t=``){C(e),D(``),k(``),j(t),N(null)}function Ae(){C(null),N(null)}function je(e){ne(e.template_id),F(e.name),L(e.workflow_steps??[]),se(``),R(null),h(null)}function Me(){ne(``),F(``),L([]),se(``),R(null)}function Ne(e){let t=In.find(e=>e.key===`from_template`);t&&ke(t,e)}async function Pe(t){if(!e){v(t),h(null);try{await J.deleteSessionTemplate(t),d(e=>e.filter(e=>e.template_id!==t)),A===t&&j(``),p(`Template deleted.`),n(`success`,`Template deleted`)}catch(e){let t=String(e);h(t),n(`error`,`Delete failed: ${t}`)}finally{v(null),x(null)}}}function Fe(e){L(t=>[...t,{tool:e,parameters:{}}])}function G(e){L(t=>t.filter((t,n)=>n!==e))}async function K(){if(!P)return;let e=re.trim();if(!e){R(`Template name is required`),n(`error`,`Template name is required`);return}if(I.length===0){R(`Template must include at least one tool`),n(`error`,`Template must include at least one tool`);return}ue(!0),R(null),h(null);try{await J.updateSessionTemplate(P,{name:e,workflow_steps:I}),await Ee(),p(`Template updated.`),n(`success`,`Template updated`),Me()}catch(e){let t=String(e);R(t),n(`error`,`Save failed: ${t}`)}finally{ue(!1)}}async function Ie(e,t,r){if(!t.trim()){N(`Target is required`),n(`error`,`Target is required`);return}p(null),N(null),o(!0);try{let i=t.trim(),a;if(e.key===`manual`)a=await J.createSession({target:i,workflow_steps:[],source:`web`,objective:e.key,metadata:{origin:`ui/sessions/create`,mode:e.key,note:r,session_name:`Manual session`}});else if(e.key===`from_template`){let e=u.find(e=>e.template_id===A);if(!e){N(`Template is required`),n(`error`,`Template is required`);return}a=await J.createSessionFromTemplate({target:i,template_id:e.template_id,source:`web`,objective:`from_template`,metadata:{origin:`ui/sessions/create`,mode:`from_template`,note:r,template_id:e.template_id,session_name:`Template: ${e.name}`}})}else{let t=await J.createAttackChain(i,e.key);a=await J.createSession({target:i,workflow_steps:t.attack_chain.steps,source:`web`,objective:e.key,session_id:t.session_id,metadata:{origin:`ui/sessions/create`,mode:e.key,note:r}})}let o=a.session.session_id;n(`success`,`Session created: ${o}`),Ae(),Oe()}catch(e){let t=String(e);N(t),p(t),n(`error`,`Session start failed: ${t}`)}finally{o(!1)}}if(de)return(0,Z.jsxs)(`div`,{className:`loading-state`,children:[(0,Z.jsx)(U,{size:20,className:`spin`,color:`var(--green)`}),(0,Z.jsx)(`p`,{children:`Loading sessions…`})]});if(z)return(0,Z.jsxs)(`div`,{className:`error-banner`,children:[(0,Z.jsx)(c,{size:16}),` `,z]});let Le=r?.active??[],q=Le.filter(e=>(e.status??`active`)!==`completed`),Re=[...r?.completed??[],...Le.filter(e=>(e.status??`active`)===`completed`)].filter((e,t,n)=>n.findIndex(t=>t.session_id===e.session_id)===t),ze=[...q,...Re].reduce((e,t)=>e+t.total_findings,0),Y=new Set([...q,...Re].map(e=>e.target)).size,Be=P?u.find(e=>e.template_id===P)??null:null,Ve=s.filter(e=>{if(!oe.trim())return!0;let t=oe.toLowerCase();return e.name.toLowerCase().includes(t)||e.desc.toLowerCase().includes(t)||e.category.toLowerCase().includes(t)}).slice(0,16);return(0,Z.jsxs)(`div`,{className:`page-content`,children:[(0,Z.jsxs)(`div`,{className:`kpi-row`,children:[(0,Z.jsx)(Q,{icon:(0,Z.jsx)(g,{size:20}),label:`Active Sessions`,value:q.length,sub:`in progress`,accent:q.length>0?`var(--green)`:`var(--text-dim)`}),(0,Z.jsx)(Q,{icon:(0,Z.jsx)(T,{size:20}),label:`Completed`,value:Re.length,sub:`archived`,accent:`var(--blue)`}),(0,Z.jsx)(Q,{icon:(0,Z.jsx)(M,{size:20}),label:`Total Findings`,value:ze,sub:`across all sessions`,accent:`var(--amber)`}),(0,Z.jsx)(Q,{icon:(0,Z.jsx)(te,{size:20}),label:`Unique Targets`,value:Y,sub:`scanned`,accent:`var(--purple)`})]}),(0,Z.jsx)(zn,{startModes:In,onOpenStartMode:ke,createMsg:f}),S&&(0,Z.jsx)(Vn,{startMode:S,templates:u,selectedTemplateId:A,setSelectedTemplateId:j,modalTarget:E,setModalTarget:D,modalNote:O,setModalNote:k,modalError:ee,creatingSession:a,onClose:Ae,onSubmit:()=>Ie(S,E,O)}),pe?(0,Z.jsx)(Bn,{title:`Active Sessions`,sessions:q,emptyText:`No active sessions. Start a session from target to run tools manually.`,onOpenSession:t,onHeaderClick:()=>me(!1),footer:(0,Z.jsxs)(Z.Fragment,{children:[(0,Z.jsx)(ie,{size:12}),`Call MCP tool `,(0,Z.jsx)(`span`,{className:`mono`,children:`handover_session("<session_id>", "optional note")`}),` to continue the session with AI.`]}),headerRight:(0,Z.jsxs)(`div`,{className:`sessions-header-actions`,children:[(0,Z.jsx)(`span`,{className:`sessions-stream-status sessions-stream-status--${V}`,children:V===`streaming`?`Live`:V===`polling`?`Polling`:`Offline`}),(0,Z.jsx)(`span`,{className:`session-help-btn`,children:`Hide active`})]})}):(0,Z.jsx)(`section`,{className:`section`,children:(0,Z.jsxs)(`div`,{className:`section-header sessions-collapsed-toggle`,role:`button`,tabIndex:0,onClick:()=>me(!0),onKeyDown:e=>{(e.key===`Enter`||e.key===` `)&&(e.preventDefault(),me(!0))},children:[(0,Z.jsxs)(`h3`,{children:[`Active Sessions `,(0,Z.jsx)(`span`,{className:`badge`,children:q.length})]}),(0,Z.jsx)(`span`,{className:`session-help-btn`,children:`Show active`})]})}),he?(0,Z.jsx)(Bn,{title:`Completed Sessions`,sessions:Re,emptyText:`No completed sessions yet.`,onOpenSession:t,onHeaderClick:()=>ge(!1),headerRight:(0,Z.jsx)(`span`,{className:`session-help-btn`,children:`Hide completed`})}):(0,Z.jsx)(`section`,{className:`section`,children:(0,Z.jsxs)(`div`,{className:`section-header sessions-collapsed-toggle`,role:`button`,tabIndex:0,onClick:()=>ge(!0),onKeyDown:e=>{(e.key===`Enter`||e.key===` `)&&(e.preventDefault(),ge(!0))},children:[(0,Z.jsxs)(`h3`,{children:[`Completed Sessions `,(0,Z.jsx)(`span`,{className:`badge`,children:Re.length})]}),(0,Z.jsx)(`span`,{className:`session-help-btn`,children:`Show completed`})]})}),_e?(0,Z.jsxs)(`section`,{className:`section`,children:[(0,Z.jsxs)(`div`,{className:`section-header sessions-collapsed-toggle`,role:`button`,tabIndex:0,onClick:()=>{Te.current=!0,ve(!1)},onKeyDown:e=>{(e.key===`Enter`||e.key===` `)&&(e.preventDefault(),Te.current=!0,ve(!1))},children:[(0,Z.jsxs)(`h3`,{children:[`Custom Templates `,(0,Z.jsx)(`span`,{className:`badge`,children:u.length})]}),(0,Z.jsx)(`span`,{className:`section-meta`,children:`Reusable user templates for starting sessions quickly.`}),(0,Z.jsx)(`span`,{className:`session-help-btn`,children:`Hide templates`})]}),m&&(0,Z.jsx)(`div`,{className:`run-error`,children:m}),u.length===0?(0,Z.jsxs)(`div`,{className:`tasks-empty`,children:[(0,Z.jsx)(g,{size:28,color:`var(--text-dim)`}),(0,Z.jsx)(`p`,{children:`No custom templates yet. Open a session and click Create Template.`})]}):(0,Z.jsx)(`div`,{className:`sessions-grid`,children:u.map(e=>(0,Z.jsxs)(`div`,{className:`session-card session-card--compact template-session-card`,children:[(0,Z.jsxs)(`div`,{className:`session-card-header`,children:[(0,Z.jsxs)(`div`,{className:`session-target`,children:[(0,Z.jsx)(g,{size:13,color:`var(--blue)`}),(0,Z.jsx)(`span`,{className:`mono`,children:e.name})]}),(0,Z.jsx)(`span`,{className:`session-tool-chip mono`,children:`Custom Template`})]}),(0,Z.jsxs)(`div`,{className:`session-card-meta`,children:[(0,Z.jsxs)(`span`,{children:[(0,Z.jsx)(M,{size:11}),` `,e.workflow_steps.length,` tools`]}),(0,Z.jsxs)(`span`,{children:[(0,Z.jsx)(w,{size:11}),` `,dt(e.updated_at)]})]}),(0,Z.jsxs)(`div`,{className:`session-tools`,children:[e.workflow_steps.slice(0,7).map((t,n)=>(0,Z.jsx)(`span`,{className:`session-tool-chip`,children:t.tool},`${e.template_id}:${n}:${t.tool}`)),e.workflow_steps.length>7&&(0,Z.jsxs)(`span`,{className:`session-tool-chip`,children:[`+`,e.workflow_steps.length-7]})]}),(0,Z.jsx)(`div`,{className:`session-card-footer`,children:(0,Z.jsxs)(`div`,{className:`template-card-actions`,children:[(0,Z.jsxs)(`button`,{className:`session-action-btn`,onClick:()=>Ne(e.template_id),children:[(0,Z.jsx)(H,{size:12}),` Use`]}),(0,Z.jsxs)(`button`,{className:`session-action-btn`,onClick:()=>je(e),children:[(0,Z.jsx)(b,{size:12}),` Edit`]}),(0,Z.jsxs)(`button`,{className:`session-delete-btn`,onClick:()=>x(e),disabled:_===e.template_id,children:[(0,Z.jsx)(ae,{size:12}),` `,_===e.template_id?`Deleting…`:`Delete`]})]})})]},e.template_id))})]}):(0,Z.jsx)(`section`,{className:`section`,children:(0,Z.jsxs)(`div`,{className:`section-header sessions-collapsed-toggle`,role:`button`,tabIndex:0,onClick:()=>{Te.current=!0,ve(!0)},onKeyDown:e=>{(e.key===`Enter`||e.key===` `)&&(e.preventDefault(),Te.current=!0,ve(!0))},children:[(0,Z.jsxs)(`h3`,{children:[`Custom Templates `,(0,Z.jsx)(`span`,{className:`badge`,children:u.length})]}),(0,Z.jsx)(`span`,{className:`session-help-btn`,children:`Show templates`})]})}),Be&&(0,Z.jsx)(`div`,{className:`modal-backdrop`,onClick:e=>{e.target===e.currentTarget&&Me()},children:(0,Z.jsxs)(`div`,{className:`modal modal--wide`,children:[(0,Z.jsxs)(`div`,{className:`modal-header`,children:[(0,Z.jsx)(`div`,{className:`modal-title-row`,children:(0,Z.jsx)(`span`,{className:`modal-name`,children:`Edit Template`})}),(0,Z.jsx)(`button`,{className:`modal-close`,onClick:Me,children:`x`})]}),(0,Z.jsxs)(`div`,{className:`modal-body`,children:[(0,Z.jsxs)(`div`,{className:`session-start-form`,children:[(0,Z.jsx)(`label`,{className:`mono`,htmlFor:`template-name-input`,children:`Template name *`}),(0,Z.jsx)(`input`,{id:`template-name-input`,className:`search-input mono`,value:re,onChange:e=>F(e.target.value),placeholder:`Template name`})]}),(0,Z.jsxs)(`div`,{className:`template-editor-grid`,children:[(0,Z.jsxs)(`div`,{className:`template-editor-col`,children:[(0,Z.jsx)(`span`,{className:`modal-label`,children:`Tools in template`}),I.length===0?(0,Z.jsx)(`div`,{className:`tasks-empty tasks-empty--compact`,children:(0,Z.jsx)(`p`,{children:`No tools selected yet.`})}):(0,Z.jsx)(`div`,{className:`template-step-list`,children:I.map((e,t)=>(0,Z.jsxs)(`div`,{className:`template-step-row`,children:[(0,Z.jsx)(`span`,{className:`mono`,children:e.tool}),(0,Z.jsx)(`button`,{className:`session-remove-tool`,onClick:()=>G(t),children:`x`})]},`${e.tool}:${t}`))})]}),(0,Z.jsxs)(`div`,{className:`template-editor-col`,children:[(0,Z.jsx)(`span`,{className:`modal-label`,children:`Add tools`}),(0,Z.jsx)(`input`,{className:`search-input mono`,value:oe,onChange:e=>se(e.target.value),placeholder:`Search tools`}),(0,Z.jsx)(`div`,{className:`session-add-tool-list`,children:Ve.map(e=>(0,Z.jsxs)(`button`,{className:`session-add-tool-item`,onClick:()=>Fe(e.name),children:[(0,Z.jsx)(`span`,{className:`mono`,children:e.name}),(0,Z.jsx)(`span`,{children:e.category})]},`edit-template-tool:${e.name}`))})]})]}),ce&&(0,Z.jsx)(`div`,{className:`run-error`,children:ce}),(0,Z.jsxs)(`div`,{className:`session-start-actions`,children:[(0,Z.jsx)(`button`,{className:`session-action-btn`,onClick:Me,children:`Cancel`}),(0,Z.jsxs)(`button`,{className:`session-run-btn`,onClick:K,disabled:le,children:[le?(0,Z.jsx)(U,{size:13,className:`spin`}):(0,Z.jsx)(b,{size:13}),le?`Saving…`:`Save Template`]})]})]})]})}),(0,Z.jsx)(Ht,{isOpen:!!y,title:`Delete Template`,description:y?`Delete template "${y.name}"? This action cannot be undone.`:`Delete template?`,impactItems:y?[`Template ID: ${y.template_id}`,`Tools in template: ${y.workflow_steps.length}`,`Future sessions can no longer use this template`]:[],confirmLabel:`Yes, delete template`,cancelLabel:`Keep template`,confirmVariant:`danger`,isConfirming:!!y&&_===y.template_id,onConfirm:async()=>{y&&await Pe(y.template_id)},onClose:()=>{_||x(null)}})]})}function Un(e,t,n,r){ht({id:Date.now(),tool:t,params:n,result:r,ts:new Date(r.timestamp),source:`browser`},e)}function Wn({isCompleted:e,sessionId:t,steps:n,selectedStep:i,selectedStepIndex:a,setSelectedStepIndex:o,selectedStepKey:s,selectedTool:c,stepState:l,runningStepKey:u,targetValue:d,setTargetValue:f,stepFieldValues:p,setStepFieldValues:m,showOptionalByStep:h,setShowOptionalByStep:g,selectedResult:_,onRunStep:y,onStopRunningStep:b,onApplyAttackChainFromResult:x,onRemoveTool:S,showAddTool:C,setShowAddTool:w,addToolSearch:T,setAddToolSearch:E,addCandidates:D,onAddTool:O}){let k=s?u===s:!1,A=_?.result,j=(0,W.useRef)(null);return(0,W.useEffect)(()=>{if(!C)return;let e=window.requestAnimationFrame(()=>j.current?.focus());return()=>window.cancelAnimationFrame(e)},[C]),(0,Z.jsxs)(`section`,{className:`section`,children:[(0,Z.jsx)(`div`,{className:`section-header`,children:(0,Z.jsxs)(`h3`,{children:[`Manual Tool Execution `,(0,Z.jsx)(`span`,{className:`badge`,children:n.length})]})}),(0,Z.jsxs)(`div`,{className:`session-workbench`,children:[(0,Z.jsxs)(`aside`,{className:`session-workbench-tools`,children:[!e&&(0,Z.jsxs)(`div`,{className:`session-tool-manage`,children:[(0,Z.jsx)(`button`,{className:`session-add-tool-btn`,onClick:()=>w(e=>!e),children:`+ Add tool`}),C&&(0,Z.jsxs)(`div`,{className:`session-add-tool-panel`,children:[(0,Z.jsx)(`input`,{ref:j,className:`search-input mono`,placeholder:`Search tool...`,value:T,onChange:e=>E(e.target.value)}),(0,Z.jsx)(`div`,{className:`session-add-tool-list`,children:D.map(e=>(0,Z.jsxs)(`button`,{className:`session-add-tool-item`,onClick:()=>O(e),children:[(0,Z.jsx)(`span`,{className:`mono`,children:e.name}),(0,Z.jsx)(`span`,{children:e.category})]},e.name))})]})]}),n.map((n,r)=>{let i=`${t}:${r}`,s=u===i;return(0,Z.jsxs)(`button`,{className:`session-workbench-tool session-workbench-tool--${l[i]??`idle`} ${s?`session-workbench-tool--running`:``} ${a===r?`active`:``}`,onClick:()=>o(r),children:[(0,Z.jsx)(`span`,{className:`session-workbench-tool-name mono`,children:n.tool}),(0,Z.jsxs)(`span`,{className:`session-workbench-tool-actions`,children:[s&&(0,Z.jsxs)(`span`,{className:`session-running-indicator mono`,title:`Tool is running`,children:[(0,Z.jsx)(U,{size:11,className:`spin`}),` Running`]}),!e&&(0,Z.jsx)(`button`,{type:`button`,className:`session-remove-tool`,onClick:e=>{e.stopPropagation(),S(r)},title:`Remove tool`,children:(0,Z.jsx)(ae,{size:12})})]})]},i)})]}),(0,Z.jsx)(`div`,{className:`session-workbench-center`,children:!i||!s||!c?(0,Z.jsx)(`p`,{className:`empty-state`,children:`No tools in this session.`}):(0,Z.jsxs)(Z.Fragment,{children:[(0,Z.jsxs)(`div`,{className:`session-target-override`,children:[(0,Z.jsx)(`label`,{className:`mono`,children:`Target`}),(0,Z.jsx)(`input`,{className:`search-input mono`,value:d,onChange:e=>f(e.target.value),placeholder:`Target to run this session against`,disabled:e})]}),(0,Z.jsxs)(`div`,{className:`session-step-row`,children:[(0,Z.jsxs)(`div`,{className:`session-step-head`,children:[(0,Z.jsx)(`span`,{className:`session-tool-chip mono session-tool-chip--${l[s]??`idle`}`,children:i.tool}),(0,Z.jsxs)($,{variant:`default`,disabled:e||k,onClick:()=>y(i,a),children:[k?(0,Z.jsx)(U,{size:12,className:`spin`}):(0,Z.jsx)(H,{size:12}),e?`Completed`:k?`Running…`:`Run ${i.tool}`]}),k&&!e&&(0,Z.jsxs)($,{variant:`danger`,onClick:()=>{b()},children:[(0,Z.jsx)(de,{size:12}),` Stop`]})]}),c.desc&&(0,Z.jsx)(`p`,{className:`session-tool-description`,children:c.desc}),(0,Z.jsxs)(`div`,{className:`session-param-grid`,children:[Object.keys(c.params).map(t=>(0,Z.jsx)(Bt,{name:t,value:p[s]?.[t]??``,onChange:e=>m(n=>({...n,[s]:{...n[s]??{},[t]:e}})),required:!0,disabled:e},t)),Object.keys(c.optional).length>0&&(0,Z.jsxs)(`button`,{className:`run-opt-btn`,onClick:()=>g(e=>({...e,[s]:!e[s]})),children:[h[s]?(0,Z.jsx)(v,{size:12}):(0,Z.jsx)(N,{size:12}),`Optional parameters (`,Object.keys(c.optional).length,`)`]}),h[s]&&Object.keys(c.optional).map(t=>(0,Z.jsx)(Bt,{name:t,value:p[s]?.[t]??``,onChange:e=>m(n=>({...n,[s]:{...n[s]??{},[t]:e}})),disabled:e},t))]})]}),(0,Z.jsxs)(`div`,{className:`session-result-panel`,children:[(0,Z.jsx)(`h4`,{children:`Result`}),e&&(0,Z.jsx)(`p`,{className:`section-meta`,children:`Completed session: results are read-only.`}),_?.error&&(0,Z.jsx)(`div`,{className:`run-error`,children:_.error}),A?(0,Z.jsxs)(Z.Fragment,{children:[(0,Z.jsxs)(`div`,{className:`session-result-head`,children:[(0,Z.jsxs)(`div`,{className:`session-step-result mono`,children:[A.success?`OK`:`FAIL`,` | exit `,A.return_code,` | `,A.execution_time.toFixed(2),`s`]}),s&&A&&(0,Z.jsxs)(`div`,{className:`session-result-actions`,children:[(0,Z.jsxs)($,{variant:`default`,onClick:()=>Un(`txt`,i.tool,p[s]??{},A),children:[(0,Z.jsx)(r,{size:11}),` Export`]}),(0,Z.jsxs)($,{variant:`default`,onClick:()=>Un(`json`,i.tool,p[s]??{},A),children:[(0,Z.jsx)(r,{size:11}),` JSON`]}),i.tool===`create-attack-chain`&&A.success&&(0,Z.jsx)($,{disabled:e,variant:`success`,onClick:()=>{x()},children:`Use Chain`})]})]}),(0,Z.jsx)(`pre`,{className:`session-result-pre mono`,children:A.stdout||`(no stdout)`}),A.stderr&&(0,Z.jsx)(`pre`,{className:`session-result-pre mono`,children:A.stderr})]}):(0,Z.jsx)(`p`,{className:`section-meta`,children:`No result yet for this tool.`})]})]})})]})]})}function Gn({show:e,templateName:t,templateError:n,setTemplateName:r,onClose:i,onSave:a}){return e?(0,Z.jsx)(`div`,{className:`modal-backdrop`,onClick:e=>{e.target===e.currentTarget&&i()},children:(0,Z.jsxs)(`div`,{className:`modal`,children:[(0,Z.jsxs)(`div`,{className:`modal-header`,children:[(0,Z.jsx)(`div`,{className:`modal-title-row`,children:(0,Z.jsx)(`span`,{className:`modal-name`,children:`Create Template`})}),(0,Z.jsx)(`button`,{className:`modal-close`,onClick:i,children:`x`})]}),(0,Z.jsx)(`div`,{className:`modal-body`,children:(0,Z.jsxs)(`div`,{className:`session-start-form`,children:[(0,Z.jsx)(`label`,{className:`mono`,children:`Template Name *`}),(0,Z.jsx)(`input`,{className:`search-input mono`,placeholder:`e.g. SMB Enumeration Pack`,value:t,onChange:e=>r(e.target.value)}),n&&(0,Z.jsx)(`div`,{className:`run-error`,children:n}),(0,Z.jsxs)(`div`,{className:`session-start-actions`,children:[(0,Z.jsx)(`button`,{className:`session-action-btn`,onClick:i,children:`Cancel`}),(0,Z.jsx)(`button`,{className:`session-run-btn`,onClick:a,children:`Save Template`})]})]})})]})}):null}function Kn(e){return Array.isArray(e.workflow_steps)&&e.workflow_steps.length>0?e.workflow_steps:e.tools_executed.map(e=>({tool:e,parameters:{}}))}function qn(e){return e.toLowerCase().replace(/[^a-z0-9]/g,``)}function Jn(e,t){let n=e.trim();if(!n)return null;let r=t.find(e=>e.name===n);if(r)return r;let i=t.find(e=>e.endpoint===n);if(i)return i;let a=t.find(e=>e.parent_tool===n);if(a)return a;let o=qn(n),s=null;for(let e of t){let t=qn(e.name),n=qn(e.parent_tool??``),r=qn(e.endpoint),i=0;t===o&&(i=Math.max(i,80)),n===o&&(i=Math.max(i,75)),r===o&&(i=Math.max(i,70)),t.includes(o)&&(i=Math.max(i,62)),r.includes(o)&&(i=Math.max(i,58)),n&&n.includes(o)&&(i=Math.max(i,56)),o.includes(t)&&(i=Math.max(i,52)),i!==0&&(!s||i>s.score)&&(s={tool:e,score:i})}return s?.tool??null}function Yn({sessionId:e,tools:t,onBack:n,onToolRun:i}){let[a,o]=(0,W.useState)(null),[s,c]=(0,W.useState)(!0),[l,u]=(0,W.useState)(null),[d,f]=(0,W.useState)(``),[p,m]=(0,W.useState)({}),[h,g]=(0,W.useState)({}),[_,v]=(0,W.useState)(null),[y,b]=(0,W.useState)({}),[x,S]=(0,W.useState)({}),[C,T]=(0,W.useState)(0),[E,D]=(0,W.useState)(!1),[O,k]=(0,W.useState)(``),[A,j]=(0,W.useState)(null),[ee,N]=(0,W.useState)(!1),[P,ne]=(0,W.useState)(``),[re,F]=(0,W.useState)(!1),[I,L]=(0,W.useState)(null),[ie,ae]=(0,W.useState)(!1),[oe,se]=(0,W.useState)(!1),[ce,R]=(0,W.useState)(!1),[le,de]=(0,W.useState)(!1),fe=(0,W.useMemo)(()=>Object.fromEntries(t.map(e=>[e.name,e])),[t]),z=a?Kn(a):[];async function B(){c(!0);try{let t=await J.session(e);o(t.session),f(t.session.target);let n=t.session.metadata??{},r=n.tool_status??{},i=n.step_results??{},a=typeof n.running_step_key==`string`?n.running_step_key:null,s={},c={};for(let[e,t]of Object.entries(r))(t===`success`||t===`failed`)&&(s[e]=t);for(let[e,t]of Object.entries(i))!t||typeof t!=`object`||(c[e]={result:{success:!!t.success,return_code:Number(t.return_code??0),execution_time:Number(t.execution_time??0),timed_out:!1,partial_results:!1,stdout:String(t.stdout??``),stderr:String(t.stderr??``),timestamp:String(t.timestamp??new Date().toISOString())}});S(s),b(c);try{let e=await J.processList();Object.keys(e.active_processes??{}).length>0&&a?(s[a]=`running`,v(a),S({...s})):v(null)}catch{}u(null)}catch(e){u(String(e))}finally{c(!1)}}(0,W.useEffect)(()=>{B()},[e]);async function pe(){if(a){F(!0);try{let e=`Tools: ${z.map(e=>e.tool).join(`, `)}`,t=await J.handoverSession(a.session_id,e);L(`LLM handover done -> ${t.handover?.category??`unknown`} (${((t.handover?.confidence??0)*100).toFixed(0)}% confidence)`),await B()}catch(e){L(`Handover failed: ${String(e)}`)}finally{F(!1)}}}async function he(){if(a){R(!0);try{await J.updateSession(a.session_id,{status:`completed`}),n()}catch(e){L(`Complete failed: ${String(e)}`)}finally{R(!1),ae(!1)}}}async function ge(){if(a){de(!0);try{await J.deleteSession(a.session_id),n()}catch(e){L(`Delete failed: ${String(e)}`)}finally{de(!1),se(!1)}}}function _e(){if(a)try{let e=a.metadata??{},t=e.tool_status&&typeof e.tool_status==`object`?e.tool_status:{},n=e.step_results&&typeof e.step_results==`object`?e.step_results:{},r=z.map((e,r)=>{let i=`${a.session_id}:${r}`,o=y[i]?.result,s=n[i],c=o?{success:o.success,return_code:o.return_code,execution_time:o.execution_time,timestamp:o.timestamp,stdout:o.stdout,stderr:o.stderr,timed_out:o.timed_out,partial_results:o.partial_results}:s?{success:s.success,return_code:s.return_code,execution_time:s.execution_time,timestamp:s.timestamp??``,stdout:s.stdout??``,stderr:s.stderr??``,timed_out:!1,partial_results:!1}:null;return{step_index:r,step_key:i,tool:e.tool,parameters:e.parameters??{},status:x[i]??t[i]??`idle`,result:c}}),i={session_id:a.session_id,target:a.target,status:a.status??`active`,exported_at:new Date().toISOString(),logs:r},o=JSON.stringify(i,null,2),s=new Blob([o],{type:`application/json`}),c=URL.createObjectURL(s),l=document.createElement(`a`);l.href=c,l.download=`${a.session_id}_tool_logs.json`,l.click(),URL.revokeObjectURL(c),L(`Tool logs exported`)}catch(e){L(`Export failed: ${String(e)}`)}}async function ve(){if(!a)return;let e=O.trim();if(!e){j(`Template name is required`);return}try{let t=Kn(a);if(t.length===0){j(`Session has no tools to save as template`);return}try{await J.createSessionTemplate({name:e,workflow_steps:t,source_session_id:a.session_id})}catch{await J.createSessionTemplateCompat({name:e,workflow_steps:t,source_session_id:a.session_id})}j(null),k(``),D(!1),L(`Template created: ${e}`)}catch(e){j(String(e))}}let V=z[C]??null,H=V&&a?`${a.session_id}:${C}`:null,ye=H?y[H]:void 0,be=V?fe[V.tool]:null;function xe(e){f(e),a&&m(n=>{let r=!1,i={...n};for(let o=0;o<z.length;o+=1){let s=z[o],c=Jn(s.tool,t);if(!c)continue;let l=`${a.session_id}:${o}`,u=n[l]??Pt(c,s,e),d=!1,f={...u};for(let t of[...Object.keys(c.params),...Object.keys(c.optional)]){let n=Nt(t,e);n===void 0||f[t]===n||(f[t]=n,d=!0)}(d||!n[l])&&(i[l]=f,r=!0)}return r?i:n})}(0,W.useEffect)(()=>{!a||!V||!H||!be||(m(e=>e[H]?e:{...e,[H]:Pt(be,V,d.trim()||a.target)}),g(e=>({...e,[H]:e[H]??!1})))},[a,V,H,be,d]);async function Se(e,n){if(!a)return;let r=a,o=`${r.session_id}:${n}`,s=Jn(e.tool,t);if(!s){b(t=>({...t,[o]:{error:`Tool ${e.tool} could not be mapped to registry`}})),S(e=>({...e,[o]:`failed`}));return}let c=d.trim(),{payload:l,missing:u}=Ft(s,p[o]??Pt(s,e,c||r.target));if(u.length>0){b(e=>({...e,[o]:{error:`Missing required: ${u.join(`, `)}`}})),S(e=>({...e,[o]:`failed`}));return}let f=z.map((e,t)=>t===n?{...e,parameters:l}:e);try{await J.updateSession(r.session_id,{target:c,workflow_steps:f})}catch{}let m=r.metadata??{};try{await J.updateSession(r.session_id,{metadata:{...m,running_step_key:o,running_tool:e.tool,running_started_at:new Date().toISOString()}})}catch{}v(o),S(e=>({...e,[o]:`running`})),b(e=>({...e,[o]:{}}));try{let t=await J.runTool(s.endpoint,l);b(e=>({...e,[o]:{result:t}})),S(e=>({...e,[o]:t.success?`success`:`failed`})),i&&i(e.tool,l,t);let n=r.metadata??{},a=n.tool_status&&typeof n.tool_status==`object`?n.tool_status:{},c=n.step_results&&typeof n.step_results==`object`?n.step_results:{};await J.updateSession(r.session_id,{metadata:{...n,tool_status:{...a,[e.tool]:t.success?`success`:`failed`,[o]:t.success?`success`:`failed`},step_results:{...c,[o]:{success:t.success,return_code:t.return_code,execution_time:t.execution_time,timestamp:t.timestamp,stdout:t.stdout,stderr:t.stderr}},last_run:{step_key:o,tool:e.tool,success:t.success,return_code:t.return_code,execution_time:t.execution_time,timestamp:t.timestamp},running_step_key:null,running_tool:null,running_started_at:null}}),await B()}catch(e){b(t=>({...t,[o]:{error:String(e)}})),S(e=>({...e,[o]:`failed`}));try{let e=r.metadata??{};await J.updateSession(r.session_id,{metadata:{...e,running_step_key:null,running_tool:null,running_started_at:null}})}catch{}}finally{v(null)}}async function Ce(){let e=_;if(e)try{let t=(await J.processList()).active_processes??{},n=Object.values(t).filter(e=>e&&typeof e.command==`string`&&e.command.trim()!==``).map(e=>e.pid).filter(e=>typeof e==`number`);if(n.length===0){L(`No active process found to terminate`),v(null);return}let r=Math.max(...n);if(await J.terminateProcess(r),L(`Stopped running process ${r}`),v(null),S(t=>({...t,[e]:`failed`})),b(t=>({...t,[e]:{...t[e]??{},error:`Process ${r} terminated by user`}})),a)try{let e=a.metadata??{};await J.updateSession(a.session_id,{metadata:{...e,running_step_key:null,running_tool:null,running_started_at:null}})}catch{}}catch(e){L(`Stop failed: ${String(e)}`)}}async function we(e){if(!a)return;let t=[...z,{tool:e.name,parameters:{}}];try{await J.updateSession(a.session_id,{workflow_steps:t}),N(!1),ne(``),await B(),T(Math.max(0,t.length-1))}catch(e){L(`Add tool failed: ${String(e)}`)}}async function Te(){if(!a||!H||!ye?.result)return;let e=ye.result,t=null;try{t=e.stdout?JSON.parse(e.stdout):null}catch{t=null}let n=(t&&typeof t==`object`?t:null)?.attack_chain,r=n&&typeof n==`object`?n.steps:null,i=Array.isArray(r)?r.filter(e=>!!e&&typeof e==`object`&&typeof e.tool==`string`):[];if(i.length===0){L(`No attack-chain steps found in result output`);return}let o=new Set(z.map(e=>`${e.tool}::${JSON.stringify(e.parameters??{})}`)),s=i.filter(e=>{let t=`${e.tool}::${JSON.stringify(e.parameters??{})}`;return o.has(t)?!1:(o.add(t),!0)});if(s.length===0){L(`All attack-chain tools are already present in manual execution`);return}let c=[...z,...s];try{await J.updateSession(a.session_id,{workflow_steps:c}),await B(),T(Math.max(0,c.length-s.length)),L(`Added ${s.length} tool(s) from attack chain`)}catch(e){L(`Apply attack chain failed: ${String(e)}`)}}async function Ee(e){if(!a)return;let t=z.filter((t,n)=>n!==e);try{await J.updateSession(a.session_id,{workflow_steps:t}),await B(),T(n=>t.length===0?0:n>e?n-1:Math.min(n,t.length-1))}catch(e){L(`Remove tool failed: ${String(e)}`)}}if(s)return(0,Z.jsxs)(`div`,{className:`loading-state`,children:[(0,Z.jsx)(U,{size:20,className:`spin`,color:`var(--green)`}),(0,Z.jsx)(`p`,{children:`Loading session…`})]});if(l||!a)return(0,Z.jsx)(`div`,{className:`error-banner`,children:l??`Session not found`});let De=(a.status??`active`)===`completed`,Oe=t.filter(e=>{if(!P.trim())return!0;let t=P.toLowerCase();return e.name.toLowerCase().includes(t)||e.desc.toLowerCase().includes(t)||e.category.toLowerCase().includes(t)}).slice(0,12);return(0,Z.jsxs)(`div`,{className:`page-content`,children:[(0,Z.jsxs)(`section`,{className:`section`,children:[(0,Z.jsxs)(`div`,{className:`section-header`,children:[(0,Z.jsx)(`h3`,{children:`Session Detail`}),(0,Z.jsxs)(`button`,{className:`session-action-btn`,onClick:n,children:[(0,Z.jsx)(me,{size:12}),` Back`]})]}),(0,Z.jsxs)(`div`,{className:`session-detail-meta-row`,children:[(0,Z.jsxs)(`span`,{children:[(0,Z.jsx)(te,{size:12}),` `,(0,Z.jsx)(`span`,{className:`mono`,children:a.target})]}),(0,Z.jsxs)(`span`,{children:[(0,Z.jsx)(M,{size:12}),` `,a.total_findings,` findings`]}),(0,Z.jsxs)(`span`,{children:[(0,Z.jsx)(U,{size:12}),` `,a.iterations,` iterations`]}),(0,Z.jsxs)(`span`,{children:[(0,Z.jsx)(w,{size:12}),` `,dt(a.updated_at)]}),(0,Z.jsx)(`span`,{className:`session-status session-status--${a.status??`active`}`,children:a.status??`active`})]}),(0,Z.jsxs)(`div`,{className:`session-detail-actions`,children:[(0,Z.jsx)(`button`,{className:`session-action-btn`,onClick:()=>D(!0),children:`Create Template`}),(0,Z.jsxs)(`button`,{className:`session-action-btn`,onClick:_e,children:[(0,Z.jsx)(r,{size:12}),` Export Tool Logs`]}),(0,Z.jsxs)(`button`,{className:`session-action-btn`,onClick:pe,disabled:re,children:[(0,Z.jsx)(ue,{size:12}),` `,re?`Handing over…`:`Handover to LLM`]}),a.status!==`completed`&&(0,Z.jsx)(`button`,{className:`session-complete-btn`,onClick:()=>ae(!0),children:`Complete Session`}),(0,Z.jsx)(`button`,{className:`session-delete-btn`,onClick:()=>se(!0),children:`Delete Session`}),I&&(0,Z.jsx)(`span`,{className:`section-meta`,children:I})]}),(0,Z.jsx)(Gn,{show:E,templateName:O,templateError:A,setTemplateName:k,onClose:()=>D(!1),onSave:ve}),(0,Z.jsx)(Ht,{isOpen:ie,title:`Complete Session`,description:`Mark this session as completed? It will move to Completed Sessions.`,impactItems:[`Session ID: ${a.session_id}`,`Session will become read-only`],confirmLabel:`Yes, complete session`,cancelLabel:`Keep active`,confirmVariant:`danger`,isConfirming:ce,onConfirm:he,onClose:()=>ae(!1)}),(0,Z.jsx)(Ht,{isOpen:oe,title:`Delete Session`,description:`Delete this session permanently? This action cannot be undone.`,impactItems:[`Session ID: ${a.session_id}`,`All workflow state and run metadata will be removed`],confirmLabel:`Yes, delete session`,cancelLabel:`Keep session`,confirmVariant:`danger`,isConfirming:le,onConfirm:ge,onClose:()=>se(!1)})]}),(0,Z.jsx)(Wn,{isCompleted:De,sessionId:a.session_id,steps:z,selectedStep:V,selectedStepIndex:C,setSelectedStepIndex:T,selectedStepKey:H,selectedTool:be,stepState:x,runningStepKey:_,targetValue:d,setTargetValue:xe,stepFieldValues:p,setStepFieldValues:m,showOptionalByStep:h,setShowOptionalByStep:g,selectedResult:ye,onRunStep:Se,onStopRunningStep:Ce,onApplyAttackChainFromResult:Te,onRemoveTool:Ee,showAddTool:ee,setShowAddTool:N,addToolSearch:P,setAddToolSearch:ne,addCandidates:Oe,onAddTool:we})]})}function Xn({page:e,demo:t,tools:n,health:r,toolsStatusWithParents:i,runHistory:a,setRunHistory:o,fetchServerRunHistory:s,clearServerRunHistory:c,commandToolRequest:l,onCommandToolHandled:u,openSessionDetail:d,activeSessionId:f,setPage:p,addBrowserRunEntry:m,logLines:h,logAutoScroll:g,setLogAutoScroll:_,logLimit:v,setLogLimit:y,logEndRef:b,loading:x,error:S,history:C,toolCategories:w}){return(0,Z.jsxs)(`main`,{className:`main${e===`run`?` main--flush`:``}`,children:[e===`settings`&&(0,Z.jsx)(hn,{}),e===`help`&&(0,Z.jsx)(Cn,{}),e===`run`&&(0,Z.jsx)(Xt,{tools:n,toolsStatus:i,runHistory:a,setRunHistory:o,commandToolRequest:l,onCommandToolHandled:u,onRefresh:s,onClearHistory:c}),e===`tasks`&&(0,Z.jsx)(Dn,{demoData:t?{processes:Xe}:void 0}),e===`tools`&&r&&(0,Z.jsx)(kn,{health:r,tools:n,toolsStatus:i}),e===`reports`&&(0,Z.jsx)(Fn,{runHistory:a}),e===`sessions`&&(0,Z.jsx)(Hn,{demoData:t?{sessions:Ye}:void 0,onOpenSession:d}),e===`session-detail`&&f&&(0,Z.jsx)(Yn,{sessionId:f,tools:n,onBack:()=>p(`sessions`),onToolRun:m}),e===`logs`&&(0,Z.jsx)(nn,{logLines:h,logAutoScroll:g,setLogAutoScroll:_,logLimit:v,setLogLimit:y,logEndRef:b}),e===`dashboard`&&(0,Z.jsxs)(Z.Fragment,{children:[x&&!r&&(0,Z.jsxs)(`div`,{className:`loading-state`,children:[(0,Z.jsx)(U,{size:24,className:`spin`,color:`var(--green)`}),(0,Z.jsx)(`p`,{children:`Connecting to server…`})]}),S&&!r&&(0,Z.jsxs)(`div`,{className:`error-banner`,children:[S,` — is the server running on port 8888?`]}),r&&(0,Z.jsx)(kt,{health:r,tools:n,history:C,runHistory:a,loading:x,error:S,toolCategories:w})]})]})}var Zn=[{page:`dashboard`,label:`Open Dashboard`},{page:`run`,label:`Open Run`},{page:`reports`,label:`Open Reports`},{page:`tasks`,label:`Open Tasks`},{page:`tools`,label:`Open Tools`},{page:`sessions`,label:`Open Sessions`},{page:`settings`,label:`Open Settings`},{page:`logs`,label:`Open Logs`},{page:`help`,label:`Open Help`}];function Qn({open:e,onClose:t,setPage:n,tools:r,onSelectTool:i}){let[a,o]=(0,W.useState)(``),[s,c]=(0,W.useState)(0),l=(0,W.useMemo)(()=>{let e=Zn.map(e=>({id:`page-${e.page}`,label:e.label,hint:`Navigation`,run:()=>{n(e.page),t()}})),a=r.slice(0,120).map(e=>({id:`tool-${e.name}`,label:`Run ${e.name}`,hint:e.category.replace(/_/g,` `),run:()=>{n(`run`),i(e),t()}}));return[...e,...a]},[r,n,i,t]),u=(0,W.useMemo)(()=>{let e=a.trim().toLowerCase();return e?l.filter(t=>`${t.label} ${t.hint}`.toLowerCase().includes(e)).slice(0,18):l.slice(0,18)},[l,a]);return(0,W.useEffect)(()=>{e&&(o(``),c(0))},[e]),(0,W.useEffect)(()=>{if(!e)return;function n(e){if(e.key===`Escape`){e.preventDefault(),t();return}if(e.key===`ArrowDown`){e.preventDefault(),c(e=>Math.min(e+1,Math.max(u.length-1,0)));return}if(e.key===`ArrowUp`){e.preventDefault(),c(e=>Math.max(e-1,0));return}if(e.key===`Enter`){let t=u[s];t&&(e.preventDefault(),t.run())}}return window.addEventListener(`keydown`,n),()=>window.removeEventListener(`keydown`,n)},[e,u,s,t]),e?(0,At.createPortal)((0,Z.jsx)(`div`,{className:`modal-backdrop`,onClick:t,children:(0,Z.jsxs)(`div`,{className:`modal command-palette`,onClick:e=>e.stopPropagation(),children:[(0,Z.jsxs)(`div`,{className:`command-palette-input-wrap`,children:[(0,Z.jsx)(O,{size:14}),(0,Z.jsx)(`input`,{autoFocus:!0,className:`command-palette-input mono`,value:a,onChange:e=>{o(e.target.value),c(0)},placeholder:`Type a command or tool name...`})]}),(0,Z.jsxs)(`div`,{className:`command-palette-list`,children:[u.map((e,t)=>(0,Z.jsxs)(`button`,{className:`command-palette-item${t===s?` active`:``}`,onClick:e.run,children:[(0,Z.jsx)(`span`,{className:`command-palette-label`,children:e.label}),(0,Z.jsx)(`span`,{className:`command-palette-hint`,children:e.hint})]},e.id)),u.length===0&&(0,Z.jsx)(`div`,{className:`command-palette-empty`,children:`No results`})]})]})}),document.body):null}var $n=1e4,er=`hexstrike_palette_hint_seen`;function tr(){let[e]=(0,W.useState)(Re),[t,n]=(0,W.useState)(e||Fe()),[r,i]=(0,W.useState)(!1),a=nt(),[o,s]=(0,W.useState)(a.page),[c,l]=(0,W.useState)(a.sessionId);function u(e){e!==`session-detail`&&(window.location.hash=e===`dashboard`?``:`/${e}`,s(e),l(null))}function d(e){window.location.hash=`/sessions/${e}`,s(`session-detail`),l(e)}(0,W.useEffect)(()=>{function e(){let e=nt();s(e.page),l(e.sessionId)}return window.addEventListener(`hashchange`,e),()=>window.removeEventListener(`hashchange`,e)},[]);let[f,p]=(0,W.useState)(e?Ge:null),[m,h]=(0,W.useState)(e?Be:[]),[g,v]=(0,W.useState)(e?Ke():[]);(0,W.useEffect)(()=>{e||J.tools().then(e=>h(e.tools)).catch(()=>{})},[e]);let[y,b]=(0,W.useState)(()=>e?qe:[]),[x,S]=(0,W.useState)(e?new Date:null),[C,w]=(0,W.useState)(!e),[T,E]=(0,W.useState)(null),[D,O]=(0,W.useState)(e?Je:[]),[k,A]=(0,W.useState)(!0),[j,ee]=(0,W.useState)(500),M=(0,W.useRef)(null),te=(0,W.useRef)(null),N=(0,W.useRef)(null),P=(0,W.useRef)(null),[ne,re]=(0,W.useState)(!1),[F,I]=(0,W.useState)(null),[L,ie]=(0,W.useState)({}),[ae,oe]=(0,W.useState)(!1),[se,ce]=(0,W.useState)(null),[R,le]=(0,W.useState)(()=>{try{return localStorage.getItem(er)!==`1`}catch{return!0}}),[ue,de]=(0,W.useState)(()=>{try{let e=localStorage.getItem(at);if(e&&st(e))return e}catch{}return`dark`});(0,W.useEffect)(()=>{document.documentElement.setAttribute(`data-theme`,ue);try{localStorage.setItem(at,ue)}catch{}},[ue]),(0,W.useEffect)(()=>{function e(e){(e.ctrlKey||e.metaKey)&&e.key.toLowerCase()===`k`&&(e.preventDefault(),oe(!0))}return window.addEventListener(`keydown`,e),()=>window.removeEventListener(`keydown`,e)},[]);function fe(e){u(`run`),ce({toolName:e.name,requestId:Date.now()})}function z(){le(!1);try{localStorage.setItem(er,`1`)}catch{}}function B(){oe(!0),R&&z()}let pe=(0,W.useCallback)((e,t,n)=>{b(r=>[{id:Date.now(),tool:e,params:t,result:n,ts:new Date,source:`browser`},...r].slice(0,200))},[]),me=(0,W.useCallback)(async()=>{if(!e)try{let e=await J.dashboard();p(e),v(t=>[...t.slice(-29),{t:Date.now(),cpu:e.resources.cpu_percent,mem:e.resources.memory_percent,network_bytes_sent:e.resources.network_bytes_sent,network_bytes_recv:e.resources.network_bytes_recv}]),S(new Date),E(null)}catch(e){e instanceof Error&&e.message===`UNAUTHORIZED`?(i(!0),n(!1)):E(`Server unreachable`)}finally{w(!1)}},[]);(0,W.useEffect)(()=>{e||!t||(async()=>{try{ie((await J.getToolCategories()).categories)}catch{}})()},[e,t]);let he=(0,W.useCallback)(async()=>{if(!e)try{let e=await J.runHistory();if(!e.success)return;b(t=>{let n=new Set(t.filter(e=>e.source===`server`).map(e=>e.serverId)),r=e.runs.filter(e=>{if(n.has(e.id))return!1;let r=e.timestamp?new Date(e.timestamp).getTime():0;return!t.some(t=>t.source===`browser`&&t.tool===e.tool&&r>0&&Math.abs(t.ts.getTime()-r)<1e4)}).map(e=>({id:-e.id,serverId:e.id,source:`server`,tool:e.tool,params:e.params,ts:e.timestamp?new Date(e.timestamp):new Date,result:{stdout:e.stdout,stderr:e.stderr,return_code:e.return_code,success:e.success,timed_out:e.timed_out,partial_results:e.partial_results,execution_time:e.execution_time,timestamp:e.timestamp}}));return r.length===0?t:[...t,...r].sort((e,t)=>t.ts.getTime()-e.ts.getTime())})}catch{}},[]),ge=(0,W.useCallback)(async()=>{if(e){b([]);return}(await J.clearRunHistory()).success&&b([])},[e]);if((0,W.useEffect)(()=>{e||!t||he().catch(()=>{})},[e,t,he]),(0,W.useEffect)(()=>{e||Fe()||J.dashboard().then(e=>{p(e),n(!0),w(!1)}).catch(e=>{e instanceof Error&&e.message===`UNAUTHORIZED`?i(!0):n(!0),w(!1)})},[]),(0,W.useEffect)(()=>{if(e||!t)return;N.current&&N.current.close(),P.current&&=(clearInterval(P.current),null);function n(){P.current&&clearInterval(P.current),P.current=window.setInterval(()=>{me()},$n)}let r=J.dashboardStream();return N.current=r,r.onmessage=e=>{try{let t=JSON.parse(e.data);p(t),v(e=>[...e.slice(-29),{t:Date.now(),cpu:t.resources.cpu_percent,mem:t.resources.memory_percent,network_bytes_sent:t.resources.network_bytes_sent,network_bytes_recv:t.resources.network_bytes_recv}]),S(new Date),w(!1),E(null),re(!0),I(null),P.current&&=(clearInterval(P.current),null)}catch{I(`Malformed dashboard data`)}},r.onerror=()=>{re(!1),I(`Dashboard stream disconnected; using polling.`),P.current||n()},()=>{r.close(),P.current&&clearInterval(P.current)}},[e,t,me]),(0,W.useEffect)(()=>{if(e||o!==`logs`)return;let t=J.logStream(150);return te.current=t,t.onmessage=e=>{O(t=>{let n=[...t,e.data];return n.length>500?n.slice(-500):n})},()=>{t.close()}},[e,o]),(0,W.useEffect)(()=>{o===`logs`&&k&&M.current?.scrollIntoView({behavior:`smooth`})},[D,o,k]),r&&!t)return(0,Z.jsx)(Ze,{onUnlocked:()=>{n(!0),i(!1)}});let _e=rt(m,f?.tools_status??{});return(0,Z.jsx)($e,{children:(0,Z.jsxs)(`div`,{className:e?`layout layout--demo`:`layout`,children:[(0,Z.jsx)(Qn,{open:ae,onClose:()=>oe(!1),setPage:u,tools:m,onSelectTool:fe}),e&&(0,Z.jsxs)(`div`,{className:`demo-banner`,children:[(0,Z.jsx)(_,{size:13}),(0,Z.jsx)(`span`,{children:`Demo mode — all data is synthetic`}),(0,Z.jsx)(`button`,{onClick:()=>{ze(),window.location.href=window.location.pathname+window.location.hash},children:`Exit demo`})]}),(0,Z.jsx)(lt,{page:o,setPage:u,lastRefresh:x,demo:e,isStreaming:ne,streamingError:F,health:f,error:T,loading:C,fetchAll:me,themeId:ue,setThemeId:de,onOpenCommandPalette:B,onSignOut:()=>{n(!1),i(!0)}}),R&&(0,Z.jsxs)(`div`,{className:`palette-hint-floating`,role:`status`,"aria-live":`polite`,children:[(0,Z.jsxs)(`button`,{className:`palette-hint-main`,onClick:B,title:`Open command palette`,children:[(0,Z.jsx)(`span`,{className:`palette-hint-title`,children:`Try Command Palette`}),(0,Z.jsx)(`span`,{className:`palette-hint-kbd mono`,children:`Ctrl/Cmd + K`})]}),(0,Z.jsx)(`button`,{className:`palette-hint-close`,onClick:z,title:`Dismiss hint`,children:`x`})]}),(0,Z.jsx)(Xn,{page:o,demo:e,tools:m,health:f,toolsStatusWithParents:_e,runHistory:y,setRunHistory:b,fetchServerRunHistory:he,clearServerRunHistory:ge,commandToolRequest:se,onCommandToolHandled:()=>ce(null),openSessionDetail:d,activeSessionId:c,setPage:u,addBrowserRunEntry:pe,logLines:D,logAutoScroll:k,setLogAutoScroll:A,logLimit:j,setLogLimit:ee,logEndRef:M,loading:C,error:T,history:g,toolCategories:L})]})})}(0,ke.createRoot)(document.getElementById(`root`)).render((0,Z.jsx)(W.StrictMode,{children:(0,Z.jsx)(tr,{})}));
