+++ server_static/index.html
@@ -5,7 +5,7 @@
     <link rel="icon" type="image/svg+xml" href="/assets/favicon-D_bVPi1p.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>HexStrike - CE</title>
-    <script type="module" crossorigin src="/assets/index-ZAq391hv.js"></script>
+    <script type="module" crossorigin src="/assets/index-Bn3Gkg7X.js"></script>
     <link rel="modulepreload" crossorigin href="/assets/rolldown-runtime-Dw2cE7zH.js">
     <link rel="modulepreload" crossorigin href="/assets/vendor-lucide-Cq1rHr8h.js">
     <link rel="modulepreload" crossorigin href="/assets/vendor-react-DDfmLfRS.js">
